Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 4.1913e-01  7.8223e-01 -1.7144e+00  1.0483e+00  1.0668e+00  3.3036e-01
#>  3.3297e-01 -1.1468e-01  1.4690e-01 -8.0840e-01  5.4413e-01  7.5378e-01
#>  6.0527e-01 -6.8181e-01  6.2525e-01 -3.1399e-01 -9.1537e-01  7.0975e-01
#> -4.9668e-01 -8.4401e-01 -1.8724e-01  7.9553e-01  9.5617e-01  2.1224e-01
#> -1.0569e+00  2.0444e-02  8.1682e-01  6.2155e-01  1.1125e+00 -1.0282e-01
#>  1.8339e-01 -1.4503e+00 -7.7243e-01  6.5622e-01  1.1451e+00  9.1079e-01
#> -7.2241e-01  1.4636e+00  1.1646e+00  4.6115e-01  2.4694e-01 -7.7103e-01
#>  3.7356e-01 -1.9876e+00 -1.4368e+00 -5.5648e-01  5.7073e-01 -7.7945e-01
#>  5.2903e-01  2.4044e-01  1.6957e-01  2.0880e-01  1.4554e-01  1.1959e+00
#>  1.6065e+00  1.7430e-01 -9.3383e-01  2.5217e-01  4.0720e-02  5.3699e-01
#> 
#> Columns 7 to 12-3.0058e-01  1.3651e-01  5.0938e-01 -1.5374e+00 -2.6510e-01  5.7450e-01
#> -5.4295e-01 -1.4979e-01 -2.2234e-02 -1.4915e-01  5.9554e-02 -3.5632e-01
#>  4.4990e-01  1.2031e+00 -3.3210e-01  2.6913e-02 -2.3042e+00 -5.5973e-01
#> -1.2723e+00 -1.1502e+00  1.0108e+00 -1.0125e+00 -7.0493e-01 -1.4019e+00
#>  1.2385e-01 -4.1238e-01 -1.6869e+00 -1.1065e+00  2.9555e-01  5.1981e-01
#>  2.5564e-01  2.6509e-01  2.8797e-01 -1.0750e+00  1.4716e+00  9.6324e-01
#>  2.0476e-01 -6.7067e-01  8.4208e-01 -2.3931e-01 -1.5886e+00  9.9010e-01
#>  1.1193e+00 -6.7415e-01  1.5565e-02  5.3145e-01 -4.5201e-01  7.8439e-01
#>  8.8614e-02 -7.8221e-01  1.1141e+00 -1.9414e+00 -8.7354e-01  1.3813e+00
#> -1.7245e-01  5.0588e-01 -3.9458e-01  7.3266e-01  1.0853e+00  2.1428e-02
#> 
#> Columns 13 to 18-4.7255e-01  1.4434e+00  1.1623e+00  6.4374e-01 -7.7601e-01 -1.1915e+00
#>  3.1910e-01  6.7131e-01  1.4685e-03 -4.9002e-01 -9.9734e-01  7.7727e-02
#>  1.0382e+00  8.8333e-02  1.2421e+00 -4.5797e-01 -3.7060e-01  1.2679e-01
#>  1.4694e+00  8.4733e-01  5.7437e-01  1.9506e-01 -1.4859e+00 -1.0333e+00
#>  6.5669e-01 -3.9713e-01  1.8985e+00  1.4219e-01 -2.2806e-01  9.8487e-01
#> -4.1537e-01 -7.8912e-01 -1.0699e-03 -5.2256e-01 -4.5008e-01 -1.2358e+00
#> -6.2210e-01 -2.6708e-01  6.1980e-01 -1.4221e+00  6.4000e-01 -1.9835e+00
#> -1.0788e+00  7.5746e-01 -1.1216e+00  1.1675e+00 -1.3671e+00 -7.5544e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.9077
#> -0.5188
#>  0.8724
#>  0.5169
#>  0.3932
#>  0.1862
#>  1.6573
#> -0.1483
#> -0.9767
#> -1.1716
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.1483
#>  0.1862
#> -0.9767
#> -0.9077
#>  1.6573
#>  0.3932
#> -0.5188
#>  0.5169
#>  0.8724
#> -1.1716
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>