Context object

Context object storing information about the model training context. See also ctx.

Active bindings

records

stores information about values logged with self$log.

device

allows querying the current accelerator device

Methods

Public methods

context$log()

context$log_metric()

context$get_log()

context$get_metrics()

context$get_metric()

context$get_formatted_metrics()

context$get_metrics_df()

context$set_verbose()

context$clone()

Method log()

Allows logging arbitrary information in the ctx.

Usage

context$log(what, set, value, index = NULL, append = TRUE)

Arguments

what

(string) What you are logging.

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

value

value to log

value

Arbitrary value to log.

index

Index that this value should be logged. If NULL the value is added to the end of list, otherwise the index is used.

append

If TRUE and a value in the corresponding index already exists, then value is appended to the current value. If FALSE value is overwritten in favor of the new value.

Method log()

Allows logging arbitrary information in the ctx.

Usage

context$log(what, set, value, index = NULL, append = TRUE)

Arguments

what

(string) What you are logging.

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

value

value to log

value

Arbitrary value to log.

index

Index that this value should be logged. If NULL the value is added to the end of list, otherwise the index is used.

append

If TRUE and a value in the corresponding index already exists, then value is appended to the current value. If FALSE value is overwritten in favor of the new value.

Method log_metric()

Log a metric gen its name and value. Metric values are indexed by epoch.

Usage

context$log_metric(name, value)

Arguments

name

name of the metric

value

value to log

value

Arbitrary value to log.

Method log_metric()

Log a metric gen its name and value. Metric values are indexed by epoch.

Usage

context$log_metric(name, value)

Arguments

name

name of the metric

value

value to log

value

Arbitrary value to log.

Method get_log()

Get an specific value from the log.

Usage

context$get_log(what, set, index = NULL)

Arguments

what

(string) What you are logging.

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

index

Index that this value should be logged. If NULL the value is added to the end of list, otherwise the index is used.

Method get_log()

Get an specific value from the log.

Usage

context$get_log(what, set, index = NULL)

Arguments

what

(string) What you are logging.

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

index

Index that this value should be logged. If NULL the value is added to the end of list, otherwise the index is used.

Method get_metrics()

Get all metric given an epoch and set.

Usage

context$get_metrics(set, epoch = NULL)

Arguments

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

epoch

The epoch you want to extract metrics from.

Method get_metrics()

Get all metric given an epoch and set.

Usage

context$get_metrics(set, epoch = NULL)

Arguments

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

epoch

The epoch you want to extract metrics from.

Method get_metric()

Get the value of a metric given its name, epoch and set.

Usage

context$get_metric(name, set, epoch = NULL)

Arguments

name

name of the metric

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

epoch

The epoch you want to extract metrics from.

Method get_metric()

Get the value of a metric given its name, epoch and set.

Usage

context$get_metric(name, set, epoch = NULL)

Arguments

name

name of the metric

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

epoch

The epoch you want to extract metrics from.

Method get_formatted_metrics()

Get formatted metrics values

Usage

context$get_formatted_metrics(set, epoch = NULL)

Arguments

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

epoch

The epoch you want to extract metrics from.

Method get_formatted_metrics()

Get formatted metrics values

Usage

context$get_formatted_metrics(set, epoch = NULL)

Arguments

set

(string) Usually 'train' or 'valid' indicating the set you want to lot to. But can be arbitrary info.

epoch

The epoch you want to extract metrics from.

Method get_metrics_df()

Get a data.frame containing all metrics.

Usage

context$get_metrics_df()

Method get_metrics_df()

Get a data.frame containing all metrics.

Usage

context$get_metrics_df()

Method set_verbose()

Allows setting the verbose attribute.

Usage

context$set_verbose(verbose = NULL)

Arguments

verbose

boolean. If TRUE verbose mode is used. If FALSE non verbose. if NULL we use the result of interactive().

Method set_verbose()

Allows setting the verbose attribute.

Usage

context$set_verbose(verbose = NULL)

Arguments

verbose

boolean. If TRUE verbose mode is used. If FALSE non verbose. if NULL we use the result of interactive().

Method clone()

The objects of this class are cloneable with this method.

Usage

context$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

context$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.