Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 1.2622  0.1650 -0.1894 -0.1861  0.5212 -1.0475 -0.8294 -1.4393  0.0091 -0.1699
#> -0.3755 -0.1719  0.3806 -0.5608 -1.5727  0.8658 -0.0154 -0.2931 -0.7496 -0.2113
#> -0.2209  0.5437 -0.5034  1.7677 -1.1448  0.0526 -1.1257 -0.5552  0.2757  1.5116
#>  0.8168 -0.8756  0.3363 -0.3370  0.0531 -0.8566  2.6308 -0.0813  0.3354 -0.6681
#> -0.6520 -1.8911 -0.2908 -0.5280 -1.1739 -0.5341  0.4595  0.4663 -1.6055  1.4493
#> -0.2476  0.8347 -0.4682  0.0090  0.1285  1.5464 -1.0656 -0.2141  0.3514 -1.7697
#>  0.4520 -1.2482 -0.2588 -0.9380  0.9737 -0.9956  1.2696  0.1609  0.4199 -0.4802
#>  1.6044 -0.6476 -0.0861  2.0878  0.5448  0.0709 -2.7548  1.0430  0.4025 -0.6299
#> -1.3504 -0.8310  0.7499 -2.2819  1.0545  0.0067 -0.3842  0.6177  0.5839 -0.8747
#>  0.7412 -0.1512 -1.4932 -1.0502 -0.6015 -0.4773  0.2367  0.8762 -0.5099 -0.1099
#> 
#> Columns 11 to 20 0.3671 -0.0014  0.3227  0.9116  0.7711 -0.9454  1.9565 -0.3173  0.3319  1.0081
#>  0.2683  0.5716  0.4725  0.2609  0.3753 -0.2032 -0.3073 -0.9958  1.7052 -0.8169
#>  0.6608 -0.7087  0.6903  1.5763  1.4961  1.0393  0.5671  0.0542 -0.5336 -0.2094
#> -1.1322 -0.5964  0.0963  0.0142  0.0652  0.8441  1.6943  0.8961 -0.5734 -0.2653
#>  0.5108 -0.5833 -0.2580  0.6759 -0.6912 -0.5908 -0.1068  0.9549  0.6544  0.5390
#>  0.2109 -0.2292 -0.8259 -0.3372  0.2325  0.7435 -0.1369  2.5693 -0.6928  1.3557
#>  0.6339 -0.5610 -1.2165  0.5422  0.6719  0.6913  0.8124  0.0255  1.1854 -0.0990
#> -0.5074 -0.4729 -0.8015 -1.7279 -1.3391 -0.0699  0.0714 -0.4202 -0.2701  0.7338
#> -0.4108 -0.5148  0.2213  0.4350 -1.3254  0.3814 -0.0447  0.7239  0.9939  0.4764
#>  0.0651  0.8383  0.4561 -0.0495 -0.7974 -0.1763  0.8792 -0.4966  0.4732  1.4211
#> 
#> Columns 21 to 30-0.7007 -0.5177  0.7326 -0.8112 -1.4317  1.5311 -1.3287 -0.7781  0.0323  1.6986
#> -1.3718 -0.0898  0.2055  1.1056 -0.6459 -0.4058 -0.7790  0.3126 -1.5165 -0.8891
#> -0.5625  0.5966  0.6083  0.5009  1.1249  0.2430 -0.6190 -0.0263 -0.0177 -0.9787
#>  1.1150  0.6256  0.3850  1.8903 -1.1473  0.5194 -0.4563  0.9321 -1.2428 -0.6445
#> -0.5241 -0.3306 -0.8651  0.4760 -0.1187 -0.0334 -0.0016  0.5962  1.5859 -0.5835
#>  0.1891  1.0190  0.3522  0.0173 -1.5031 -0.0308  0.8826 -1.2441  0.0957 -0.8260
#>  0.7897 -0.2124  0.2266  0.7327  0.4710 -1.9599 -0.5246 -0.7419 -0.6626  0.1802
#> -0.3777  0.3274  0.2190  0.7369 -0.3005  0.1547 -2.1464 -0.8314 -0.2130  0.9341
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.6308
#> -0.6488
#> -1.9627
#> -0.1898
#> -0.8919
#>  0.8388
#> -0.1869
#>  1.4009
#> -0.8711
#> -0.0715
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.0715
#>  0.6308
#>  0.8388
#> -0.8919
#> -1.9627
#>  1.4009
#> -0.1898
#> -0.8711
#> -0.6488
#> -0.1869
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>