Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 8.5123e-01 -2.3463e-01  3.8317e-01 -1.2767e+00 -5.5830e-02  5.8379e-01
#>  4.6536e-01 -9.2533e-01  8.7663e-02 -6.1327e-01 -1.2365e+00 -8.0685e-01
#>  1.3937e+00  6.3947e-01 -1.5583e-01  1.5702e+00  6.6039e-01  3.1843e-01
#> -1.6431e+00  3.2083e-02 -6.3983e-01 -3.4872e-01 -1.4586e-01 -1.1498e+00
#>  1.3709e+00 -4.6546e-01 -8.4086e-01  1.2490e+00 -8.9286e-02  1.4321e-01
#> -7.3411e-01 -1.0717e-01 -2.7517e-02  1.4303e-01  1.2054e+00  3.8857e-01
#> -4.4807e-01  1.2764e+00  1.4735e+00 -1.6790e-01 -5.9479e-01 -4.9319e-01
#> -4.4617e-01 -1.4564e+00 -5.4988e-01  9.1221e-01 -1.4604e-01  5.5934e-02
#> -1.3748e+00  2.2325e-01 -7.2240e-01  8.6238e-01 -4.1740e-01 -1.5160e+00
#> -1.8137e-02 -9.9196e-01  1.4639e+00  4.0881e-01  1.1027e+00  1.1588e+00
#> 
#> Columns 7 to 12-6.4568e-01  5.5918e-01 -4.4009e-01  2.1005e+00  8.1982e-01  7.9639e-01
#> -5.9700e-01  5.4878e-02  5.2911e-01  3.3657e-01 -5.3830e-01 -1.0370e-01
#> -6.9307e-02 -1.9114e-01 -1.3790e+00  1.5733e+00  3.8147e-01 -5.5850e-03
#> -3.3821e-01  9.9584e-01 -4.5066e-01  3.0392e-02 -1.2103e+00 -5.8858e-01
#>  1.5827e-01  1.6087e+00 -1.5641e+00 -1.4405e+00  1.3252e+00  3.8523e-01
#> -5.1680e-03 -5.0635e-01 -4.0108e-01  5.8603e-01  4.7939e-01 -1.6152e-01
#>  5.3945e-01  3.5919e-01 -4.2674e-01  5.7090e-01  9.6642e-01 -9.7820e-01
#>  1.6146e+00 -7.3540e-01 -4.3930e-01 -2.6951e-01 -1.1070e+00  1.3129e+00
#>  3.5196e-01  3.2290e-01  2.6045e-03 -1.1448e+00 -6.6565e-01 -9.1625e-01
#>  1.9254e-01 -2.6240e-01 -3.0609e-01  6.5206e-01 -2.0282e-01 -1.5683e-01
#> 
#> Columns 13 to 18-2.6551e-01 -4.1790e-01  5.6584e-01 -2.0419e-01 -7.5727e-01  7.1263e-01
#>  1.3771e-01  3.7786e-01  4.0937e-01 -5.0517e-01 -4.3723e-02 -1.4528e-01
#>  1.2086e+00  6.0847e-01 -7.2407e-03 -1.8323e+00 -1.3514e-01 -5.8288e-01
#>  1.3770e+00  1.8609e-01  1.8976e+00 -7.9110e-01 -2.2914e+00 -1.3370e+00
#> -3.3583e-01  7.7796e-01  1.0837e+00  2.5957e-01  6.7310e-01  1.2131e+00
#> -1.9217e-01  9.8788e-01  6.0967e-01 -1.6051e+00  3.7847e-01 -7.0697e-01
#>  1.2775e+00 -9.9641e-01 -1.2597e+00  2.4924e-01 -4.5665e-01 -1.1268e+00
#> -3.5592e-01  2.3698e-01 -4.8083e-01 -2.4817e-02 -9.4267e-01  1.0301e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.9037
#> -0.5006
#> -1.0617
#> -0.0916
#> -0.5620
#> -1.6712
#>  0.3114
#> -1.3528
#> -0.0332
#>  0.0102
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.9037
#> -0.0332
#> -0.0916
#> -0.5006
#>  0.0102
#> -1.0617
#>  0.3114
#> -0.5620
#> -1.3528
#> -1.6712
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>