Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.3027  0.3493  0.7590  0.1593  0.1583  1.6682  0.8075  1.4383 -1.0759 -0.1779
#>  1.0119  1.1554 -2.2477  0.7016  0.7003 -1.6007 -0.2726 -0.3389 -0.8318  1.8920
#> -0.9468 -1.9456 -1.1004 -0.5070 -0.5986  1.2058 -0.4429 -0.7691  1.9440  0.8910
#> -0.0791 -0.1029  0.1607 -0.1227  0.2940  1.0537 -0.0533  0.8479 -0.5914 -0.3198
#>  1.0430 -0.0734  0.2181 -0.1631 -0.2136  2.2046 -1.9850  1.6951  0.2090  1.1684
#>  0.1621 -1.4150 -1.0329 -0.7733  1.0304  0.7755 -0.6452 -0.3381  0.7409  1.5811
#>  1.4737 -0.1296  0.1741  0.3487  0.4270  2.2563 -0.8423  0.5251 -0.0978  0.9572
#> -0.2359  1.4703 -1.1088  0.9901  0.8106 -1.1741  0.4954 -0.0741 -1.4070  1.4873
#> -1.5284  1.2176  0.2078 -0.8546 -0.2749 -1.0773 -0.4002  1.0111 -1.0308  0.1733
#>  0.2851 -0.6071 -1.5972 -1.6845 -0.0447  0.5275 -0.8179  0.0313  0.7890  1.2268
#> 
#> Columns 11 to 20 1.1979 -0.9284 -0.1536 -1.6745  0.8793  0.2642  0.7027  1.0175  0.1707  0.8144
#>  0.4780  0.3979 -0.6855 -1.2321  0.4189  0.7145  1.7671  0.1689  0.0348  0.8233
#>  0.2088  0.5810 -0.0111 -0.2119  0.1249 -0.8847 -0.4718  0.6737  0.2588 -0.6120
#> -0.8596  0.9901  1.5065  1.5140  0.3939  0.5499 -1.0707  0.5513 -0.6908 -0.2550
#> -0.2509 -1.3487 -0.9369  1.9951  0.6928 -0.3696 -0.7695  0.3083 -0.1014 -1.2543
#>  0.3493  1.9181 -0.1291 -1.6270  0.9057 -0.6219 -0.6827  1.8225 -0.6163 -0.3755
#> -0.1137 -0.7875  1.7484  2.4340  0.2553 -0.5718  1.8708 -0.2583 -2.2060 -1.2672
#> -0.7957 -0.4415 -1.1918 -1.3750 -1.2465 -0.1085  0.9110 -0.7262  0.6180  1.6529
#> -0.2853 -0.4212 -0.0086 -1.3761  0.3582 -0.1722 -1.0024  0.6337 -0.2389  0.2421
#>  1.3103 -0.1656 -1.4682 -1.0280  0.9851  0.6591  0.4406 -0.7370 -0.9352  1.4618
#> 
#> Columns 21 to 30-1.7070  0.3240  1.5246 -1.4222 -1.6898  1.9259 -0.6330 -0.3868  0.3976 -0.2530
#>  0.5013  0.2112 -1.4058 -0.8741  1.6740 -0.8692 -0.7819  0.8914 -0.3951 -0.8239
#>  1.6933 -1.1818 -0.5713 -0.1493 -0.3122 -0.3390 -0.1833  0.0989  0.1992 -0.2582
#>  0.1501 -0.8459 -0.0106 -1.0550  1.4526 -0.2242  1.0304 -0.2891  0.4190 -0.1798
#>  0.9115 -0.5622 -1.0989  0.8230  0.6390 -0.1360 -0.0610 -0.6350  0.1384  0.0561
#>  1.0744  0.4972  0.6161 -0.3063  0.8490  0.2255  0.5604  1.4924 -0.1004 -1.1842
#>  0.8681  0.4006  1.6428  0.8648 -1.4093  1.3758  0.2833  2.4920 -0.5461 -0.2139
#> -1.1181  0.1497  0.4784  0.6771  0.0307  1.8533  0.8432  0.7625  0.1460 -0.2766
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.0181
#>  0.6909
#>  1.4002
#> -0.2775
#>  0.5429
#> -0.0387
#> -0.7911
#> -0.6387
#> -0.7237
#> -0.3503
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.7237
#> -0.3503
#>  1.4002
#>  0.5429
#> -0.6387
#> -0.2775
#> -0.7911
#>  0.6909
#>  0.0181
#> -0.0387
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>