Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.3832  0.3757 -1.7820 -0.9403 -0.2578  1.1272 -0.2090 -0.5876  0.7918  0.5623
#> -0.7346  0.3725 -2.5947  0.3087  0.1456 -0.1801  0.8556 -0.0831  1.4210 -0.1603
#>  1.4346 -0.0181  0.1295 -2.4967  1.1335 -0.9923 -0.0702 -0.4500 -2.3220 -0.0974
#>  0.2042 -0.3565 -0.8727 -0.9573  1.0605  0.2102  0.5992  1.0815 -0.5524  0.1293
#> -1.7639  1.6133 -0.2659  0.3079  1.0210  0.7446 -0.8527  1.0705 -1.3854 -0.2073
#> -0.6933 -0.0988 -0.7756  0.7295 -0.1054 -0.6752  0.1174 -0.7770  0.8679 -0.6456
#> -0.7233  1.9003 -0.8408 -1.0239 -0.8244  1.2847 -0.3660 -0.4357  1.9856  1.1637
#> -0.2057 -0.4094 -2.6271 -0.1486 -1.4806 -0.0330  0.5784  0.5738  0.1435 -0.4151
#>  0.7487  0.0918 -1.1931 -0.9123 -0.7047 -0.2056  1.5052  0.3174  0.5179  0.4239
#> -0.9566  0.4548 -0.7667  0.7011 -0.8541  1.4729 -0.2416 -0.2423  0.1447  1.1747
#> 
#> Columns 11 to 20 1.1452 -0.0972  0.1581  1.0014 -0.4992 -1.1005 -1.1664  0.8151 -0.9911 -0.5563
#> -0.9777  0.2722 -0.6222  0.3564  1.3731 -1.6972 -0.4033  1.0989  0.3925  2.6512
#>  0.6308  1.5424  0.3364  0.3042 -1.5012  0.3343 -0.3403 -0.7121 -0.5652 -1.6640
#>  0.4157  0.5565 -0.3846  0.7348  0.0851 -0.5705 -0.2472  1.8984  0.1849 -1.2988
#>  2.3113  0.0184  0.0369  0.4404 -0.0650  1.4029  1.0360  0.1091 -0.8215  0.7929
#>  1.4782 -0.3575 -0.3739 -0.3914 -0.8995  0.1469 -1.7596  0.2358  1.2267 -0.8946
#>  1.6964  0.9234 -0.7311 -0.7342  2.5682  0.0853 -0.0961  0.2106 -1.1501 -1.7907
#> -0.2508 -0.7542 -1.0841  0.9237 -0.2126 -2.2639 -0.4368 -0.5109 -0.7023  0.6954
#> -0.1295 -0.0328  0.7424 -0.6954  0.3493 -0.6875 -1.3338  0.0236 -0.4263 -1.1374
#>  2.3249  0.2359  0.9225 -0.0042 -1.6021  1.6831  0.2530  0.0481 -1.2330  0.8737
#> 
#> Columns 21 to 30 0.6633 -0.0396  0.9274 -0.9415  0.0561 -0.9819  0.4492 -0.6339 -0.9968  0.5080
#>  1.0848 -1.0717  0.3259  0.0629 -0.8290  0.5985  1.3610 -1.1675 -0.8486 -0.7186
#> -0.2967 -0.6125 -2.3663  0.3152  0.5648 -0.4092 -0.6833 -0.1111  0.9008  0.9165
#>  1.0023 -0.7010 -1.2286  0.6006  0.1645 -0.0440 -1.3143  0.0739  0.7549 -0.3544
#> -0.5118 -1.2288 -1.2653 -0.6880 -0.6474 -0.2991 -0.3604 -0.9489 -0.2533 -1.4780
#>  1.9551  1.1765  1.3119 -0.6307  0.1829  0.7226  1.0549  0.4521  0.1569 -0.9561
#>  0.0347  0.1780 -0.3600  1.2480  0.7700 -0.0325  0.3203  1.9128 -0.5012 -0.1063
#>  0.2932  0.1153  0.7910 -1.2043  0.5387 -1.1120 -1.2846 -1.1045 -0.3117  0.1688
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.0431
#> -1.2457
#>  0.1956
#>  0.2035
#>  0.0465
#>  0.5178
#>  0.6680
#> -0.7679
#> -0.5719
#>  0.0528
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.0465
#>  0.5178
#>  0.2035
#>  0.1956
#>  0.6680
#> -1.2457
#> -0.0431
#> -0.5719
#> -0.7679
#>  0.0528
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>