Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 1.1164e+00 -3.3064e-01  1.5071e+00  4.5902e-02  6.0988e-01 -3.0253e-01
#> -3.3489e-01 -3.0798e-01 -1.3028e+00 -1.9119e+00 -1.3955e+00  1.3550e+00
#> -1.1646e+00 -1.4533e+00  4.9251e-01  1.1599e+00  4.6352e-01 -7.1828e-01
#> -2.8076e-01 -8.5921e-01  6.7653e-01 -3.4712e-01 -4.9409e-01 -1.5753e+00
#>  5.7367e-02  1.0660e+00  8.6956e-01  9.5275e-01 -9.9809e-01 -1.1265e-01
#> -7.5193e-01 -1.3826e+00  5.2502e-01 -2.5846e-01 -1.0821e+00 -4.1955e-01
#>  9.3742e-01  4.7443e-01  2.8379e+00 -8.0775e-01 -5.4684e-01 -3.8196e-02
#>  4.1847e-01  8.6421e-01 -2.2255e+00  2.2677e-01  4.8857e-01  5.4457e-01
#> -2.2925e-02 -4.5387e-01 -9.5194e-01  6.7255e-01  5.5701e-02  1.4670e+00
#>  1.1598e+00  2.2433e-02 -2.8177e-01  1.2721e+00 -4.5161e-01  1.0513e+00
#> 
#> Columns 7 to 12 4.6882e-01 -1.5638e+00 -1.6107e-02  4.0087e-01 -1.4204e+00  1.0913e+00
#>  1.4839e+00 -2.9597e-01 -2.9655e-01  1.8385e-02 -1.1001e+00 -1.8578e-01
#> -1.8331e+00 -1.3850e+00  1.8233e+00  3.8770e-01 -1.0423e+00  8.8752e-01
#> -6.8667e-01 -1.0641e+00 -6.4984e-02  1.1106e+00  7.2116e-01 -1.5075e+00
#> -1.6294e+00 -1.2864e+00  1.3188e+00  3.7370e-01  2.0915e-01 -4.9354e-01
#>  7.5512e-01  1.1305e+00 -5.2266e-01  1.0152e+00  3.1419e-01 -9.8664e-01
#> -2.8192e-01  2.9237e-01 -1.2640e+00 -1.4029e+00  4.0075e-01  5.7529e-01
#>  2.0055e+00 -7.1843e-01 -1.6072e+00 -5.9227e-01  3.5950e-01  2.7087e-01
#> -9.1492e-01 -5.1655e-01  2.3011e-01 -6.5434e-01 -4.9891e-01  9.7243e-01
#>  4.5352e-01 -2.1555e+00  9.9476e-01  2.6008e+00 -1.5096e+00 -9.4421e-02
#> 
#> Columns 13 to 18 9.6410e-01 -7.2680e-01 -1.3202e+00 -4.2303e-01 -9.5113e-01  3.2723e-01
#>  1.2501e+00 -3.8753e-01  1.7767e-01  9.4882e-01  1.1194e+00 -6.0511e-01
#>  2.3032e+00 -1.6017e-02 -6.9195e-01 -9.2308e-01 -9.0492e-02 -9.3677e-01
#> -6.4497e-01 -2.2071e-01 -2.4864e-01  4.2220e-01  5.5673e-01  6.1535e-01
#>  1.4666e+00  6.9939e-01 -8.4547e-01 -7.1894e-01  7.1443e-01  6.1286e-01
#>  8.8888e-01 -1.1342e+00 -7.9987e-01 -7.0769e-01  6.7247e-02 -2.0897e+00
#>  2.1521e+00  7.0980e-02 -1.7028e-01  2.5471e+00  1.6204e+00 -3.3256e-01
#> -7.8555e-01  3.7784e-01  6.2644e-01  3.0415e+00 -3.2352e-03 -8.8306e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.9461
#>  0.1357
#>  0.6791
#>  0.9512
#>  0.7742
#> -1.2008
#> -0.4232
#>  0.8343
#> -0.3844
#> -0.0039
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.0039
#>  0.9512
#>  0.7742
#> -1.2008
#> -0.3844
#>  0.6791
#> -0.4232
#>  0.1357
#>  0.8343
#> -0.9461
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>