Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.6309 -1.0709 -0.7854  0.1633  0.4796 -0.2011 -1.7913  1.9053  0.5345 -0.0029
#>  0.2078 -0.8543  0.0201  0.0603  0.4242 -1.2117  0.1083 -0.3842  1.1642  0.0903
#> -0.8217  1.1453  0.1575 -0.2022 -0.4091  0.8766 -0.3141 -0.8926  0.0999  0.1646
#>  0.0983  0.8260 -0.3693  0.0337  0.2243 -0.1194  1.5663 -0.2932  0.7247 -0.7522
#>  0.9845  0.9501 -1.3438 -0.5629  2.3981  0.1408 -0.6971 -0.8635 -0.2633 -1.9517
#>  1.6332 -0.4442 -0.6846 -0.1647 -0.7174 -0.2058  1.8295 -1.1684 -0.6237 -1.1590
#> -0.4216  1.0571  1.1323 -1.2560  0.1129  0.0790  1.1646  0.8973  0.3233 -0.3296
#>  0.4384 -1.6231  0.3949 -1.6581 -1.0688 -0.1946  1.1280 -0.2347 -0.3626 -0.1528
#> -0.2988 -0.5395 -1.6824  1.5670 -1.0152  0.8951 -0.1574  1.1205  1.0923 -0.9119
#> -0.2824 -0.1988 -0.4141 -0.3259  0.6681  0.8693 -0.0062 -0.7665  0.8299 -0.0584
#> 
#> Columns 11 to 20 0.0978  1.4239 -0.0919 -0.9945  0.3492  0.6480 -0.2004  2.1310  1.1208  1.3037
#>  0.3265  0.0597 -1.9034 -0.7129  0.1185 -1.1803 -0.6407  0.4970  0.7805 -1.2549
#>  0.0268 -0.7833 -2.7022  1.1250  0.4357 -1.3885  0.5197 -0.7955  0.4605  0.6631
#>  1.6939  0.0234 -0.3135 -1.0576  0.6375  1.0834  0.7042  1.1617  0.3404 -0.0916
#>  0.1701  2.1872 -0.8765  0.8634  0.4286 -0.8261 -1.7056 -0.0778  0.4651  0.5208
#> -0.6004 -0.7363 -1.8504  0.0007  0.2035 -0.5365 -0.8628  0.3929 -0.4298  0.8848
#> -0.5020  0.0058 -1.6460 -0.1157  0.2478  1.1058  2.0826  1.4355 -0.8454 -0.0646
#> -1.6124 -0.5712 -0.5393  1.3793 -1.1770  0.5932 -0.2320 -1.1323 -0.5651 -1.3765
#>  0.0497 -0.2189  2.2703  0.7086 -1.9724  0.2205  0.6920 -1.4505  0.3589 -2.4307
#>  1.2002  0.0644 -0.0736  0.4385  0.7088  0.6364 -0.1166 -0.6049 -1.9749  0.9176
#> 
#> Columns 21 to 30-0.0910 -1.5840  0.2101 -0.7864 -0.0519  1.2549  0.3184 -1.8225 -0.3473 -0.4020
#> -0.4322 -0.5340  1.3127 -0.1528  0.0020 -0.4506  1.4646  0.3542 -0.2125 -0.7440
#>  1.1845  0.7184  1.2206 -0.4541 -0.5739  0.1210  1.0628 -0.5248 -0.5389  0.8496
#>  0.9332 -0.0431  0.3945 -0.7800  0.4573 -0.4841 -0.1204 -1.4719  1.9513 -0.7550
#> -0.2569 -0.4834  0.6135 -0.0714 -0.3360 -0.8666 -0.1596 -0.6362  0.0225 -0.8190
#>  1.3187  1.7297 -1.2198  0.6323 -0.3252  0.2035  0.0227 -0.9477 -0.1115 -0.3241
#>  0.2267 -1.2203  0.4793  0.6394 -0.7236 -0.2301  0.8272 -0.5652  2.4154  1.0198
#>  1.7744  0.4793 -0.9683  0.1447 -0.7204 -1.5253  1.7043  0.6079 -1.1229 -0.4561
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.0111
#> -1.0959
#> -1.3503
#>  0.7063
#> -0.1707
#>  1.3496
#> -1.1074
#>  0.3132
#> -1.2930
#>  0.8074
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  1.0111
#>  0.7063
#> -0.1707
#> -1.1074
#> -1.0959
#>  0.3132
#> -1.3503
#>  1.3496
#> -1.2930
#>  0.8074
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>