Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-2.2742e+00 -4.3856e-01 -3.1202e-01 -6.8002e-01  1.3930e+00  1.8106e+00
#> -1.1820e-01  4.4465e-01 -3.8090e-01  8.3434e-01 -6.0633e-01 -6.6036e-01
#>  1.3166e+00  6.4660e-01  8.7875e-01 -4.9884e-01 -6.8039e-01  4.4428e-01
#>  9.5649e-01  3.8333e-01 -5.0071e-01  1.3603e+00  5.2972e-01  1.3124e+00
#> -1.3423e+00  8.6793e-01 -1.0591e-01 -1.2607e-01  4.1793e-01  1.1150e-01
#> -9.8199e-02  1.3093e-02  7.9021e-01 -5.6023e-01  6.8929e-01  8.4208e-01
#> -6.6575e-01 -4.5649e-01 -1.1848e+00  5.4164e-01 -6.4244e-01 -2.2084e-01
#> -2.4722e-01 -4.1652e-01  1.5945e-01  4.8361e-01  5.9850e-01  7.0767e-01
#>  1.8022e+00 -6.9823e-01 -6.4734e-02  5.7230e-01 -3.0274e-01  1.9269e-01
#> -4.0200e-01 -5.3684e-01  1.1800e-01 -7.3915e-01 -7.1443e-01 -8.0102e-01
#> 
#> Columns 7 to 12 2.1991e-01  2.3001e-01 -9.8211e-01 -3.7914e-01 -1.5411e+00  1.2909e+00
#>  1.0509e+00 -7.6430e-01 -3.6411e-02  1.5104e-01 -2.3064e-01 -6.3914e-01
#>  1.1574e-01  4.9109e-01  1.3314e+00  2.2266e-01  1.7234e+00 -3.1267e-01
#>  2.9727e+00  5.5781e-01  6.0781e-01 -7.9646e-02  1.0893e+00  1.1855e+00
#> -3.9380e-01 -7.7849e-01 -5.2227e-01  1.4351e+00 -1.0459e+00 -1.3716e+00
#> -1.1112e+00  4.2023e-01 -6.7848e-01 -7.2601e-01  5.1148e-01 -4.0055e-01
#> -4.7833e-01 -4.5492e-01 -3.1419e-01  6.6482e-01 -1.5464e-01 -6.3259e-01
#> -1.5553e+00  7.8386e-01  3.4969e-02 -1.0571e+00 -8.7297e-01  1.4117e-02
#>  1.4913e+00  9.1385e-01  8.6206e-03  9.6383e-01  1.2117e+00  1.1528e+00
#>  4.1185e-01 -8.1497e-02 -1.2321e+00 -1.7549e-01  7.9203e-01  1.0850e+00
#> 
#> Columns 13 to 18-1.2274e+00  5.5193e-01  7.2007e-01 -1.4389e+00 -7.9695e-01 -2.2775e-01
#>  1.6421e+00 -7.3536e-01 -9.2900e-01  8.9276e-01 -1.0425e+00 -1.0390e-01
#>  9.1798e-01  7.8415e-01 -2.2824e-01 -8.1196e-02  1.6979e-01  7.1401e-01
#> -1.4483e-01  1.6163e+00 -7.1815e-01  5.1317e-01 -1.2311e+00  4.2962e-02
#> -1.1840e+00 -1.2477e+00  4.7300e-01 -9.0966e-01 -7.6165e-01 -2.7549e-01
#> -3.0283e-01 -1.6414e+00 -4.4684e-02  5.8112e-01  3.1318e-01  1.0429e+00
#> -5.1813e-01  1.0005e+00 -5.6379e-01  1.6112e-01  1.4051e+00 -2.0184e-01
#>  1.3978e+00  4.0883e-01 -5.0532e-01 -1.3213e+00 -2.5157e-01  1.3266e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.5871
#>  0.9745
#>  0.3772
#> -0.7096
#> -1.6641
#> -0.3057
#> -2.0163
#>  0.2686
#>  0.6565
#> -0.2270
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.6641
#> -0.5871
#>  0.6565
#> -0.7096
#>  0.9745
#>  0.3772
#> -2.0163
#> -0.3057
#> -0.2270
#>  0.2686
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>