Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-4.5064e-01  2.7403e-01  1.1819e+00 -1.1495e+00 -4.0606e-01 -1.7401e+00
#>  1.3267e+00 -8.6428e-01  1.6693e+00 -9.4829e-01 -1.1844e+00  1.3886e+00
#>  2.3062e-01 -5.8445e-01 -4.6561e-01 -1.2403e+00  7.9463e-01 -2.4332e-01
#>  6.9625e-01  1.2031e+00  1.2340e+00  1.5485e-01  7.8099e-01 -1.5724e+00
#> -1.7571e+00  2.2330e-01  2.7117e-01  2.0501e-01 -5.2698e-01  7.9407e-01
#>  1.9822e+00  2.5416e-01 -1.2153e+00  3.4645e-01  6.3213e-01 -1.2206e-01
#> -1.8482e+00 -8.9096e-01 -5.7454e-01 -3.7890e-01  1.1905e+00 -4.8498e-01
#>  3.8349e-01  2.2780e+00 -2.5071e-01  9.9074e-01 -2.9681e-01 -2.0075e-01
#>  1.3113e+00  1.3221e+00 -1.2217e+00 -6.5114e-01 -6.6539e-02 -2.1964e-01
#>  6.8650e-01 -1.0359e-01  4.0716e-01 -1.1701e+00  1.6621e-01  1.1586e+00
#> 
#> Columns 7 to 12-9.8762e-01 -1.6388e+00 -1.2514e-01  1.5619e-01  1.9668e-01 -9.2782e-01
#>  9.2512e-02  7.4998e-01 -1.0095e+00 -1.2709e+00 -1.2687e+00 -1.8259e+00
#> -5.8583e-01  2.0116e-01  6.8236e-02  3.7958e-01  6.8914e-01  4.2434e-01
#>  3.6073e-01 -1.3012e+00  1.2337e-01  1.9477e-01 -3.8940e-01  6.5488e-01
#>  4.0920e-01  7.3387e-02  1.9697e-01 -1.3660e+00 -1.8476e+00 -1.1190e+00
#> -2.1471e+00 -9.0785e-01  3.4046e-01  1.1239e+00 -4.1914e-01 -1.0531e+00
#> -1.3291e-01 -8.3302e-01 -3.7567e-01 -1.0600e+00 -6.3342e-01  1.2768e+00
#>  2.7415e-01 -6.6700e-01 -5.1957e-01 -5.5411e-01  4.8176e-01 -1.3552e+00
#>  8.5042e-02 -1.0548e+00  9.1145e-01  8.0412e-01  1.2262e+00 -9.5917e-02
#>  1.0375e+00  8.9151e-01  9.4625e-01  7.0754e-01  8.7061e-02  1.6959e+00
#> 
#> Columns 13 to 18 1.3229e+00 -1.1767e+00  1.1779e+00  4.6743e-01  4.2296e-02 -7.9049e-01
#>  6.2424e-02 -3.1265e-01  1.3873e-01  3.4917e-01 -5.6423e-01 -8.1010e-01
#> -1.7127e+00 -6.6570e-01  3.7632e-01 -2.7629e-01  1.5819e+00  4.5437e-01
#> -1.0876e+00 -9.7795e-02 -6.4690e-01 -4.5880e-01  3.8072e-01  5.3344e-01
#> -1.0116e+00  1.2681e+00 -1.1659e+00  7.6828e-01  7.6063e-01  3.4054e-01
#>  6.8393e-01  2.7849e+00 -4.2305e-01  1.2052e+00 -2.8858e-02 -9.3545e-01
#>  1.2186e+00  2.3603e-02  1.0512e+00  1.5537e+00  7.3489e-01 -1.0721e+00
#>  1.4042e+00  4.3918e-01 -1.0844e-01 -1.4868e+00  6.9436e-01 -1.5459e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.0886
#> -0.5328
#> -0.4580
#>  2.3141
#> -0.7785
#>  1.7995
#>  0.9443
#> -0.0880
#>  1.0736
#> -1.1235
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.5328
#> -0.7785
#>  1.0736
#>  1.7995
#> -1.1235
#> -0.4580
#> -0.0880
#>  0.9443
#>  1.0886
#>  2.3141
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>