Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.2350  0.3498  1.0195 -0.3418  1.3656 -0.5723  0.8518  0.7920 -0.3594 -1.7229
#>  0.8632 -0.8368  1.0108 -0.3757  0.3914  0.6564  0.4138  0.6199  0.5032  0.2086
#>  0.1175 -1.5238  1.1072 -1.5739  0.7266  1.1696 -0.5831 -0.7631 -2.5049  0.9934
#>  0.3733  0.1346 -0.5880  0.8714  0.4914  0.4581  0.6996 -0.2689 -1.2482  1.7361
#> -0.8272  0.1650  0.6978  0.3497  0.4222  1.2581  1.7176  0.1079 -0.6470 -0.3341
#> -0.6676 -0.0921 -0.4210  0.5137 -0.0063 -1.2522  0.4532  1.6599 -0.5210 -0.5020
#> -0.6599 -0.9676  0.3322 -0.0107 -0.7798 -0.2118 -0.9815  0.2745  0.2905  0.2896
#> -0.2868  0.5271 -0.1984 -0.4264  0.8942  0.2753  0.4062  0.5139 -0.3036 -1.1289
#> -0.6191 -0.2912 -0.8277  0.9391 -0.9296 -0.9067 -1.2938 -0.9630  2.1618 -0.5012
#> -1.4822  0.1940  0.4396  1.5408  0.3835  0.7108 -0.2801 -0.3080 -0.4454  0.9517
#> 
#> Columns 11 to 20-0.3896 -2.0683  0.6876  1.1451 -0.2786  0.7871  0.6952 -0.2198 -1.1245  2.5164
#>  0.1129 -0.0319  0.5236  0.2116  0.4691  0.8519  0.1989 -0.4917 -0.3429  0.1124
#>  1.6249  0.3706  0.9376 -0.2862  0.6115  0.5790  0.5247  0.5560  0.7518  0.6729
#>  0.1406  0.3539 -0.4036 -1.4701  1.0563 -1.2082 -0.1629  0.6826  0.0305 -0.6474
#>  0.7554  2.2712  0.0209  0.2945  0.2289  0.9642 -1.3118 -0.7058  0.3611 -2.6792
#> -0.5293 -0.4167  0.8634 -0.2204 -0.0069  1.1930 -1.3298  1.2476  1.0205 -0.6738
#> -0.5866 -0.8377 -0.0784 -0.5555 -0.6860  0.2216  0.1338 -0.1057  0.1886  2.0645
#> -0.2577 -0.5098 -0.1766 -0.1776  1.1494 -0.3178  0.1810 -0.7389 -1.0089 -1.2237
#> -1.3506  0.3009  1.1411  0.7817  0.4784 -0.1432 -1.4294 -0.5635 -1.3388  0.7327
#>  1.2856  1.2440  0.6160 -2.1123 -0.5798  1.3497  0.7890  0.5421 -0.4942 -1.2586
#> 
#> Columns 21 to 30-1.5377  0.4248 -0.3839  0.0434 -0.0877  1.2582  0.2292  0.1689  0.5296  0.4780
#>  0.4906  0.8391  1.0834 -0.0206  1.3310 -0.1178 -0.1709 -0.1599  1.2315 -0.2860
#>  0.2233  1.1960 -1.7514  0.5184  1.9036  0.5852 -0.2466 -0.4191 -1.6440 -0.5973
#> -1.6804 -0.1455  1.3004  0.4991 -0.5285 -0.2174 -0.0348  0.9081  0.6589 -0.6149
#> -0.5061  1.2322 -0.2202 -0.7180  0.2548 -0.6482  0.7224 -0.2748 -0.9589 -0.2262
#> -1.2880 -0.3862  0.9343  1.0858 -0.7028  1.8633 -0.2980 -2.4976  0.7817  0.2419
#> -1.5938  0.3248  0.4746  0.9344  1.0752 -0.5121  0.8797  0.4647  0.4049  0.7603
#>  0.5809  0.0102 -0.3521 -0.1084  2.0716  1.0153 -0.7117 -0.6187 -0.1130  0.0826
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.3976
#>  1.2447
#> -0.0480
#> -0.0961
#> -1.2174
#>  1.5903
#>  0.4806
#> -0.5285
#> -0.4042
#> -0.5256
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.2174
#> -0.0961
#>  1.2447
#> -0.0480
#> -0.5256
#>  1.5903
#> -0.3976
#> -0.4042
#>  0.4806
#> -0.5285
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>