Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 1.1788  1.4670 -0.3657 -0.7143  1.0326  0.8401 -0.2818  0.3147  0.0966 -2.3724
#>  0.3963 -0.3477  1.3797  0.1176  1.5528  0.8718 -0.1038  1.0266 -0.3956 -1.2384
#>  1.0193 -0.4533  0.7315  0.7390  0.5915  2.5774  0.6173 -0.8025 -0.6404 -1.2933
#> -0.8738  0.1810  0.0059 -0.3364  0.1677  0.1269  0.0966  0.9770 -1.0869  0.4863
#> -0.9647 -0.5238  0.9967  0.7324  0.3659 -1.4491 -0.4452  1.1533 -1.5509  1.8984
#>  1.3263  1.4210  0.9498  0.5385 -0.9353  0.6058 -0.5343  0.9524  1.1091  0.0215
#> -0.5810 -0.9676  0.3519  0.7088  1.4821  0.0635  0.1029 -0.9950 -2.0403 -1.2639
#> -0.2265 -0.0999  1.4882 -0.3843  0.1630  0.3746  1.2733 -0.2076 -0.7805 -0.0565
#> -2.3493 -1.5622 -1.5535  0.4987 -0.2788 -0.4790 -1.2299  0.0902 -0.1817  1.9422
#>  0.0393 -0.6391  0.6510  0.3130  0.2733  0.5300  0.2031 -0.4220  0.4970  0.1105
#> 
#> Columns 11 to 20-0.3484 -0.3550  0.1957  0.8019 -1.1656  1.2486 -0.6442  1.1571 -1.1284 -1.0227
#>  0.1625 -0.6666  2.3821 -0.0843 -0.6865 -2.8304  0.0487 -0.4505 -1.8141  1.1077
#> -0.7039 -0.6959 -1.1493 -0.6216  0.2387 -0.2800 -1.5938 -0.7761  0.6853 -0.2394
#>  0.4090 -0.0800  1.1340  0.3174  1.1599  0.5525  0.6526 -0.6451 -1.5417  0.5491
#>  0.9251 -1.1067  0.8761 -0.6947 -0.4800  0.0860 -0.3353 -1.0206 -0.1809 -0.8180
#> -0.3031 -0.2824 -1.5244 -1.1668  0.5962  0.7707  0.7102 -0.2345 -0.3460  1.3174
#> -1.5474  0.1482 -0.3664  0.0388 -0.4264 -0.5528  1.0447  0.8300  1.3396  0.8471
#> -0.1281  0.2672  0.3959 -1.1890 -0.1681  0.9322 -0.8535 -0.1103 -0.8844 -0.0530
#>  1.8013 -0.2656  0.0482 -0.3111  0.1418 -1.2753 -0.4779 -0.7448  0.4216  0.8460
#> -1.4902 -0.0785 -0.2396 -0.2596  0.6598  1.6172  0.5653 -1.6612  0.7273 -0.8822
#> 
#> Columns 21 to 30-0.2120 -0.9546 -0.1494  0.0443  0.2149 -0.5224  0.1274 -1.1670 -1.5838 -0.5471
#> -0.1093  2.7348  0.4241 -1.1301  0.6156  1.4233 -1.2648  0.4278 -0.4462 -0.1017
#> -0.5719 -0.1362 -0.5651 -0.5884 -0.0016 -0.4822  1.0851 -0.0961 -0.2505  0.5132
#> -0.5004  1.5237 -0.2745 -0.0934 -1.0723  0.2452  0.3877 -0.0657 -0.2149  0.1890
#> -1.4155  1.2233  0.7619  0.6000  0.8233  0.3493 -0.3782 -1.4433  0.4839 -2.5128
#> -0.3185  1.1704  0.5761 -1.1746 -1.9996 -1.4041 -1.7882  1.3245 -0.1606  1.3170
#> -0.6456 -0.0388  0.0020  0.3745 -1.0361 -2.1813  0.1489  0.2077  1.0806 -1.4377
#>  0.3813 -0.4322 -0.4156 -0.6676 -0.1218  0.4124  0.3893  0.3430  0.2438  0.8794
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.7466
#>  1.9374
#> -0.1116
#> -0.2129
#>  1.8581
#>  0.6654
#>  1.0991
#> -0.2130
#>  0.1538
#>  1.4219
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  1.7466
#>  1.9374
#> -0.1116
#>  1.8581
#>  0.1538
#>  0.6654
#> -0.2130
#> -0.2129
#>  1.4219
#>  1.0991
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>