Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.2048 -0.2939 -0.4795  0.7202 -1.9698 -0.5648 -0.5866 -0.6413  0.9305  1.9333
#>  0.9045 -0.9143 -0.2110  0.8027  0.7300  0.0773 -0.8567 -0.5894  0.3113  1.6430
#> -0.6045 -0.3825  0.7626 -0.6181  0.4820 -0.8466 -0.3048 -0.9057  0.9136  0.2024
#>  0.3312  0.4599 -1.7967  0.0403  0.4394  0.7393 -1.6227  0.6240 -1.5762  0.1226
#> -0.5010 -2.0228 -1.6562  0.5454 -0.1313 -0.4892 -0.3192 -0.5671 -0.1744 -0.1895
#> -0.0461 -1.6718 -0.7552 -0.4499 -1.0773  0.9616  1.8802  0.1860  0.4674  0.5514
#>  0.6868 -0.0697  1.1523 -0.9649 -0.3945  0.5363 -0.4977 -2.0084  0.6512 -0.2238
#>  0.7669  0.8621 -0.1225  0.7797 -0.5376 -0.2825 -0.2409  0.9736  1.3590 -1.3354
#>  1.4358 -0.7556  1.0264  1.2301 -0.1581  0.6879 -0.4055  0.3461  0.6769 -0.1283
#>  2.0706 -0.0046  0.0579 -1.2839  1.6485  0.6679  0.7761 -0.8188  0.3009 -0.0210
#> 
#> Columns 11 to 20 1.9280  0.6421  0.7008 -1.1458  0.3837  0.0864  0.0765 -0.1160 -0.3574  0.4574
#>  0.1827 -0.3866  0.1401  0.4019 -0.6840  0.5145 -0.3349 -1.1971 -0.2077 -0.5337
#> -0.2440  0.5585  0.7847  0.1803  0.1965  0.3832  0.3531  1.2150 -0.8197  1.2251
#> -1.3122  0.0596 -0.7007  2.2918  0.1105 -0.4351 -0.6855  0.4967  0.2033 -2.1407
#> -0.0857 -0.0749 -0.4728  0.1802 -0.0085 -0.5427 -0.7939  1.1635 -1.0073 -0.7733
#>  0.2149  0.0528 -0.3490  0.4457 -1.0258  0.9238 -0.7917  0.3490 -1.1951 -1.2585
#>  1.0286 -0.8566 -0.8786 -0.0829  0.9796  1.5668 -0.0168 -0.5049 -1.3459  0.7173
#>  0.3963  0.2241 -0.2311 -0.2833  1.7688  0.4730  0.5866  0.0211  1.4454 -0.0584
#>  0.2412 -1.7240  0.1192 -0.8602  1.1761 -0.0433 -0.4662  1.5076 -0.5863 -1.1663
#> -0.7777 -0.8199 -1.0757  0.6246  1.2080 -0.1261  0.3615 -0.7307  1.0115 -1.1734
#> 
#> Columns 21 to 30-1.1169 -0.2876 -0.6757 -1.5989  0.4886 -1.5605 -0.5425 -0.8179 -0.1108  0.3376
#>  0.1162 -0.2143 -0.4168  0.9563 -0.3560 -1.8929  0.6506  0.6660 -1.1291  0.8782
#>  0.8103 -1.6647 -0.7219 -1.1634 -0.9293  0.3484 -0.3272 -0.0585  0.0722  0.6267
#>  1.5560  0.1253 -1.0253 -1.0519  0.8381  1.3364 -2.2204  0.9114 -0.6021 -0.5660
#> -0.7026 -1.6809 -0.0351 -0.4092 -0.1876  1.2005  0.5559 -0.1405 -0.6131  2.7969
#>  0.5357 -0.0759  0.7964  1.4892  1.5510 -0.9863  0.8020 -0.1832  0.3301 -0.6788
#> -1.6506  0.8557 -1.1140  1.4475  0.2911 -1.5374  0.2496  0.4259 -0.7849  0.7098
#>  0.8960  1.2554 -0.5609  1.6191  0.5086 -1.1001  0.4482  0.5483 -1.3620 -1.6155
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.2055
#> -0.1858
#> -0.3353
#> -0.9608
#> -0.4442
#> -0.7862
#>  0.1821
#>  0.1036
#> -0.9173
#>  0.5370
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.9173
#> -0.4442
#>  0.2055
#> -0.9608
#> -0.3353
#> -0.1858
#> -0.7862
#>  0.1036
#>  0.5370
#>  0.1821
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>