Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 2.6625  0.3853  0.1934  0.7297 -0.1950 -0.0242  1.6568  1.7336 -0.5521 -0.0177
#> -0.9447  0.1813  0.9366  0.7792  0.0084  0.4351  0.8051 -0.9465 -0.2840  0.8448
#> -0.7903  0.6576  0.9979  1.4207 -1.2233  0.6388  0.0440  0.0329 -0.9858 -0.0834
#>  0.4505  0.5239 -1.0068  0.5570  0.0211  0.2595  1.3417 -0.9829  0.0678 -0.4293
#>  1.1004  1.2353  0.8260 -1.2920 -0.2842  0.5580  0.3703  1.8449 -0.0870 -0.0237
#>  0.0502 -0.6411  0.1013  0.8128 -0.7832  2.1553  0.5556 -0.9129 -0.0755  0.4454
#> -0.7110  0.2072 -0.8815  0.6118 -0.8958  1.0720  0.8396 -1.2150 -2.5525 -0.6068
#> -0.5055 -1.8374  0.3430  0.4955 -0.8371 -0.6309  1.4100  0.0556 -0.1323  0.1375
#>  1.6693 -1.4627 -0.2675 -0.4742 -0.2029  0.2553 -1.5411  1.4665  1.0864 -0.6217
#>  0.4342 -0.4212  2.2097 -0.0430 -1.3236 -0.2203 -0.1575 -1.9859  0.7315  0.4399
#> 
#> Columns 11 to 20-0.0355 -0.8303 -1.8310 -0.9789  0.0287  1.5046 -0.0229 -0.6890  0.6211  0.7080
#> -0.6214 -2.2984  0.6443  1.1477 -0.7392 -0.8650  0.1470  0.4511 -0.9449 -2.2172
#> -0.4221 -1.5162 -0.8012  0.8106 -2.0280 -0.3390 -0.1243 -0.4690 -0.7840 -0.1079
#>  0.1671 -0.1238 -0.0176  1.3500 -1.0305 -0.2203  1.5085  1.4288 -0.2824  0.4975
#> -1.2804 -0.0891 -0.3699 -0.8303 -1.0208  1.4891 -1.5748 -0.7493  0.6770  1.6014
#> -0.1848 -0.8562  0.7612 -1.8687  1.0196 -0.3396 -0.9749  0.3485  0.2017  0.3874
#> -1.4239 -0.2845 -1.6458  0.4783 -0.2280 -0.2107 -0.9370 -0.5795 -0.9035  0.7907
#>  0.9405 -0.5830 -0.1564 -0.1063  0.1782  1.4519  0.4857  0.0179  0.7876 -0.0978
#>  0.3160 -1.5651  0.3633 -0.5317  0.1127  1.0000  1.1014  0.6219 -1.2383 -0.8433
#> -0.7594 -1.6102  0.1655 -1.5627  0.0731 -0.9243 -0.6874  0.9381 -0.9301  0.7629
#> 
#> Columns 21 to 30 0.6197  0.3510  1.0710 -2.0288  0.2645  0.6579  1.1783 -0.4299 -0.5379  0.6714
#> -1.6353 -0.5961 -0.4186 -0.1671 -1.3017  0.4569 -0.3605 -1.3208  0.1779  0.5833
#> -1.3851 -0.0024  0.4040  0.2756 -1.1018 -0.3256 -0.1045 -0.0354  0.3063 -1.1113
#>  0.0106  0.0657  1.1337  0.2539 -1.3990  0.8490 -0.8886 -0.5097  0.4214  0.8079
#>  0.7927 -0.2953  0.8337  2.3895 -0.1380  1.3501 -0.9694 -0.4464  0.1799  1.8440
#>  0.7615  0.0574 -0.5371  0.1183  0.5175 -0.4823  0.2619  0.6492 -0.9772  0.5389
#> -0.5113 -1.1126  0.3308  0.0811 -1.2742  0.3893 -0.9253 -0.2572 -0.5847 -0.2962
#> -0.2261 -0.7006 -0.1055 -0.1027 -1.4204 -0.1636  0.8856  0.6951  0.8050  0.4794
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.1329
#>  1.0845
#> -0.8780
#>  0.7792
#>  0.2463
#>  0.0885
#>  1.5917
#> -0.0851
#>  1.5553
#>  0.6264
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.0885
#>  1.0845
#>  1.5917
#> -0.0851
#>  1.5553
#>  0.6264
#>  0.7792
#> -0.1329
#>  0.2463
#> -0.8780
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>