Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.8388 -0.1471  0.1097  1.6436  1.3128 -0.4051  1.5187  0.8309  0.3212 -0.5679
#> -0.6954 -0.8294  2.7801  0.7230  0.1532 -0.3798  0.1140 -1.0874  0.2542  0.5661
#>  2.2783  0.0799  0.5876 -0.1017  0.4808  0.9400  0.7175  0.0565  1.2281  0.0201
#>  1.7500  0.9018  1.1081 -0.4190 -0.7407 -0.2910 -1.2944  0.8223  1.2762 -0.8313
#>  0.1800 -1.1927  0.3044 -0.4467 -1.4173  0.2514  1.4129  0.0960  0.6265 -0.8949
#>  0.2792  1.3013 -0.1224  2.0214  1.1186 -0.7001 -0.5429 -0.3636  0.6481  0.7344
#>  0.7802 -0.0870 -0.3569  0.5338 -1.3881 -0.7810 -0.1274  1.6686  0.4437 -2.5279
#> -0.8837  0.6725  0.1419 -0.7596  0.0892  0.4828  0.0837 -0.1266 -0.9111 -0.0697
#>  0.4717 -0.3290  2.5163  0.1817  1.5818 -0.7564 -1.0554  0.2482 -0.5283  0.4709
#> -0.2986 -0.1546  0.5405  0.8302 -0.3231 -0.3083 -1.1648 -1.2471  0.5237 -0.1486
#> 
#> Columns 11 to 20 0.2727  1.2699  0.1713 -0.3182  0.5910  0.5465 -0.1231  0.0009 -0.2081 -0.5437
#>  0.6754  0.7646 -0.0685 -1.1823  0.1017 -0.7200 -0.1523  1.4860  0.3436 -1.9730
#>  1.1430 -0.5927  0.8823  1.1834  0.0876  0.8238  0.0991  1.5732  0.4597 -0.6421
#> -0.4729 -0.2984 -0.0400 -0.3871  0.2453  1.7366 -0.6961 -0.1785 -0.2437 -0.3637
#>  1.5198 -0.2517  1.2212  0.2570  0.9639 -1.0084 -0.8773  0.2175 -0.4350 -0.6661
#> -0.1868  0.0505  0.8357 -0.3817 -0.3384  0.0998  1.0655 -0.1832  0.5041  0.1139
#>  0.5255  0.1193  0.5777  0.9852  0.1386  0.5317  0.8167 -1.5185 -2.0595  0.0531
#> -0.2753 -0.8017 -0.0297  0.8124  0.4591  0.1694 -1.0554 -0.6397  0.3336 -0.2268
#>  0.2080 -0.4048  0.3972  1.1689 -0.8711 -0.0200 -0.2119  1.3721  1.2337 -0.9065
#>  0.2234 -1.1390  0.3515  0.7074 -0.1015  0.7105 -0.2372 -0.3942  0.2199  0.6481
#> 
#> Columns 21 to 30 1.1764  0.1624  0.6420  0.0201  1.0500  0.6198  0.1663  0.1318 -0.5397 -0.3018
#> -0.3404 -0.8539  0.1045 -0.8704 -0.0791 -0.6554  0.3694 -0.5862  0.5364 -0.5479
#> -0.8251  0.3864 -0.0365  1.1942 -0.0436  0.2761  0.0079 -0.1690 -0.0297 -0.1689
#> -1.3571 -0.4883  1.4461 -0.5396  1.4952 -0.7551  0.5560 -0.6053 -2.0049  1.0421
#>  1.4065 -0.8630  0.7240 -2.1863  1.0683  0.8637  0.3342 -1.2185 -1.4461  1.1135
#> -0.8200  0.9331  1.6923 -0.2750 -2.3291 -0.8556  1.5093 -0.4848  0.5352 -0.5032
#> -1.0352  0.8014 -0.6603 -1.1477 -0.8246 -2.4030 -1.2437 -1.2736  0.3930 -0.1299
#>  0.3961 -1.1269  0.5512  0.3408 -0.4455  0.0269 -0.3414 -0.7978  2.1611 -0.7009
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -1.4118
#>  1.8059
#>  1.1682
#>  0.6860
#>  0.6175
#> -0.6389
#>  0.5179
#> -0.3830
#>  1.7077
#>  0.5071
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.4118
#> -0.6389
#>  1.7077
#>  0.5179
#>  1.1682
#>  0.6860
#> -0.3830
#>  1.8059
#>  0.6175
#>  0.5071
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>