Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.2939 -0.1716  0.4528  0.5083  0.7136 -0.7393  0.6000  0.2201 -0.9965 -0.7268
#> -0.2705 -0.2382  1.2585 -0.0272 -0.4097 -1.3051 -1.2948  1.4649  0.7435 -0.1294
#>  0.7967 -0.2589  0.0420  1.3873  1.1764 -0.6935  1.1186 -0.7974  1.7014 -0.7185
#>  0.3574  0.9930  0.5721 -1.1953 -0.3471  0.3122  0.6877  0.0949  0.0931  0.2170
#>  0.8491 -0.1154  1.4497 -1.2286 -1.0975  0.1036  0.3457 -0.5101 -0.1182 -0.0926
#> -0.1602  0.3143  0.7377  0.4662  0.3118 -0.3363  0.5983 -0.1225 -0.5773  0.1370
#> -0.7525 -1.2596 -1.2079  1.1861 -0.3735 -0.8022 -0.3604  0.2238 -0.7980 -0.8232
#> -1.7860  0.9214  0.2053 -0.0487  0.0867 -0.5380 -0.1493  0.3550 -2.3782 -0.7187
#>  0.1488 -0.0805 -0.3137  1.5757  1.6169  0.1133  1.0179 -0.6022 -0.8372 -1.4353
#>  0.1194  0.9009 -0.2756 -1.8402  0.9931  0.4968 -1.2867 -0.6867  0.7418 -1.0792
#> 
#> Columns 11 to 20-0.8721  0.7000  0.1876  0.7881  1.2541 -0.1432  1.9267  0.7669 -0.5642 -0.4031
#> -0.0147  0.6991 -0.0026  1.1447 -0.2631  0.6545 -0.2257 -0.7556  1.0162  0.2604
#>  1.2487 -0.2463  0.2167  0.2916  0.5012  0.4156 -0.1843 -0.6420 -0.4648 -0.6364
#>  0.3798  0.4222 -0.9439 -0.4301  0.8694  1.4714  0.5623  0.2659  2.4340 -1.1046
#> -0.0378  0.4548 -0.2055 -1.3030  1.0360 -2.1157 -1.0829  0.8692  0.7547 -1.5108
#> -0.3935 -1.0189  0.3949  0.6594 -1.3401 -0.7600 -0.5853  1.0938 -0.3371 -0.6840
#> -1.1442  0.5889 -0.4083  0.1345  0.8838 -0.2139  0.9548  0.1798  0.3408 -1.2909
#>  0.2184  2.8148  0.9084  0.1117  0.1970  2.5532  1.4172  1.6306  0.3397 -0.2592
#> -0.1005  1.0207  0.2532 -0.3932 -0.5646  0.8024 -0.7494  0.1033 -1.8491 -0.0167
#> -0.4346  1.5261 -0.7315  1.2414  0.1774 -0.2861 -1.3551  1.1710 -0.1492  0.7812
#> 
#> Columns 21 to 30 1.5354 -0.7741 -0.2682  1.6360  1.3899 -0.0308  0.1538 -0.3024  0.1302 -2.6203
#> -0.0653 -0.8395  0.1961  0.9587  0.0344  1.7035 -0.0634  0.1900 -0.5460  0.8725
#> -0.0935  0.3076 -1.3328  1.0010 -0.9201 -0.6827  0.3742 -1.2926  0.4828 -1.3111
#> -2.0340  0.2135  0.3899 -0.6661  0.4662  0.2157  2.5137 -0.0331  0.1229  1.6645
#>  1.8071  0.1896 -0.2106  0.1256 -0.0706  1.3270 -1.0926 -1.0337  0.9058 -0.4991
#>  1.7652 -1.7103 -0.3695  0.3398  1.3955  0.3456  0.2389 -0.4748 -0.2237  0.7654
#> -0.6945 -0.3795 -0.0389 -0.4402  1.1743  1.0984  1.7154  0.2803 -0.2272  1.5774
#>  1.2344  0.9816  0.5143  0.4696 -0.0544  0.3288  0.5246 -1.0121 -0.9106  0.3072
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.1661
#>  0.8975
#>  2.0945
#> -0.9164
#>  1.1669
#> -1.2365
#> -0.6173
#>  0.4757
#>  0.4486
#> -0.9626
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.9626
#> -1.2365
#>  2.0945
#> -0.6173
#>  1.1661
#>  1.1669
#>  0.8975
#>  0.4486
#> -0.9164
#>  0.4757
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>