Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 1.1915  1.3475 -0.0815 -0.5089 -0.0991  0.2454 -0.0370  0.6298  0.4429 -0.7567
#>  2.6873 -0.4327 -0.6697 -0.3913 -0.3428  0.9259 -0.4633  1.3494 -0.1266 -0.7907
#> -1.3607 -1.0588  0.1291  0.4225  0.0983  0.6117 -0.9942  0.9552  0.0995 -0.1004
#>  0.2678  0.2245  0.1983 -0.6438  0.7010 -0.6974  1.9230  0.2334 -0.0641 -1.8017
#>  1.2900 -0.3519  1.3609  0.3995  0.1465 -0.3298  0.1429  0.9460  0.5467  1.2899
#>  0.0915 -0.3228 -0.8486 -2.1707  0.8765 -0.3452 -0.0832  1.4473  0.2246  1.2830
#>  0.5356 -0.2826 -0.2636  0.3289  1.5007 -1.0736  0.4302  0.0441 -1.7748  0.3040
#>  1.1399  1.0263  0.6513  0.2968 -1.1288  0.4068  0.2266 -0.4338 -0.2478 -1.5117
#> -0.8538 -0.1644 -0.9345  2.1766  1.5723 -0.8661 -1.3587 -0.5363 -0.1260  0.6990
#> -0.8581  0.8595 -0.3744 -1.8865  0.4603 -1.2159 -0.1430 -0.8710  0.4479  1.2461
#> 
#> Columns 11 to 20 1.7032 -0.0170 -1.3513 -1.1502 -0.4909  0.5559  0.2481 -0.2343  0.5491 -1.3063
#>  1.2403  0.3240 -0.0657  0.2048 -1.5396 -0.0285  0.9882  0.6039  1.9256  0.4197
#> -0.5204  0.2827 -1.2560 -1.4162 -0.6179  1.1723  0.9600  0.0801  1.5386  0.7160
#>  0.4677 -0.2298 -0.2883  0.3759  0.6840 -1.0600  0.5592  0.7009  0.4096  0.5683
#>  2.4156  0.6234 -0.3347 -0.7367 -0.8439 -0.3997  0.2642  0.3076  0.8909 -0.5206
#> -0.8075  1.0014 -0.3800  0.6558 -0.0234  0.9600 -0.6411 -0.0874 -0.2202  0.5648
#>  1.4544  1.2724 -0.1986 -0.8290  0.0248 -0.9205  0.3435  0.2182 -0.8963 -0.1859
#>  0.4270  0.0570  1.6512 -0.1308  1.7720  1.1585  1.4879  0.2285  0.6409 -0.4698
#>  0.3364  0.9181  0.8503  0.8850  0.5084  0.2013 -0.9439  1.1192  0.1790  0.4337
#>  0.6998  3.0292  1.4261  0.6340 -1.0991 -3.2307  0.6676  2.2878  1.4438  1.2437
#> 
#> Columns 21 to 30-0.1512 -0.2777  0.6830  0.6672  0.1980  0.1371 -0.3370 -1.4759  1.1090 -0.6342
#>  0.9011 -2.3254  0.2065 -0.0264 -0.2376  0.1118  0.1678 -0.6386  0.5693 -2.0783
#>  0.9033 -0.1213  0.7565  0.4158 -0.5855  0.2188  0.9923  0.4952  0.1176 -0.4054
#> -1.5148  0.9607  0.6684  2.9774 -0.1041 -0.6313 -1.5187  0.6721  0.2734  0.8502
#> -0.8484 -0.7545  0.7938 -0.4705 -0.8687  0.4417 -0.7230  0.1589  1.7037 -0.7744
#>  2.0919 -0.7987 -1.7697 -0.8117 -1.2891  1.1493 -0.1674 -0.8355  0.3876  0.7100
#>  0.0506  0.6357  1.4426  0.8043  0.1093  1.1019 -0.2655  1.1279 -0.6135  0.2301
#> -2.0637  0.3513 -1.6248 -0.2571  0.2438  0.6726 -0.9231 -0.8322  0.0213 -0.5426
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.0905
#> -0.5453
#>  1.6682
#> -0.6370
#> -0.2097
#> -0.4043
#> -0.3566
#> -1.5187
#> -0.7411
#> -0.3525
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.5453
#>  1.6682
#> -0.7411
#> -0.3566
#> -1.5187
#> -0.2097
#>  1.0905
#> -0.6370
#> -0.4043
#> -0.3525
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>