Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.7399 -0.3892  0.5983 -0.7852 -0.0519 -0.1588  1.2552  0.0435 -0.1729 -0.4012
#> -0.1946  0.6219  1.2824 -0.3115 -0.2113  0.3070  0.4546 -1.1491 -0.4712  0.5308
#> -0.0331  1.6101  0.5300 -1.2916  1.5993  1.9095 -0.7871 -1.1150 -0.1966 -0.3476
#>  0.3998  1.1649 -0.9658 -0.9054 -1.0022 -1.1152  1.0623 -1.7172 -0.9123  1.8638
#> -0.2439 -1.1307  0.1682 -0.2930  0.0421  0.0166 -0.2570  0.9621  0.9780  0.5202
#> -1.7965  0.2172  0.9170 -0.3865  0.1964  0.5691  0.5405  0.4030  0.0453 -1.2824
#> -0.6130 -0.4669  0.5261  0.0210  0.7953  1.4963  1.1468 -0.8847  0.7760  1.3920
#> -0.1235  0.2679  0.7420  0.2956  0.3026  1.2862  0.5969  1.2632  0.7881  0.5394
#> -0.7020  0.2905  0.4339 -0.2090  1.4903  0.5738 -0.3614 -0.9161  0.5252  1.5695
#>  0.3586 -2.8672  3.5762 -0.0250 -1.4045 -0.4058  0.5319 -0.7845 -0.3527 -1.0471
#> 
#> Columns 11 to 20 0.6746 -0.1142 -2.7642  0.4194  1.2654 -1.7030  0.2243 -1.0283  0.1292 -1.2135
#>  0.7369  0.1502 -0.7808  0.1059 -0.3646 -2.0335 -0.9801 -1.0485  0.3996  0.9453
#>  0.0609 -0.3558  0.6153 -1.8612 -0.4509 -0.1536 -0.5919 -1.3270 -0.7320 -0.0265
#>  1.4232  1.3861 -0.1101  0.4385 -0.1826 -0.2921  0.3961 -0.2643 -1.3186 -0.4277
#> -0.1474  0.9380  0.6808  1.1488  0.0492  1.4794 -1.2945 -1.0296 -1.4244 -0.8128
#> -0.2355 -1.4898 -1.6446 -0.7280 -0.4276  1.2169 -1.5958 -1.1810  0.2662 -0.4365
#>  1.8748 -1.0646 -0.1286  0.5364  1.3258  0.8373  0.7239  0.5634  0.3456  0.6427
#>  0.5757  0.5715 -0.5877  0.8907 -0.8347  0.2203 -0.5914 -1.2525  0.3127  0.3276
#> -1.3258 -0.1801 -1.9053 -0.6394 -1.2424 -0.2359  1.1895  0.9126  0.8376  0.0427
#>  0.5506 -1.6712 -0.7576  0.6125  0.6498  0.4789 -0.2647  0.9481 -1.0043 -0.1550
#> 
#> Columns 21 to 30 0.0250 -0.0911 -0.3010 -0.5792  0.7179 -0.8628  0.7264 -1.3600  1.4192 -0.6955
#> -0.4974  0.2053  0.0050  0.2603  0.3018 -0.3599 -0.9405  1.0859  0.7866  0.0697
#>  0.5164  0.5504  0.7887  0.2388 -0.0964 -1.5985 -0.0320  0.1793 -0.4651  0.8910
#>  0.6926 -1.2132 -0.6745  1.6344  1.1210  0.7885 -1.0741 -0.5289  2.4695  0.8091
#> -1.2931  0.9320  0.5066  0.5818 -0.8557  0.6729 -0.8810 -0.6550  0.1794 -2.0904
#> -0.1724 -0.0775 -0.3265 -0.4801 -0.1345 -0.5827  0.7886 -0.3019 -2.0725 -0.2351
#> -0.7866  1.0139 -0.9934  0.4858 -0.0328 -0.3610  0.1071 -0.5045 -0.1833  0.2079
#> -0.5306 -0.9146  0.7502  1.0543  0.7272  0.7702 -0.0353  0.4253  2.0539  0.8041
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.1608
#> -0.5863
#> -0.1588
#> -0.7619
#> -0.9362
#> -0.6765
#>  0.9972
#>  1.1358
#>  0.6090
#> -1.7447
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.1608
#>  1.1358
#> -0.5863
#> -0.7619
#>  0.9972
#> -0.6765
#>  0.6090
#> -0.9362
#> -0.1588
#> -1.7447
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>