Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-1.2461e+00 -9.3290e-01  8.3062e-01 -1.3499e+00  1.1685e-01  4.4261e-01
#>  6.3104e-01  6.3915e-03 -8.7281e-01  5.3078e-01 -5.7607e-01 -7.0721e-01
#> -9.6812e-01  1.2862e-01 -8.3993e-01 -5.9082e-01  2.7955e-01 -4.6288e-01
#> -6.1352e-01 -1.0569e-01 -1.9454e+00  8.6928e-01  8.0006e-01  1.8695e-01
#> -1.8452e+00 -1.0223e+00 -2.0673e+00  6.5999e-01 -7.8533e-02 -1.9387e+00
#>  1.0611e+00  2.7769e-01 -9.2933e-01  2.1205e-01  1.1149e-01 -1.4077e+00
#> -1.6910e+00 -1.2725e+00 -2.0850e-01  4.5977e-01 -1.8418e+00 -8.3559e-01
#> -6.6341e-01 -2.8185e+00  1.0511e-01 -5.8665e-01  1.3224e+00  3.6570e-01
#>  6.6780e-01  3.0635e-01 -4.8479e-02  1.4024e+00  1.0554e+00  1.4415e+00
#>  2.1722e-01  2.9571e-01  4.8366e-01  7.6254e-01  1.1028e+00 -4.9217e-02
#> 
#> Columns 7 to 12-1.8120e+00 -9.4209e-02 -2.5700e-01 -1.3966e-01 -1.0691e+00 -3.0305e-01
#> -7.6059e-01  4.3977e-01 -5.4546e-01 -2.1632e+00  1.7676e-01 -9.4204e-02
#> -1.1552e+00  1.3569e+00 -2.2032e+00 -4.3199e-01  2.1796e+00  1.9585e-01
#>  9.3000e-01  1.4727e-01 -1.0194e+00 -1.4417e+00 -8.6096e-01  1.7167e+00
#> -2.2197e+00 -1.6018e+00 -1.4909e+00 -3.7962e-01 -1.3782e+00  1.2227e+00
#>  3.6090e-01  9.9905e-01  1.0816e+00  3.2719e-01 -1.4802e+00  2.0648e+00
#>  2.7126e-01  1.3634e+00  7.0970e-01 -1.2869e+00  4.0354e-01  2.9781e-01
#> -5.6100e-02 -8.8022e-01 -9.2955e-01  1.0419e+00 -2.0408e+00  3.8047e-01
#>  9.8746e-01 -1.8201e+00  1.9516e-01  7.7747e-02 -1.1984e-01 -3.6244e-01
#> -1.5373e+00 -1.2126e+00  3.9317e-01 -7.4328e-01  4.5744e-01 -8.4676e-01
#> 
#> Columns 13 to 18 1.5239e-01 -1.0130e+00  3.8807e-01  5.1969e-01  1.9038e+00 -9.0134e-01
#>  3.4947e-01  1.8570e+00 -1.4333e+00 -1.2673e-01  1.0253e-01  2.8777e-01
#> -4.4079e-01 -7.4045e-02 -5.3552e-02 -1.1342e-02 -2.0575e-01  5.7991e-01
#> -3.6868e-01  1.5156e-03  1.0175e+00  2.8778e-01  3.1020e-01  4.3615e-02
#>  6.2759e-01  1.5356e+00 -4.0328e-01  1.7083e+00  2.4319e+00  2.2231e-02
#>  2.5187e-01  7.0928e-01 -4.8524e-01  6.6803e-01  1.3722e+00  2.6523e-01
#>  6.7052e-01  4.8685e-01  9.3000e-01 -8.4472e-01  7.6880e-01 -9.6462e-01
#> -4.9258e-01  8.8292e-02  5.8750e-01  3.3479e-02  5.7140e-01  1.5285e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.5557
#>  1.4739
#> -0.9994
#> -0.3600
#>  0.8481
#>  2.5732
#> -1.1444
#> -1.5477
#>  0.9113
#>  1.2245
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.9113
#> -0.3600
#> -1.1444
#>  0.8481
#> -0.9994
#>  2.5732
#>  1.5557
#>  1.4739
#> -1.5477
#>  1.2245
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>