Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.4610 -0.3945 -0.3799  1.6444  0.6492 -0.1950  0.5261  0.0043 -0.6646 -0.3270
#>  0.0892 -0.5047 -1.1244  0.2758  0.0803  0.2589  0.4476 -1.5530  1.2574 -0.0304
#> -1.3493 -0.2408 -0.2302 -1.9278 -1.8161 -0.1963  1.1828 -0.3229  1.7473  0.8234
#>  0.7668  0.2524 -0.2126 -0.2393  0.2097  0.3182  0.5009 -0.9711 -1.9580  1.1996
#> -1.4431  0.2316 -0.6911 -0.7949 -0.4255  0.5436  1.1515 -0.8108  0.0589  0.5536
#>  0.1911  0.6597 -1.5432  1.0964 -1.4818 -0.6567  0.1728  1.1761  0.6023 -1.0521
#> -2.3868  0.1398 -0.3281 -0.2502  0.0988 -0.3660 -1.1207  0.1437 -0.4539  1.5881
#>  0.0552 -1.4962  0.4906  1.0792  0.1087  2.3383 -0.4368 -0.8255 -0.1977  0.2980
#>  0.8514  0.0477 -0.6167 -0.5293 -1.3278 -0.7991 -0.4482  0.5737 -0.6214  0.4522
#>  0.3241  0.7816 -1.0612  0.6237 -0.6652  1.3204 -1.0149  0.5827 -0.6226  0.3301
#> 
#> Columns 11 to 20 0.7158  2.7128  1.8616 -1.2894 -1.3575  0.6181  0.9349 -0.3882 -1.5209 -0.5492
#>  0.1286  0.6573  0.7076  1.5378 -0.3876 -0.0261 -0.0064  1.5712 -0.5429  1.3568
#>  1.7456  0.6972  0.3889  0.0889 -0.1791  0.7780  2.7446  0.1050  0.4026 -0.0748
#>  0.3999 -0.2955 -0.8080  0.5055  0.5426  0.7245 -0.7291  0.1526 -1.2140  1.0948
#>  1.2669 -1.7673 -0.8114 -0.0132 -1.0167  1.0133 -0.3340 -1.7585 -0.9362 -0.4103
#>  0.8728 -0.7307 -1.5248 -1.4267  1.3940 -0.5738 -1.3194  0.1331  0.7789 -1.2932
#> -0.8542  1.8572  0.6762  0.5968 -0.6043 -0.8434  0.5815 -1.3626 -1.4918  0.5321
#>  0.1547 -1.0422  0.5311 -0.1959  1.2628 -0.6828  2.3471 -0.4015 -0.6491 -1.0358
#>  0.6962 -1.6123  0.6839  0.1205  0.3454 -0.7730  0.0014 -0.0047 -1.4542 -0.0636
#>  0.6148  0.3822  0.0895  1.2374 -1.1290 -1.7810 -1.1131  0.0428  1.8303  0.9593
#> 
#> Columns 21 to 30-0.3339 -0.8042  0.2032  0.7578 -0.0696  1.2467 -0.4625  1.5351  0.3614  0.6541
#>  0.5914 -1.1657 -0.4467 -1.1220  0.9590 -0.5791 -0.2081 -0.1566 -0.4766  0.4725
#> -0.6038  1.4510  0.2236 -0.2479  0.0224 -1.3777 -0.1754  0.6922 -1.7598  1.0684
#> -1.2516  0.5729 -0.5520  0.3855  2.4642  2.2716  1.7630  0.4034  0.4599  0.5649
#>  0.3909  0.2743 -0.0202  0.8382 -1.0001  0.3754 -1.2099 -0.5827 -0.8563 -0.4314
#>  0.5777 -0.1465  0.3288 -0.2886 -1.1826  0.5462  0.9720 -0.3401 -1.2086  0.1964
#>  0.4225  0.1376  0.5335  0.1311 -2.1980 -0.2419  0.2756  1.9791 -1.4217  0.2717
#>  0.7183  0.6429 -0.3814  0.4703 -1.1703  0.5372  0.3755  0.7168 -0.6356 -0.1975
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.5959
#>  0.8487
#> -0.6288
#>  0.2438
#>  0.4132
#>  0.0826
#>  1.6119
#> -1.2292
#> -1.8854
#>  1.8747
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.2292
#>  0.2438
#>  1.6119
#>  0.0826
#> -0.6288
#>  1.8747
#>  0.8487
#>  0.4132
#> -1.8854
#> -0.5959
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>