Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.0471 -0.3960  0.1831 -1.3533  0.8482 -0.3397  0.2712  0.2076  0.6414 -0.3867
#> -1.1847 -0.4114  1.6663  1.6568  1.6280  0.6788  0.4035  1.4302 -0.3156  1.1555
#> -0.1444 -0.1544  0.9179 -2.3987  1.4588 -0.9783 -0.7982  0.0952 -0.5717 -0.7066
#>  0.2920  1.9798 -0.7682  0.1725  0.9738  0.9840  0.8623  1.0153 -1.0577  1.7361
#>  0.2379 -1.0642 -1.5048 -0.2829 -0.9043  0.6475 -0.6551 -0.6280  1.0132  0.6255
#>  0.4308  1.7266  0.1520 -0.3705 -1.1746  0.2547 -0.0598  0.0227  1.6180 -0.6547
#>  0.1892  0.2331  1.3495  1.2575  0.6889  0.6204  0.6986  0.3958 -0.6146  0.2813
#>  0.2405 -0.5854  0.5224 -0.1970  0.3656  0.9756  0.6754  1.3868  0.1776  0.5066
#> -0.9542 -0.9670 -0.1713 -0.7556  2.1828  0.5050 -0.8082 -0.3513 -0.2268 -0.7296
#> -0.6593  0.4649 -1.4361  1.5572  0.6728 -0.2841  1.3680  0.2478 -0.1393 -0.1782
#> 
#> Columns 11 to 20 0.4217 -0.4263 -0.7064  0.0872  0.7933 -1.4815  0.8983  0.0833 -0.5042 -0.4328
#> -0.6651 -1.1458 -0.2817  0.2328 -1.1159  0.0540  0.9313 -1.0396 -1.0475 -0.4261
#> -0.9094 -0.6527 -0.3066 -1.5690 -0.1850 -0.1579 -0.4476 -1.0572  0.4411  0.5723
#> -1.7139  1.4749 -1.0576  1.1793 -0.6476  0.0919  1.2532 -0.9785  0.6205 -0.4211
#> -0.1768 -0.8604  0.6973 -0.2501 -1.3664  0.9249  1.2895  2.5905  1.2916  1.3818
#> -1.1209 -0.6283 -0.1723 -0.1147 -0.4707 -0.4402 -0.0368  1.0924 -0.2243  0.7058
#>  0.2826 -0.2471 -0.8957  0.5307 -0.3570  1.7517 -0.1657 -0.0284 -0.8001  2.2384
#>  0.1538  1.5186  0.8812  1.2411  0.7741  0.3825  0.0517 -0.1919  2.2015  1.0135
#> -0.4136 -0.5627 -0.2856  0.2589  1.2913 -0.0482  1.5316 -1.1487 -0.5348  0.1740
#> -0.5969 -0.6000  0.2033  0.4452  0.9769  1.3474  1.0047  0.1087  0.7992 -1.0555
#> 
#> Columns 21 to 30-0.1935  0.5565 -0.6298 -0.3984  0.1182  0.2037  0.1195  1.5663 -1.2078 -0.2108
#> -0.5918  1.6134  1.5566  0.6819 -0.1198  0.1461 -0.4163 -2.1621 -0.6545  0.5421
#> -0.1395 -0.7489  1.8462 -0.8655 -0.6889 -0.8857 -0.3398  1.0896 -1.3994 -0.8709
#> -0.2663 -1.0177 -0.9308 -0.3023 -0.5663  0.5708  0.2384 -0.5835 -0.4456  1.0058
#>  0.2796 -1.4203  2.2239 -1.5310  0.5873  0.8132 -0.7452 -0.4338 -0.6509 -0.8671
#> -0.8411 -1.3950  1.3048 -1.6133  0.3291  0.0785  0.1777  0.3471  0.0881  0.3414
#> -0.1221  0.1552 -2.6603 -0.4103 -3.4103  0.6510 -1.0009  0.5235 -1.1094 -0.0936
#> -0.5031 -0.6861  0.1147 -0.8525 -0.8737 -0.0919  0.9933  1.2692 -1.1781  0.2487
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -2.4131
#> -0.7169
#> -0.4950
#>  0.2414
#> -0.3643
#>  0.5187
#>  0.3147
#>  2.1676
#> -0.9031
#> -0.8404
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.4950
#>  2.1676
#> -0.9031
#> -0.8404
#> -2.4131
#> -0.3643
#>  0.5187
#> -0.7169
#>  0.3147
#>  0.2414
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>