Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.4541  0.5002 -0.3446  0.3818 -0.0144  0.4346 -0.3108 -1.2076 -1.7950 -0.6227
#>  1.0930  0.7916 -0.4570  0.6409  0.2898 -0.5048  0.2589 -0.3488  0.5376 -0.6359
#>  0.9030 -0.0834 -1.1850  0.3938 -1.2827 -1.2151  1.0309  0.4329  1.1631 -1.3999
#>  0.1217  1.2595  1.1944  1.0270  1.2664 -0.0314 -0.8075  1.7688  1.4993 -0.0851
#>  0.4597  0.8015  0.4849  0.3015 -0.0439  1.1234  0.4176  0.0732 -0.2838 -0.5951
#>  0.3462 -1.0578  0.8884  0.0657  0.0465 -1.1571  0.2211 -1.0165  0.0465  1.1462
#>  0.2661 -0.7308  0.1577 -0.7900 -0.6379 -0.1892  0.4562 -0.0242 -0.9420 -0.6627
#>  0.2765  1.7471 -0.1957 -2.1457 -0.7146 -0.3930 -0.0754 -0.4255  0.6386  0.3479
#> -0.6428  0.0921 -0.6172 -0.8799  0.2482  0.6776  1.8323 -0.3014  0.7236 -0.2312
#>  1.1366 -0.0984  0.1138 -0.4246 -0.3196 -0.2069 -0.8103  0.9229  0.1288  0.3711
#> 
#> Columns 11 to 20-1.1978  0.5996 -0.4163  1.5834 -0.1004  1.4396 -0.7776 -1.7034  0.7265 -0.3271
#> -1.0554 -1.1587  1.4444  0.3247  0.9706  0.1679  0.2283 -0.9296  1.0349  0.2818
#> -0.2432  0.7425  1.9298  0.3131 -0.2495 -0.1255 -0.7841  0.2544  0.7975  1.0236
#>  0.9192 -1.0567  0.7421  1.3616 -0.3888  2.1947 -0.0861  0.2157 -0.2360  0.6237
#> -0.5117  0.9306  0.3624 -0.9365  1.3540 -0.7887  0.0788  0.3952 -0.4240 -1.2237
#> -1.3847 -0.0188  2.3638 -0.9355  0.4904 -0.8346 -0.4302 -0.7936 -0.7876  0.2399
#>  0.3844  0.5797 -0.8387  0.6110 -0.9091 -0.3268  0.5782  0.6224  0.1975 -0.3950
#> -2.3031 -0.6148  1.9780  0.8338  0.1802 -0.7544 -1.9411 -0.1648 -0.5381  0.1840
#>  0.8539 -0.0283  0.3045 -1.5070  1.2545 -0.3362 -0.4902 -0.2650 -1.1030 -0.5104
#> -0.4684 -1.5237 -0.2510 -2.3816 -0.2438 -0.1700  1.3189 -1.4456 -0.2542  0.7062
#> 
#> Columns 21 to 30-1.5354 -0.4757  0.4873 -0.6444  1.0353  2.0746 -0.6439  0.6643  0.0165  1.2210
#> -0.2541 -0.0877  0.1276 -0.9676 -1.6433  1.0258 -0.7076  0.6683 -0.0565  0.9538
#>  1.4578 -0.4155  0.3601  0.4839  1.8695 -1.5925 -0.0658 -1.2617 -1.1245 -1.9116
#>  0.5510 -0.5368  0.1822  1.2111 -0.1021 -0.2776  0.5951 -1.1391 -1.2270  0.5782
#>  1.5246  1.4026 -0.2054 -0.2511 -0.7132 -0.6255 -1.5960 -0.0933  0.3074  0.3792
#>  1.0378  0.3627 -0.3119 -0.9145  0.1805  0.2882 -0.4786 -0.2251 -0.1043  1.8018
#> -0.7745 -0.5846 -0.2284 -0.4133 -0.0836  1.2724  2.0768  0.3619 -0.0334 -0.9204
#> -0.2834 -2.7180  1.5339 -1.1551 -0.6864  0.5307  0.5017  0.0828 -0.0022  0.2602
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.7524
#> -2.0763
#>  1.6060
#>  1.0018
#> -0.5382
#>  0.3319
#> -0.2225
#>  0.3608
#>  1.2112
#>  0.8987
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.2225
#>  0.7524
#>  0.8987
#>  1.0018
#>  0.3608
#>  1.2112
#> -2.0763
#>  1.6060
#>  0.3319
#> -0.5382
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>