Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 1.1393  0.0993  0.2189 -0.2654 -0.7432 -0.9969 -1.2753  0.5388  0.5730  0.2095
#>  0.3582  0.1142  0.6296  1.0551  2.1062 -1.7690 -1.8887  0.4381 -0.1461 -1.5384
#>  0.3171  0.4339  1.6651  0.6048  0.4125  0.0881  0.4461 -1.6529  1.1526 -0.5967
#>  0.8984 -1.0272  0.3716 -0.8413  0.2836 -0.2622  0.1638  0.5714  0.8996 -0.3916
#>  0.2602 -0.1375  1.2566 -1.7051 -0.3983  0.4102  0.6852 -0.4518 -1.3997 -1.1431
#>  1.5858 -0.5452  0.9573 -0.4355 -0.9121  0.2738 -1.5407 -0.3501  0.7244  0.9511
#>  0.1820  0.5924  0.6709  1.2956  1.4954 -1.1035  0.0082 -1.3616  0.6332  0.3337
#>  0.0645  0.3202 -0.2724 -0.5346  0.4736  0.1301  0.0420  0.0624  0.1931 -1.1690
#>  0.0742  1.1964  0.1973  0.8757 -0.3198  0.4474  0.1053  0.2626  0.3517 -0.3764
#> -1.2348  0.8410  0.4748  0.8293  0.6034 -0.1404  1.0503  0.2744  0.0553  0.0101
#> 
#> Columns 11 to 20-1.1628 -0.9000 -0.1200 -1.7747  2.0785  0.2508 -0.2756  0.9503 -0.2875 -0.9612
#>  0.4514  0.4650 -0.1955  0.0284 -1.5111 -1.1480 -0.6545 -0.7647 -0.4443  1.3862
#> -0.0846 -0.1375 -1.2416 -1.1809 -1.0445 -0.7300  1.0265 -0.0616  0.2704  0.7250
#>  1.2937 -0.0212 -0.1863  2.1415 -0.6190 -0.3932 -0.2443  0.5312 -0.1467  0.3124
#>  0.2674 -1.0153  0.2602 -0.3196 -0.5767 -1.4015  0.5534 -1.4207  0.6381  0.5604
#>  0.6666 -0.8641  0.3172 -1.2492 -0.2380  1.2477  0.0292  2.0571  0.3930 -0.8251
#> -1.6735 -1.9595 -0.2864 -0.6369 -1.3433  1.5303 -0.3180  1.1233  0.4945  0.8848
#> -0.0305 -0.2461  1.3660  0.1224 -1.9653  0.1442 -0.6087  0.5048 -2.6664  1.3331
#> -0.9083 -0.1662  1.2504  0.0953  1.1607  1.0706  0.4981  0.3725  1.3551 -0.8204
#>  0.7000 -0.4391 -0.5123 -1.9343 -0.0952 -1.5392  0.1590  0.4879 -0.7269 -1.3653
#> 
#> Columns 21 to 30 0.6593 -0.9278 -0.2411  0.8735 -0.2638  0.1710  1.2417 -0.0843 -0.3250 -0.4481
#> -0.7164  0.2065 -0.0711 -0.1089 -0.4489  1.1127 -0.2569  1.9696 -0.6500 -0.6558
#> -0.4417  1.0230 -0.5084  0.1311 -1.3141 -0.3737  0.0621 -1.7182  0.6729  0.4252
#> -0.6836  1.5845  0.4761 -1.6866 -1.0847 -0.9799  0.7991  1.3185  0.5447  1.6422
#> -1.0973 -1.5845  0.3143 -2.5648  0.6592  0.1162  0.9846  1.4157  0.9013  1.2128
#>  1.0941  0.2957  0.5872 -0.8143 -1.0980 -0.8505  0.7287 -0.0129 -0.3398 -0.4599
#> -1.8190  1.0450  1.0871 -1.0077  0.0792  0.4431  1.0031 -2.3264 -0.6882 -0.3862
#> -0.3753 -1.0501 -0.2500 -0.3912  0.0619  1.3641  0.3387 -1.1738 -0.8219 -0.8540
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.6201
#> -0.9265
#> -0.1409
#>  0.2495
#>  1.8456
#>  0.7852
#> -0.7732
#>  0.0345
#> -0.3194
#>  0.5101
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.6201
#> -0.3194
#>  0.2495
#>  0.7852
#>  1.8456
#> -0.7732
#> -0.1409
#>  0.5101
#>  0.0345
#> -0.9265
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>