Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-8.6082e-01 -7.8806e-02  1.9245e+00  6.8804e-01  6.9887e-01 -8.4450e-01
#>  1.3861e+00 -1.9949e-01  5.2210e-01  2.7854e-01 -6.6785e-01  2.2387e-01
#> -7.6426e-01  2.7076e-01 -6.7413e-01 -9.3596e-01 -2.5374e-01  8.6298e-01
#>  8.0512e-01 -2.6108e-01  7.5967e-01  1.2459e+00  3.7697e-01 -1.6079e+00
#>  5.8751e-01 -1.1762e+00 -1.1527e+00  4.3746e-01 -1.8238e+00 -5.4201e-02
#>  7.2034e-01  5.4933e-01 -1.1879e-01 -1.5244e+00 -6.4757e-01  1.5475e+00
#>  5.9149e-01 -9.4957e-01  1.2790e+00  3.3994e-01  1.2602e+00 -1.2770e+00
#>  5.9479e-01 -2.8007e-01 -8.9491e-01  4.8623e-01  1.4879e+00 -4.6115e-01
#>  5.6121e-01 -1.5925e-01 -9.1020e-01 -5.4900e-01  7.5694e-01  1.1214e+00
#> -2.4273e-01  5.5416e-01  5.2655e-01  6.1916e-01 -1.2841e+00  2.9820e-01
#> 
#> Columns 7 to 12 8.8958e-01 -6.5575e-01 -9.3499e-01 -9.2691e-01 -1.0460e+00 -1.9959e-01
#> -2.7485e-01 -3.8797e-01  9.8359e-01  1.5072e+00  9.6217e-01 -3.9360e-01
#> -5.0523e-02 -1.0400e+00  1.2301e+00  3.3220e-01  2.4788e-01 -5.1029e-01
#> -3.5306e-01  1.0653e+00  1.8812e+00  5.9615e-01 -9.0495e-01  4.5232e-01
#> -1.2325e+00  7.7903e-01 -8.1954e-01 -8.5182e-01 -8.3420e-02 -2.1520e+00
#>  2.7993e-01 -1.9192e-01 -1.7076e+00  3.6462e-01  6.2233e-01 -1.0834e+00
#>  6.9701e-02 -5.6126e-01 -2.7326e+00  1.3215e+00 -1.4140e+00  2.7138e-01
#>  8.5111e-01  3.8518e-01 -1.1404e+00  6.8668e-01 -4.2637e-01 -5.7483e-01
#> -3.0779e-01  1.0837e-01  1.1650e+00 -5.6728e-01 -3.5224e-01 -8.7674e-01
#>  4.4812e-05 -3.2135e-01 -6.8465e-01  1.9593e+00  2.2908e-01  9.6052e-01
#> 
#> Columns 13 to 18 1.1539e+00 -2.0476e-01  1.7435e+00  4.6071e-01 -5.7770e-01  5.9283e-01
#> -1.0260e+00 -9.7207e-02  2.1354e+00  1.9578e+00  7.0506e-01  1.4208e-01
#> -1.8285e-01  1.2825e+00 -2.1050e-01  1.4670e-01  2.2600e-01  6.3807e-01
#>  1.9413e-01  1.0984e+00 -1.2806e+00  4.6220e-02 -7.7638e-01  1.4289e+00
#> -6.6939e-01  9.9155e-01  1.3335e+00 -6.3257e-02 -1.5242e-01  4.5087e-01
#> -4.2777e-01 -3.8921e-01 -1.7355e+00  6.1378e-01 -6.6066e-01  2.2810e+00
#>  4.1598e-01  7.7193e-01 -7.9969e-01  7.5646e-01 -1.3182e+00 -1.1350e+00
#> -1.2046e-01  3.3021e+00 -7.1881e-01 -7.0439e-01  9.7505e-01 -1.4283e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.4051
#> -0.5336
#> -1.8154
#> -0.2333
#> -0.4749
#>  0.1056
#> -0.0158
#> -0.5855
#> -1.3174
#> -0.0093
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.2333
#> -1.8154
#> -0.0093
#> -0.4749
#>  1.4051
#> -1.3174
#> -0.5855
#> -0.0158
#>  0.1056
#> -0.5336
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>