Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 5.6676e-01  6.2325e-01 -3.8728e-01 -1.5728e+00  2.1272e+00  3.9756e-01
#>  1.3353e-01 -3.3821e-01 -8.9423e-01 -8.8234e-01 -2.2186e-01  1.0839e+00
#> -8.4559e-01  9.7151e-02 -6.6803e-01  2.5170e-01 -4.2755e-01 -1.2013e+00
#>  7.7831e-01 -7.5631e-02  3.0310e-02 -3.7941e-02  1.6405e+00  9.0960e-01
#>  2.0453e+00 -9.9076e-01 -5.2568e-01 -2.2468e-01  3.0600e-01  5.4187e-01
#>  5.1603e-01  7.5529e-01 -2.0267e+00 -4.9173e-01 -6.3315e-01  1.5932e-01
#> -1.1884e+00 -2.2367e-01  4.9242e-01  5.8167e-01  1.0748e+00  5.5370e-02
#>  7.0970e-02 -1.8690e+00 -1.3096e+00  1.4884e+00 -1.3856e+00 -2.6528e+00
#> -3.1004e-01 -1.8209e+00 -1.7792e+00 -6.5043e-01 -1.3018e-01 -1.5728e+00
#> -9.2817e-01 -6.0697e-01  2.4412e-01 -1.3588e+00  1.1955e+00 -6.8833e-01
#> 
#> Columns 7 to 12 2.5248e-01  5.9760e-01 -8.2978e-02  6.7677e-01  3.4674e-01  4.2238e-01
#> -6.2521e-01  7.6254e-01 -2.8358e-01  3.1860e-01  9.1043e-01 -1.8678e+00
#>  5.3440e-01 -1.3008e-01  1.1122e-01 -7.1764e-01  1.4741e-01 -6.5163e-01
#> -1.4888e+00  1.6559e+00  2.9930e-03  6.8538e-01 -2.6357e-01  1.8291e+00
#> -9.2188e-01  1.1157e+00  1.4423e+00  1.4117e+00  6.2070e-01  5.5953e-01
#> -7.3411e-01  7.0384e-01  3.3126e-01  1.3378e+00  1.7040e+00 -1.3195e+00
#> -1.1636e+00  6.0859e-01  1.6662e-01 -3.1503e-01 -8.6661e-01  1.8334e-01
#>  5.9566e-01 -1.3938e+00 -1.0168e-02 -1.5078e+00  2.6400e+00 -1.6926e+00
#>  1.4183e-01  1.3655e+00 -1.5842e+00  1.0410e+00  4.4420e-01 -4.3664e-03
#>  2.0299e+00 -1.5588e+00  7.2734e-01 -3.4725e-02  9.8284e-01 -1.3474e-01
#> 
#> Columns 13 to 18 1.9344e+00 -4.8151e-01 -7.8655e-01  1.2518e-01 -3.2420e-01  1.6730e+00
#>  1.1779e+00 -7.0587e-01  8.4791e-01 -1.6691e-01  4.3034e-01 -7.0923e-01
#> -9.0019e-01 -1.7663e+00  2.2678e-01  6.6315e-01 -4.3417e-01  4.3427e-01
#>  1.2672e+00 -5.4747e-01  9.4612e-01  1.3264e+00 -6.2011e-01  9.9055e-01
#> -1.6189e+00 -4.5014e-01 -3.4264e-01  2.3604e-02 -2.4182e-01  4.2150e-01
#> -3.2051e-01  9.5392e-01  4.8100e-02  5.3385e-01  7.0211e-01  5.3834e-01
#> -7.5549e-01  6.7074e-02  7.7430e-02 -9.6041e-01 -5.9655e-01  7.9696e-01
#> -9.5731e-01  2.6364e-01  7.1603e-01 -3.2559e-01 -6.1679e-01 -9.5858e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.8183
#> -1.6612
#>  0.0645
#>  0.9606
#> -0.6594
#>  0.9223
#>  0.4477
#>  0.5539
#>  0.9437
#>  0.4016
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.9606
#>  0.9223
#> -1.6612
#>  0.9437
#>  0.0645
#> -0.6594
#> -0.8183
#>  0.5539
#>  0.4477
#>  0.4016
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>