Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-9.2247e-01 -1.2815e-02 -6.1802e-01 -3.2562e-02  8.6357e-01 -2.4183e-01
#> -6.9456e-01  7.3533e-01  6.2588e-01  7.0242e-01 -7.0534e-02 -2.4178e-01
#> -1.5092e+00 -2.6282e+00 -6.1397e-01 -9.4997e-01  3.6357e-01  1.2298e+00
#>  2.7976e-01 -1.4617e-01 -7.5300e-01 -8.7853e-01  1.3445e+00  9.9036e-01
#> -2.7111e-01 -3.8931e-01 -9.2454e-02  2.0481e-01  1.1760e+00  9.8344e-01
#>  1.9328e+00 -1.2154e+00 -4.0962e-02  3.2196e-02  2.1684e+00  1.0536e+00
#> -4.9443e-01 -5.0192e-01  6.9316e-02 -1.5099e+00  2.0120e+00 -5.8831e-01
#>  1.5851e+00 -4.6320e-01  1.2727e+00  4.5362e-01 -1.6535e+00 -7.6227e-02
#>  1.1236e+00 -2.7106e-02 -3.7196e-01  7.9898e-01 -7.4922e-01  1.4065e+00
#> -7.0321e-01 -8.1048e-01  4.5265e-01 -1.9664e+00  1.1984e+00  5.3083e-01
#> 
#> Columns 7 to 12-5.9786e-02  1.0259e-01  1.9850e+00  9.4303e-01 -3.1519e-01  4.7030e-01
#> -7.8137e-01  1.1502e+00  1.1997e-01 -6.3207e-01 -2.0948e+00 -1.8088e-01
#> -4.2415e-01  9.5407e-01 -1.2310e-02 -1.0077e+00  6.7865e-02 -1.3388e+00
#>  1.2348e+00 -2.1930e+00 -2.9768e-01 -1.6907e-01  2.7630e-02  6.2069e-01
#> -1.6609e+00 -2.1990e+00  5.2509e-01 -6.7213e-01  1.7356e+00 -1.7452e+00
#>  8.8089e-01  1.0634e+00  3.1363e-01 -7.0527e-01 -4.0226e-01  6.5657e-01
#>  1.1206e+00 -3.8013e-01 -1.5426e+00 -1.2347e+00  1.1051e+00 -2.3710e-01
#>  7.7520e-01 -8.8392e-01  6.8168e-01  1.2417e+00 -2.0922e+00  9.9962e-01
#> -6.5171e-01 -1.1491e+00 -7.4613e-01  4.5294e-01 -3.9596e-01 -1.1822e-01
#>  6.2012e-01  3.0141e-01  1.7139e+00 -2.7488e-01  1.4372e+00 -3.0062e-01
#> 
#> Columns 13 to 18-9.9892e-02 -7.2227e-01 -5.6123e-01 -7.9673e-01 -3.8143e-01 -5.9890e-01
#> -2.3839e-01  6.6169e-01  1.7523e+00  1.4937e-01  1.3127e+00 -1.7956e-01
#> -1.1312e+00  4.3941e-01 -7.0857e-01  5.2472e-01  1.0344e+00  8.9731e-01
#>  4.7065e-01  9.5893e-02 -1.2389e+00 -4.5991e-03  1.4168e+00  9.3533e-01
#>  6.3746e-01 -7.2276e-01  1.7191e+00  8.1692e-02  1.4298e+00  8.1216e-01
#>  1.0852e+00  7.1517e-01  1.2852e+00 -2.5990e+00  1.1138e+00 -4.5792e-01
#>  5.2382e-01  3.0123e-01 -2.1725e-01  1.4007e+00 -5.2686e-01  5.9708e-01
#>  5.2110e-01  9.3902e-01 -1.0739e+00  1.1393e+00  2.1741e+00 -1.6465e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.1645
#>  0.9837
#> -1.9027
#>  0.3870
#>  1.3750
#>  0.8127
#>  1.2238
#>  0.4257
#> -0.0921
#> -0.4870
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.3870
#>  0.4257
#>  0.9837
#>  1.2238
#>  0.8127
#> -0.4870
#> -0.0921
#> -0.1645
#>  1.3750
#> -1.9027
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>