Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 5.5721e-01 -5.2166e-02  1.7797e+00 -1.8861e+00 -1.4358e+00  1.5979e+00
#> -1.6089e+00 -6.3579e-01  2.1093e+00 -1.6231e+00 -1.2099e+00  1.2311e+00
#>  4.8387e-01  1.2621e-01 -1.4176e-01  2.0697e+00  8.2255e-01 -1.3939e-01
#>  3.4483e-01  6.2533e-01 -3.7123e-01  3.8865e-01  1.2123e+00 -5.1277e-01
#>  9.9745e-01  9.5490e-01 -7.6970e-01  1.5246e+00 -5.0048e-01 -1.0042e+00
#> -2.9178e-01  5.5845e-02  1.3709e-01  3.1181e-01 -1.7443e-01 -1.2663e+00
#> -8.1689e-03  3.2289e-01  4.0838e-01  4.9088e-01  6.0705e-01 -1.1643e+00
#>  2.2631e-02 -2.2312e-01  1.0005e+00 -8.1882e-01  6.1623e-01  7.0849e-01
#>  5.8433e-01  1.5419e+00 -1.1952e+00 -7.8664e-02  1.0717e+00  2.6806e+00
#> -1.5062e+00 -9.6501e-01 -9.7541e-01  6.1731e-01  9.5494e-01 -8.5491e-02
#> 
#> Columns 7 to 12 1.6006e+00  1.9019e+00  8.1998e-01  7.8871e-01  9.6176e-01 -9.2333e-01
#>  2.2801e+00  1.3547e+00  5.7696e-01  1.2252e+00 -1.2466e+00 -6.5016e-01
#>  6.5789e-01 -8.0962e-01  7.1157e-01  7.9691e-01 -1.8297e-01 -9.7525e-01
#> -7.7234e-01  1.0940e+00  1.0938e+00  5.6303e-01 -1.1330e-01 -3.6798e-02
#>  5.4255e-01  7.6438e-01 -9.1615e-01 -6.5596e-01 -1.5952e-01  2.3628e-01
#> -2.8630e-01  1.6870e+00 -3.6743e-01 -1.1579e+00 -1.9258e+00 -1.8557e+00
#>  5.2262e-01 -3.2806e-01 -8.6161e-02 -1.1738e-01  5.2589e-01  7.0990e-01
#>  1.0203e+00  4.0920e-01 -2.6337e-01 -7.6577e-02 -8.3448e-01 -1.0251e+00
#> -1.9481e-01 -3.4621e-02 -2.6708e-01  4.0083e-01 -1.2155e-01 -1.2188e+00
#>  5.7693e-01 -6.2163e-01  1.8246e+00  7.8966e-02  1.3534e-01 -4.0642e-01
#> 
#> Columns 13 to 18 1.0477e+00  1.1657e+00  6.1262e-01  2.3804e-01  1.1071e-02  1.7847e+00
#> -1.7350e+00  1.5428e+00  5.7291e-01  1.5808e-01  1.0970e+00 -5.7280e-01
#>  2.7553e-01  1.2624e+00 -1.0640e+00  1.3352e-01  1.8938e+00 -2.6654e-01
#>  1.3160e+00 -4.7216e-01 -1.0657e+00  6.3279e-01  9.4945e-01  1.4578e-01
#> -1.0938e+00 -1.0485e+00  6.5000e-01  3.7005e-01  9.5743e-01 -1.7306e+00
#>  6.1449e-02 -9.3433e-01 -1.7407e-01  4.7130e-01  4.6941e-01 -1.7182e-01
#> -1.7290e+00 -1.0889e-01 -1.0333e-02 -9.5815e-01 -6.1368e-01 -2.4934e-01
#>  1.3694e+00  5.8803e-01  5.5112e-02 -1.7869e+00 -4.0434e-01 -1.5908e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.2862
#> -0.1121
#> -0.3542
#> -0.1502
#> -0.4027
#> -1.1900
#> -0.9690
#> -3.1857
#>  0.0276
#>  1.1755
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -3.1857
#> -0.1121
#> -0.1502
#>  1.1755
#> -0.9690
#> -0.4027
#>  0.0276
#> -1.1900
#>  0.2862
#> -0.3542
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>