Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.3339  0.2766 -1.9041  1.7917 -0.1344  0.8407 -0.3183 -0.1293  0.3099  0.4668
#> -0.0581  0.3172 -0.5595  0.1346 -0.7373  0.9025  1.1411  1.5891  0.6507  1.9587
#> -0.7155 -0.3677  0.3456  0.1549 -1.8618 -0.9547 -0.7143 -0.1759 -1.7690  0.9796
#>  0.3359  2.1807 -0.0860 -0.0544 -1.0292 -0.5421  0.0253 -1.2106 -0.6167 -0.0826
#> -0.8885  1.3839  0.0240  1.1596  0.7816 -0.4044 -0.2899  0.7994 -1.8941 -0.8663
#>  1.3382 -0.6346  0.0904  0.2452  0.3132 -0.2097  0.5375 -1.3671  0.2003 -0.3115
#> -0.4612 -0.9193  1.9243  0.4155 -0.1347 -1.0236 -0.9571 -0.6900  1.3890  1.0298
#>  0.5348  1.0898 -1.6901 -0.1616 -0.8906 -0.1376 -0.3369 -1.5882  0.0305 -0.1001
#> -0.6456  0.3319 -0.3280  1.0123 -1.5406 -0.4034  0.6647 -0.4960  0.0262  1.3575
#>  0.3709  0.8655  0.2074 -0.3996 -0.4882  0.3775 -0.2231  0.4091  1.6517  0.9494
#> 
#> Columns 11 to 20-0.2193  0.9071 -1.1677 -0.3050 -0.3629  0.6817 -1.0255  0.0360  0.4459 -1.2311
#> -0.8969  0.6956 -0.4013 -2.0603 -0.1767 -0.7965  1.0387  1.3055  0.3103  0.4106
#>  0.0581 -2.0278 -0.7088  0.9276  0.4757 -0.0785 -1.9945  0.3203  0.2308 -1.3717
#> -0.2296 -0.2386 -1.0007  0.2686  0.8614  0.6579 -0.2686  0.5040  0.3352  0.4079
#> -0.5237 -1.7025 -1.9960 -0.0665  0.4088 -0.8512 -1.3613 -0.4796 -0.9972  0.4776
#>  0.6194 -1.4572 -0.1122 -1.0227  1.4454 -0.7210  0.2560  1.8734  0.6210 -0.5835
#>  0.3277  1.7074  1.0172 -0.1577  0.3278  0.1705 -0.3214  0.8479 -0.9254  0.1731
#> -1.1090  0.4984  0.8804 -0.6413  0.6248 -0.2069  0.8145 -1.3985 -1.2907 -2.2704
#> -1.6107  1.2687  0.2806 -0.0915 -0.1273  1.7109  0.5448 -1.7696  1.0161  0.3247
#>  0.2407  2.5008 -0.0776  0.0195 -0.3098 -0.6034 -2.2514  0.5321  0.0660  1.3895
#> 
#> Columns 21 to 30-1.7002  0.2236 -0.7959 -0.7010  0.0102  0.4510 -1.7623  0.6363  0.2974  0.0300
#> -1.0864 -0.6517  0.5927 -0.2855 -0.9319  0.3360 -0.6110  0.7409  1.0392 -1.4804
#> -0.3689 -0.7815  0.8441 -0.0452 -0.2140  0.5410  0.6769 -0.7159 -2.3666  0.5011
#> -0.4441  2.5121  0.8634  0.2118  1.6795 -0.9861 -0.1425 -0.6362 -0.0094  0.8733
#>  0.3474 -0.2449 -1.0377 -0.0916  0.7254 -1.1359 -0.1962  0.8588 -0.6687 -0.3886
#>  0.4976 -0.5496  1.0117 -0.2181 -0.0913  0.4857 -0.7263  0.1139  0.3704  0.1092
#> -0.1381 -1.4563  0.0928 -0.1017 -1.0692 -0.2777 -0.1778  2.7436  0.0210 -0.0542
#> -0.8402 -0.7044 -0.0391  0.6265  0.1393  0.6809 -1.5656  0.5106  0.3147 -1.2449
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -1.7634
#> -1.4077
#> -0.6723
#> -0.5647
#>  0.0015
#>  1.1661
#> -3.0650
#>  0.4069
#>  0.2263
#>  0.9861
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.4077
#> -3.0650
#> -1.7634
#>  0.0015
#>  0.9861
#>  1.1661
#> -0.6723
#>  0.2263
#> -0.5647
#>  0.4069
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>