Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.3853  0.6044 -1.4013 -0.6588 -1.4272  0.5675 -0.5223 -0.7837  0.7011 -0.6788
#> -0.6889 -0.3792  1.5805  0.4323  0.9920  1.2376 -0.3938 -0.6521 -0.1037  1.0847
#> -1.1800 -0.7838  0.0117 -0.3618  0.8281  0.0816  0.0088 -0.1923  0.3116 -0.1607
#>  0.9390 -0.6209 -0.4414 -0.0759 -0.9379 -0.4381 -0.3576 -1.3129  1.0387 -0.2890
#>  0.4612 -1.4379 -0.2003  0.0359 -0.6435  1.6715  0.3041 -1.4099  0.3242 -0.7352
#> -1.6715 -1.3216  0.3134  0.2698 -0.7189  0.4224 -0.4521 -0.9822 -0.0340 -0.9928
#> -0.1883  0.9858 -1.0129 -0.5601 -1.0505  0.2485 -0.3650 -1.1132 -0.2081  0.7062
#>  0.8157 -0.2904 -1.5657  0.4879 -1.0433  1.6754  0.2183 -1.2111  0.9526  1.2055
#>  0.5070  1.1287  0.5918 -2.0217 -0.7289  1.2814  0.3450 -0.8741 -0.0840 -0.3566
#>  0.1941  1.3595  2.2871 -0.1863  0.1997  1.3669 -1.4616 -1.5411  0.5658 -0.8817
#> 
#> Columns 11 to 20 0.4917  0.4481  0.1140 -1.7307  0.7451  0.2033  1.0206 -0.1179  0.6884 -0.0030
#> -1.4956  2.0429 -2.4853  1.1968  2.8699 -0.3851 -0.2078  0.6108  0.5675 -1.5463
#> -0.3939  2.1726 -0.6563  0.1587  0.6644 -0.8479 -2.8526  0.5325 -0.1223 -1.4486
#> -0.3293 -1.4843 -0.5311 -0.8768  0.9912  1.3233 -0.7889  0.9068 -0.6045 -0.0769
#>  0.1151  0.7726 -0.6274 -0.4205 -0.5234 -1.1563 -0.2636  0.8364 -0.7542  0.5711
#> -0.6956 -0.8864 -0.4939 -1.8258 -0.5252  1.5051  0.8138  0.0936 -0.0818 -0.4757
#> -0.4633 -0.6361 -0.3591  0.3738 -1.6406 -2.0849 -1.5819  0.1153  1.9630  1.2529
#>  1.0116  0.3787  1.0642 -1.2028 -0.6950  0.0339 -0.2316 -1.4259  0.9941 -0.3353
#> -0.9100 -1.1689  0.1310 -1.8411  0.4218 -1.5918 -0.9576 -0.5696 -0.9434  0.4330
#> -0.9497 -1.7219  1.0272 -0.8391 -0.1852  1.8511 -0.1097  0.5283 -0.5685 -0.1747
#> 
#> Columns 21 to 30 0.8800  0.8496  0.1127 -1.6653  0.2365  0.0037  1.1036  1.8154 -1.6581 -0.0120
#>  0.3708  0.1574 -0.1001  0.3237  0.3858 -0.6112 -1.4021 -0.0981  0.6681 -0.0843
#> -0.4178 -0.9709 -0.0472 -0.9336  0.2304  0.0632 -0.3918  0.6802 -0.1449  0.2820
#> -1.3476 -0.1947 -1.2842 -1.5655  0.2656 -0.7660 -0.4068  0.3887 -1.9822 -0.9312
#>  2.1419  0.2054 -1.2628 -0.7875  0.0927  1.1969 -0.3652  0.2068  0.1807 -0.0215
#> -0.5337  1.0397  1.0377 -0.5159 -0.6194  0.3504 -0.5037  0.0028 -2.0291  0.7083
#>  1.8304 -0.8043 -0.7214  0.8252  0.3165  0.5417 -0.3191 -0.2474  0.3740  0.1852
#> -0.2783  1.6344 -1.2445  0.6217 -1.2195 -1.2578  0.4044 -1.0174  2.3290 -0.4473
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.8219
#> -1.7523
#> -0.1563
#> -0.4960
#> -1.0309
#>  0.6857
#> -1.3548
#> -0.1528
#>  0.2805
#>  1.5041
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.0309
#> -0.1563
#> -1.7523
#> -0.8219
#> -1.3548
#> -0.4960
#>  1.5041
#>  0.2805
#> -0.1528
#>  0.6857
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>