Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 1.4153e-01 -1.2423e+00 -4.2530e-01  9.5276e-01 -1.5392e+00 -1.4710e-01
#>  1.2466e+00  9.5967e-01 -5.1133e-01 -9.6976e-01  1.9275e+00  3.9306e-01
#>  9.6124e-01  1.1228e+00  6.2509e-01  1.7336e+00  5.0262e-01  4.5372e-01
#> -2.8984e-01  1.0875e+00  6.0134e-01  1.6763e-02  3.0314e-01 -1.0918e+00
#> -7.1908e-01 -3.2471e-02  1.8824e+00  8.0106e-02 -7.2135e-02 -1.4344e+00
#>  2.5207e-01 -1.0943e-01  1.0442e+00  5.4580e-01 -1.5853e+00 -8.3012e-01
#> -1.9857e+00 -1.1473e+00  5.5587e-01 -5.3899e-01 -6.7591e-01 -1.1672e+00
#> -1.0558e+00 -6.5991e-01  1.5484e-01  3.2897e-01 -6.8677e-01  2.3925e-01
#> -1.5598e+00 -9.0786e-01 -1.5851e+00  8.6070e-01 -1.6376e+00 -1.4142e-02
#> -5.7694e-01  3.6113e-01  8.7526e-02 -2.6386e-01 -2.7311e-01  1.2895e+00
#> 
#> Columns 7 to 12 4.9061e-01 -5.4312e-01  3.3496e-01 -1.5955e-01  5.2982e-01 -4.6370e-01
#>  1.0255e+00 -1.3720e-01 -4.2237e-01 -1.1313e+00  1.9048e-01  5.0690e-01
#> -5.5514e-01 -1.2938e+00  4.2353e-01  3.4266e-01  1.0467e+00  5.7940e-01
#>  3.2124e-01 -8.9770e-01  8.7726e-01  9.7179e-01 -8.2251e-01 -3.8427e-01
#>  2.5530e-01  8.7088e-01  1.3717e-02 -7.9884e-02  2.4048e-01 -7.7813e-01
#>  2.9403e-01  7.3381e-01 -8.1614e-01  1.5238e+00  3.9267e-01  9.4313e-01
#>  1.9463e+00 -3.2273e-01  2.5708e-01  3.6312e-01 -2.9152e-01  1.7608e-01
#> -1.4904e-01  1.5427e+00 -1.7450e-01  2.2554e+00  2.4035e-01 -9.0184e-02
#> -3.6470e-01 -5.7249e-01 -5.8431e-01  1.9775e+00  7.6392e-01 -8.4846e-01
#> -3.3388e-01  7.8328e-01 -1.8293e+00  1.8201e+00 -7.8135e-02 -2.1152e-01
#> 
#> Columns 13 to 18-6.8170e-01 -3.9092e-01 -3.6412e-01  9.1169e-01 -7.4863e-01  6.5247e-01
#>  1.0919e+00 -1.2914e-01 -8.5321e-01 -5.3449e-01  1.1190e-01  5.2328e-02
#>  2.3007e+00  2.6828e-02 -4.2921e-01 -9.7520e-01  7.1998e-01  4.4824e-01
#>  1.6352e-01  1.4017e+00  1.0246e+00 -1.1392e+00  7.1311e-01 -7.2681e-01
#> -5.6241e-01 -1.2028e-01 -1.9623e+00 -1.5961e+00  5.7967e-01  1.1210e+00
#> -2.5073e-03 -4.0816e-01 -1.1887e+00  1.7342e-01  2.8500e-01 -1.2817e+00
#> -7.7091e-01  5.7988e-01 -1.8345e+00 -1.8844e-02 -2.6083e-01 -4.9685e-01
#>  2.3589e-01 -2.1774e+00  1.0843e+00  1.0927e-01  5.1907e-01  1.5492e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.4240
#> -0.7962
#> -0.8371
#> -0.7065
#>  0.7369
#>  1.2693
#>  1.5491
#> -1.1492
#>  1.1434
#> -0.1714
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  1.1434
#>  1.4240
#> -0.7962
#> -0.8371
#>  1.5491
#> -1.1492
#>  0.7369
#>  1.2693
#> -0.1714
#> -0.7065
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>