Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.3284  2.3338 -1.3417  1.1281 -0.4488 -0.7338 -0.7415 -0.2289 -0.4620  0.1246
#> -1.1247  0.3881  1.1121  0.9585 -0.2988  0.6337  0.0636 -0.5311  0.4611 -0.4563
#>  0.6057  1.8179  0.3030  1.3878 -0.8253  0.3369 -1.1910 -2.1267 -0.3963  0.1680
#>  0.2244  1.0574  1.3038 -2.0589 -0.2963 -1.9449  0.4130  0.5367 -1.0985  0.0689
#>  0.2164  0.3596 -0.8533 -1.0470 -1.3026 -0.7000 -0.5153 -0.1398 -0.5404 -0.5428
#> -0.6886  1.7022 -0.6257 -0.2195 -0.3792  0.4442  0.6005  0.6890  1.8315  1.4338
#> -0.2497 -1.1254  0.3921  0.7717 -0.1186  0.4436  1.6973  0.2579 -2.1567 -0.0252
#>  0.2315 -0.2016  0.6336 -1.6960  1.6547 -1.0550  0.4728 -0.4102 -0.7926  1.4853
#> -0.1204  1.1590  0.1291 -0.1294  0.1133 -1.8242  0.1052  0.5622 -1.0813  0.5789
#>  0.3027 -0.3843  0.1098  0.2575 -1.1770 -0.0012  0.3958 -0.8655 -0.0442 -0.2920
#> 
#> Columns 11 to 20 1.1188 -0.2037 -0.7219 -0.1152 -0.2299  1.5915  0.0932 -1.3389  0.1359  1.5214
#> -0.2751 -0.7347 -0.8684  0.3835  1.6337  0.9313 -0.3421 -0.1702 -0.0214 -0.2515
#> -1.8685  0.3457 -0.5107  0.2936 -1.0133 -1.1920  0.9081 -0.9664 -1.1328  0.1980
#>  0.4346  0.4294 -1.2684  0.4278 -0.8286  1.3694  1.1955  0.0448 -0.2055 -0.5865
#> -0.8252 -0.0503 -0.4449 -0.1161  1.0412 -1.2013  0.9899 -0.1721 -0.3019  0.4327
#>  0.0335  2.6302 -2.1975 -0.0528 -0.6396 -0.5547 -0.9383  1.1209  0.1322 -0.0100
#> -0.6784  0.2552  0.9089 -1.3591 -0.3842 -0.9269  1.2495 -0.2553 -0.0214  0.7625
#> -0.3541 -0.8004  1.1738 -0.3372 -0.9757  0.0716 -0.9849 -0.8310 -0.2040  0.3927
#> -0.5780 -0.8265 -0.5064 -0.0639  1.1755 -0.4682 -0.8747 -1.1526  0.0246  1.5382
#> -0.5196 -0.9368  0.0216  2.6370  0.7601  1.4180  0.3147 -0.5963 -1.5616  0.0948
#> 
#> Columns 21 to 30-0.2238 -0.2202  1.0963  1.1009 -0.6857  1.4041 -0.2988  1.5129  0.3921 -1.1235
#> -0.1462  1.0291 -0.6929  0.1970 -0.9967  0.8398  0.0543 -1.1102  0.0872 -0.9816
#> -0.7726 -0.5232  0.3953 -0.7663 -1.0105  0.8040  1.9318  1.0778 -0.2136 -0.3180
#> -0.6926 -0.8096  0.5211 -1.4885  0.5566 -0.1421  1.0015  0.0244  0.1321 -0.6068
#>  0.3075  1.1107  1.5262  0.5378 -0.9423 -0.5483  0.5264  0.0937  0.4411  0.6939
#>  1.1237  0.3297  0.0685  1.0609 -1.0158 -0.8228  0.2382  0.5847  0.1635  1.6293
#>  1.2384 -0.9379  0.5659  1.3027  0.1935 -1.2970  0.2040  0.4756 -0.9959  0.1362
#>  0.3444  0.0903  0.0933 -0.7451  0.5282  1.0808 -0.2937  0.4472  2.0964  1.2387
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.9777
#> -2.0125
#> -0.8277
#>  0.4617
#> -1.9906
#>  0.2552
#>  0.6712
#> -0.7601
#> -0.9768
#>  1.1296
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.9768
#>  0.9777
#> -0.8277
#>  0.2552
#>  1.1296
#>  0.6712
#> -0.7601
#> -1.9906
#>  0.4617
#> -2.0125
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>