Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-6.0361e-01  6.4629e-01 -8.1094e-01 -1.1255e+00 -9.0437e-01 -1.3860e+00
#> -9.6464e-01 -2.1454e-01  1.4928e-01  6.3560e-02 -1.6420e+00  3.2559e-01
#> -1.0241e+00  8.8175e-01  1.1791e+00 -1.1503e-01 -8.9774e-01 -7.4474e-01
#> -1.0792e+00 -1.7628e+00 -1.2132e+00  3.3215e-01  2.9562e-01 -1.6384e+00
#> -1.5722e+00 -1.4153e+00  5.5036e-01  4.0297e-01 -1.0831e+00  5.3443e-01
#> -1.4588e+00 -1.4609e+00  9.2576e-01 -1.6471e-01  1.6825e-01 -1.2758e+00
#>  4.5976e-01  3.0150e-01 -3.3320e-01 -7.7282e-01 -1.2288e+00  9.7038e-01
#> -8.5970e-01 -1.2355e+00  1.6269e+00 -2.1236e-01 -2.9685e-02 -1.6938e+00
#>  9.0700e-01 -2.2095e+00  6.9243e-01  1.9512e-01  2.4530e-01  1.4217e+00
#>  4.3756e-01 -8.3114e-01 -1.1505e+00 -1.2570e-01 -3.8006e-01  8.4138e-01
#> 
#> Columns 7 to 12 2.3044e-01 -9.4001e-01  3.7115e-01 -1.1913e+00  1.4269e+00  1.4354e+00
#>  3.8161e-01 -1.5236e-02 -1.5074e-01 -1.1869e+00  2.7619e-01 -7.4600e-01
#>  1.1292e+00  8.5080e-01 -3.7440e-02 -3.4553e-01  6.9893e-01  1.6907e+00
#>  7.0924e-02  3.9156e-01 -6.9490e-01  8.5289e-01 -1.3834e+00 -9.2141e-01
#> -7.4851e-01  1.9368e-01 -2.0355e-01 -8.2114e-01 -1.3667e+00  2.9626e-01
#> -7.6605e-01  5.7544e-01  7.7699e-01  1.8051e-01  8.9699e-02  2.3803e-01
#> -3.4253e-01 -9.6892e-01 -9.5503e-01 -3.7958e-01 -5.1565e-01  3.7501e-01
#> -6.2172e-01  4.0715e-02 -4.3632e-02  1.5084e+00 -1.0437e+00 -1.7686e-01
#>  5.1470e-01  5.1463e-01  1.1159e+00 -5.0756e-01 -3.0860e+00  2.8814e-01
#> -9.4280e-01  6.9432e-01 -1.1224e+00 -1.3399e-01  6.0824e-01  3.8680e-01
#> 
#> Columns 13 to 18 4.1159e-01  3.7380e-01 -2.7033e-01  7.0117e-01  1.4235e+00  1.3768e+00
#> -6.0415e-02  1.5301e+00  4.0602e-01  3.5020e-01  1.3147e-01  4.0205e-01
#>  7.6502e-01 -1.6216e+00  1.7621e+00 -2.0395e-01 -2.4070e-01 -4.8663e-01
#>  9.4106e-02  1.8321e+00  1.2591e+00 -2.9505e-02  1.0025e+00 -9.0058e-03
#>  5.6673e-01  1.2001e+00  1.2696e+00 -7.8492e-01 -4.4196e-01 -3.6271e-01
#> -6.9541e-01 -1.2706e-01  3.3506e-01 -4.6217e-01  4.1549e-01 -8.3892e-02
#> -4.7222e-01  6.2531e-01 -4.9147e-01  6.1144e-01  1.6570e+00  2.3585e-01
#>  8.4109e-01  1.1527e-01  1.1033e+00 -1.4251e+00 -8.7652e-01  1.6141e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.5059
#>  0.0400
#>  0.6419
#>  1.2185
#>  1.2473
#> -1.5125
#> -1.1064
#> -0.4763
#> -0.6577
#> -0.5430
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.0400
#> -0.5059
#> -1.1064
#>  1.2473
#> -1.5125
#>  0.6419
#> -0.5430
#>  1.2185
#> -0.6577
#> -0.4763
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>