Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.3851  1.5219 -0.9184  0.6196 -0.1012 -0.5261 -0.8837 -0.6506 -0.2008 -0.0904
#> -0.6214 -0.5214  0.7306 -0.3995  0.5097 -0.9790  0.2016  0.2301  0.6046 -0.9127
#> -0.5543 -2.9844 -0.4410  1.2187 -0.2201  1.2892 -1.3277 -0.2442 -0.4487  0.9587
#>  1.3590 -0.1936 -0.6599  0.2081 -0.4367 -0.6401 -2.4341  0.0029 -1.0597 -1.2968
#> -0.4359 -0.5536 -0.1008 -0.4101 -0.3107 -1.1785 -0.1430  1.4475  1.1621 -0.5601
#> -0.9222 -0.5574 -0.1988  2.1730 -0.2718 -0.1111 -0.3776  0.1430  0.5059  0.6844
#>  0.2767  0.2386 -0.5891  1.8395 -0.9365 -0.5219 -0.6907 -1.5060 -1.1682 -0.1545
#> -0.7841  0.1391 -0.0905 -0.2759 -0.4532 -1.6289  1.0145 -0.2351 -0.9076  1.1697
#>  1.6041 -2.8994 -0.4849  1.5116 -0.2652  0.9123 -0.8715  0.2720 -0.3333 -0.5197
#>  0.4684  0.2335 -0.7403  1.3648  0.1696  1.4019  0.3187  0.2826 -0.5445  0.1425
#> 
#> Columns 11 to 20-0.5648  0.7195 -1.3127  0.3469 -0.4457 -0.7545  1.5942 -0.7418  0.9042 -0.3167
#> -0.8014  0.4709  0.3001 -1.7000 -0.2099 -0.2111  1.0612  0.8295 -1.3065  1.5536
#>  2.1431  0.5859 -0.1704 -0.3552  0.6759 -0.3021  0.4214 -0.6249 -0.3953  0.7242
#>  0.2479 -1.4141 -0.1670 -0.4337 -0.8735  0.1000 -0.7234 -0.6539  0.0829 -1.1979
#>  2.1049 -0.8303 -1.8021 -0.2634  0.6690  0.9945 -0.6582 -0.0463  0.3379  0.1006
#>  1.6299 -0.4782 -0.1860 -0.8626  0.6208  0.6187  0.3976 -1.1996 -0.2676 -0.5984
#>  0.1763 -0.0102  0.4232  0.1013 -0.6881  0.4248 -1.2868  0.4772 -0.2777 -1.1187
#> -0.9138 -0.0582 -1.0304 -0.7609  0.5276 -0.0428  0.1173  1.8134  1.0568 -0.2449
#> -0.8478 -0.3072 -0.8842  0.6739 -1.9979 -1.7262  0.6097  0.8727  0.1098  2.4355
#> -0.8392  1.1671 -0.2430  0.8130 -0.3504 -0.4138  0.9033 -1.2036 -1.6019 -0.0240
#> 
#> Columns 21 to 30-0.3340  0.0754  0.7157 -0.8999 -0.1430 -0.0167  0.9213  0.3801 -0.4000  0.4564
#> -0.3199  0.2601 -0.8069 -1.2122  0.5243 -0.4746  0.2992 -0.7467 -0.5347  0.8178
#> -0.8970  0.6254 -1.1560  0.5015 -2.3630  0.2458  1.4790 -0.9210 -0.7368 -0.0445
#>  0.8814 -1.9924  0.8176 -1.2193 -0.9326  1.3646  0.9815 -1.4619  0.5499 -0.4665
#>  0.8356  1.0923 -0.4417  0.4824  0.2082 -0.9840  2.1992  0.3333  0.0282 -2.1276
#> -0.1625  0.3499 -0.1091  1.1563 -1.1674  0.4803 -0.1988  0.2631  0.1797  0.3273
#> -0.1436  0.3641  0.4966 -0.0453 -0.8134  0.8165  1.0071  0.9810 -1.2333  0.7224
#> -0.0958 -0.0202 -0.3708  0.7309 -1.6195  0.5253 -1.8615  0.0796  0.5925  0.7401
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.0029
#>  0.7950
#> -0.1347
#> -0.5683
#> -1.1740
#> -0.4395
#>  0.8149
#> -1.1359
#>  0.4710
#>  0.7064
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.8149
#>  0.4710
#> -0.4395
#> -1.1740
#>  0.7950
#> -1.1359
#> -0.5683
#> -0.1347
#>  0.7064
#> -0.0029
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>