Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 1.5142e+00  4.0901e-01 -1.6555e+00 -8.9678e-01 -5.8064e-01 -6.8705e-01
#> -4.4078e-01 -2.1479e+00 -4.3890e-01  1.6934e-01 -6.9693e-01 -1.7988e+00
#> -8.9653e-02 -1.3933e+00 -1.0388e+00  3.6654e-01 -3.0805e-01  2.6509e-01
#> -5.2303e-02 -2.2839e-01  5.8525e-01  1.9377e+00 -6.9435e-01  1.1539e+00
#> -9.9967e-01  4.2458e-01 -4.3384e-01 -7.5287e-01  5.3525e-01 -8.5194e-01
#>  5.6734e-01 -1.5962e+00  9.5673e-01  1.6509e+00 -1.6310e-01 -2.5839e-01
#> -1.5095e+00  3.9315e-01  6.3172e-01  2.9330e-01 -1.7017e+00  1.0351e+00
#> -5.7839e-01  2.0497e-01  3.6289e-01 -1.3326e+00 -6.8274e-01  1.4414e+00
#> -1.0133e+00  4.1170e-03  1.8555e+00  1.2167e+00  1.4479e+00 -1.9892e+00
#>  8.6568e-02  2.2619e+00  1.0549e+00 -8.9993e-02 -1.6232e+00 -2.0710e+00
#> 
#> Columns 7 to 12-7.9677e-02 -1.0157e+00  3.0721e-01 -3.8995e-01  1.5751e-01  4.3613e-01
#> -7.5101e-01 -9.1126e-03  1.7959e+00 -1.0444e+00 -2.0452e-01  3.8945e-01
#>  1.3920e+00 -4.4917e-01 -8.2794e-01  1.1840e-01  3.5642e-01  8.8140e-01
#>  3.6976e-03  1.3859e-01 -8.0253e-02 -3.0675e-01 -6.6394e-01  6.7393e-02
#> -5.6724e-03  1.4311e-01  6.5577e-01 -6.9616e-01 -1.3116e-01 -5.0466e-01
#>  9.0835e-01 -1.2740e+00  6.5208e-01  9.8030e-01 -1.1214e-01 -3.7400e-01
#>  1.2770e+00  6.7793e-01  1.7554e-01  1.5612e-01  1.8588e-01  7.0938e-01
#> -1.2988e+00  2.3348e-01  1.4425e+00  3.0777e-01 -1.5365e+00  1.1166e+00
#>  3.6915e-01 -1.6312e+00  2.3571e-01 -9.9206e-01  3.4669e-02 -3.6680e-01
#> -9.9494e-01  7.8221e-01 -1.8817e+00 -8.4095e-01  7.8489e-02 -6.1893e-01
#> 
#> Columns 13 to 18-5.1967e-01  5.1524e-01  8.2781e-01 -1.0560e+00  2.7403e-01  1.0963e+00
#>  1.0475e+00  4.4877e-01  1.1649e-01 -1.1872e+00  3.2092e-01  5.4191e-01
#> -5.3863e-01 -1.3605e-01 -2.4751e-01 -5.3881e-01  2.7546e-02  9.1000e-02
#>  6.1786e-01 -1.9640e+00  1.1924e+00 -1.2167e+00  1.1790e-01 -8.5185e-01
#> -1.4488e+00  8.9737e-01  1.4742e-01 -1.1808e+00  2.2643e-01 -5.3876e-01
#>  4.9922e-01  2.9353e-01 -2.4487e+00 -1.9206e-01  7.9402e-01 -2.9357e-01
#>  1.3446e+00  1.1265e+00 -5.9911e-01  7.6755e-01 -1.4092e+00  4.9819e-01
#> -1.4055e+00 -6.6210e-01  9.0986e-02 -1.2257e+00 -8.2243e-01  6.1130e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.1431
#>  0.3886
#> -0.0832
#> -1.0233
#> -1.2499
#>  1.3776
#> -0.5550
#>  0.7812
#> -0.3674
#>  1.5199
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.3886
#> -1.2499
#>  0.1431
#>  0.7812
#> -0.3674
#> -1.0233
#> -0.5550
#> -0.0832
#>  1.3776
#>  1.5199
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>