Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.0806 -0.7500 -0.7730 -0.5681  0.2101  0.3127  0.1822 -0.5558  0.6633 -0.0175
#>  0.2466  0.4280 -0.2121  0.0371 -1.2640  0.7861  0.5376  1.2183 -0.3778 -1.6489
#> -0.2841  1.3402 -1.4925  0.8857  0.6275 -0.4837 -0.4788 -1.5670  0.1072 -0.3592
#>  0.3329 -0.5059 -1.2180 -0.4024 -0.8164  0.1172  0.5192  0.7845 -0.4153 -0.1834
#>  0.6140 -0.8311  1.1401 -0.7022  0.1329 -0.3704 -0.2151 -0.2533 -0.2939 -0.1767
#>  0.8187  0.2313  1.1822 -0.3139  0.9678  0.6197  1.3420 -1.1793 -1.9032 -0.7914
#> -0.5004  0.7580  0.8665 -0.0284  0.6813  0.0893  0.7781  0.8802 -0.0094 -1.7354
#>  1.2367 -0.5267  0.8330 -0.9264  1.3704  0.1658  1.3846 -1.0466 -0.1946  0.0997
#>  0.4756  0.7301 -0.4034 -1.0955 -0.6107 -0.1937 -0.5011 -1.7791 -1.0308  0.1108
#> -0.0219  0.1388  0.8700  0.6203  0.3382  0.0063 -0.5497  0.0585  0.8684 -0.2345
#> 
#> Columns 11 to 20-0.3618 -0.8462  1.2814 -0.8728 -0.2306 -0.4565  0.0102 -0.8847 -0.4949 -0.2611
#> -0.3407 -0.3410  0.4023 -1.0977 -0.4258  0.2583 -0.8629 -0.2283 -0.7592 -0.0197
#> -1.0688  0.4219  1.7603  0.4541 -0.9258  0.7016 -0.5899 -1.5918 -0.5589  0.9255
#>  1.5051  0.6279 -0.4422 -1.4603  0.1698  1.6756 -0.6134 -0.1524  0.1517 -1.2240
#> -0.5046  0.1827 -0.3950 -1.9523  0.7665  0.5206 -0.2778  0.0578 -0.6889 -0.1105
#> -0.9112 -1.0759 -0.3349 -0.2989 -0.0589  1.7845  0.9301 -1.5439 -0.0381 -1.1854
#> -0.9856  2.4626 -0.3074 -1.5671 -1.5552  1.0224  0.6521 -0.4058 -0.1940 -0.7800
#> -0.7818  0.0586  0.7894  0.5724 -0.0084  1.6065  0.6205 -0.2238 -0.3456 -0.7675
#> -1.1475 -0.3111 -0.7081 -0.9278  1.8049  0.4236  0.1413 -2.4716 -0.2339  0.2827
#> -0.0746 -0.2402 -2.1875 -2.4039 -0.4224  1.7809  0.8097  0.5007  0.2725  0.7950
#> 
#> Columns 21 to 30-0.3635  0.7441  0.1418  1.6956 -1.7871  0.1919 -0.0231 -0.2087  0.5481 -0.0344
#>  0.7620  1.3103 -0.7342 -0.2606 -0.1522  1.1586  0.4597  0.1326  0.2008  0.9314
#> -1.1206  0.2905 -0.2283 -0.6222 -0.2662 -0.2558  0.7260 -0.5888  0.3590 -0.1327
#> -0.1274  0.6231  0.2917  1.1411  0.3157 -1.1089 -0.7222 -0.1439  0.9525  0.0981
#> -0.0705  0.5643  0.4672 -0.2578  0.5278  0.6525 -1.5534 -0.9301 -0.5930 -0.1004
#> -0.4063 -1.1713 -0.6707 -0.0578 -1.2631 -1.0343  0.7426 -0.8686  1.9559  0.0204
#> -0.5961  0.0680 -0.0757 -1.1016 -0.5493  0.4441 -0.3838  1.8652  0.3375 -0.7904
#>  0.7230 -0.6486 -1.1789 -0.0549 -0.8665  0.4595  0.1277  0.0063 -0.7028  0.7706
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.9485
#> -0.2528
#> -1.3393
#> -0.6107
#>  0.0107
#> -0.9859
#> -0.3175
#>  0.0133
#> -1.3726
#> -0.6532
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.9859
#>  0.0107
#> -0.9485
#> -0.3175
#> -0.6107
#> -1.3393
#> -0.2528
#>  0.0133
#> -0.6532
#> -1.3726
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>