Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-4.9192e-01 -6.3190e-02  1.0863e+00 -1.0424e+00  1.4247e+00  9.5794e-01
#> -4.8316e-01 -7.8610e-02 -1.3913e+00  9.0487e-01 -7.1783e-01  7.4476e-02
#>  1.4147e+00 -6.0492e-01 -2.0227e-01 -5.3188e-01  7.2357e-01 -2.6152e-01
#>  7.8401e-01  1.3074e-01 -3.7700e-01  8.4978e-01  2.6218e-01 -7.7859e-01
#> -6.3937e-01 -3.8520e-01 -7.3973e-01 -1.3811e-01 -1.3087e+00  1.0239e+00
#> -4.3401e-01 -2.1526e+00 -5.4880e-01  2.8204e+00  7.8604e-02 -1.6776e+00
#> -1.3022e+00 -7.3508e-01  1.5950e-01 -8.9609e-03 -3.4524e-01 -6.0831e-02
#> -7.1744e-02 -9.5758e-02 -1.2296e-01  8.2036e-01  5.4920e-01  1.4391e+00
#> -8.1588e-01 -1.7436e+00  1.7093e-02  5.3120e-01 -3.0932e-01 -1.0660e+00
#>  1.6034e+00  8.2070e-01 -1.1486e+00 -7.4785e-02 -1.3200e-02 -1.3927e+00
#> 
#> Columns 7 to 12 9.6024e-01  9.5029e-01 -9.2564e-01  2.4773e-01 -1.0281e+00 -1.5204e+00
#>  2.1641e+00 -6.1233e-01  3.3843e-01  2.5390e-01 -1.4178e+00  1.8073e-01
#>  1.2568e+00 -1.6925e+00  2.9455e-01 -3.1042e-01 -9.0673e-01  7.7568e-01
#> -3.7052e-01  3.9637e-01 -3.4947e-01 -2.8855e-01  2.4556e-01 -2.2530e-01
#> -9.0290e-01 -1.5215e+00 -1.2834e-01  1.5325e+00  8.1966e-01 -3.6942e-01
#> -8.7864e-01 -1.2104e+00  2.5265e+00  1.7199e-01 -1.0807e+00 -8.6539e-01
#> -3.4222e-02  6.9698e-01  8.1224e-01  7.7678e-01 -6.9694e-01 -7.6653e-01
#>  7.9845e-01 -1.2888e+00  2.0721e-01 -3.0796e-01  1.0050e+00  1.6515e+00
#>  1.6437e+00  1.2365e+00  5.4036e-01  1.9283e+00  7.7729e-01 -1.6404e-01
#>  2.9648e-01  1.0773e+00  1.2677e+00 -1.4778e+00  4.7429e-02  4.0326e-01
#> 
#> Columns 13 to 18-5.9388e-01 -1.1705e+00  5.0530e-01 -1.1849e+00 -2.6828e-01  6.8366e-01
#> -1.3598e+00  7.1562e-01 -5.2261e-01  2.3817e-01  4.6355e-01  5.4196e-01
#> -2.3476e-01  1.5613e-01  1.0442e+00 -1.1751e-01  4.2462e-01  8.1886e-01
#> -1.0346e+00 -1.0416e+00  8.1134e-01  8.6262e-01  1.4128e-01 -6.8576e-01
#>  1.1962e-01 -7.5440e-02  1.6545e+00  6.7197e-02  8.9322e-01  8.0199e-01
#> -8.6920e-01 -9.2687e-01 -1.0190e+00  6.2433e-01 -3.4607e-01  4.3506e-01
#>  2.0697e+00 -4.7314e-01  1.7054e+00 -1.0355e+00 -6.6761e-01 -4.7746e-01
#>  9.7394e-01 -2.1193e+00 -7.7741e-01  7.5800e-01  5.8035e-01 -2.7260e-02
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.4038
#>  0.1423
#> -0.0973
#> -0.3053
#>  0.3292
#>  0.3731
#> -0.1541
#>  1.2835
#> -0.8082
#>  1.1705
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.1541
#> -0.8082
#>  0.1423
#>  1.2835
#>  0.4038
#> -0.3053
#> -0.0973
#>  0.3292
#>  0.3731
#>  1.1705
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>