Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 2.4230e-01  3.9914e-01 -1.5341e+00  8.1238e-02 -3.1426e-01  1.2046e+00
#> -1.2000e-01 -2.8859e-01 -6.9993e-01 -5.1452e-01  1.5115e+00 -9.6425e-01
#>  3.8860e-01  2.8038e-01 -8.5587e-01 -2.3745e-01 -1.0134e+00 -5.1850e-01
#>  7.0871e-01 -4.7179e-04  9.8453e-01 -9.7362e-01 -8.2629e-01  2.1049e-02
#> -6.6614e-01 -3.2368e-01 -9.4344e-01 -6.0514e-01 -7.2009e-01  1.1582e+00
#>  6.4044e-01 -1.1805e+00 -1.3129e+00 -8.9617e-01 -2.1999e-01  8.3407e-01
#>  2.3339e-01 -8.3543e-01 -1.3443e+00  6.1642e-02  5.7769e-01  9.2812e-02
#> -2.4868e-01  2.0863e-02 -2.4675e-01 -1.3757e+00 -1.2871e+00  3.8259e-01
#> -5.0574e-02 -1.7024e+00  9.1820e-01 -1.1869e+00  6.9568e-01  4.8932e-01
#>  1.5547e+00 -1.4155e+00 -3.4938e-03 -1.5975e-01 -1.1251e+00  3.5297e-02
#> 
#> Columns 7 to 12-1.2447e+00  1.7745e-01  1.6696e+00  4.5097e-01  3.3327e-01  1.1206e+00
#>  7.2130e-01 -8.6058e-01 -1.8429e-01  8.9707e-02  1.0398e-01 -4.5704e-01
#> -2.6312e-01 -5.0422e-01  3.4939e-01 -4.3912e-01 -5.6867e-01 -2.4172e-01
#> -1.1491e+00  8.9127e-01  2.6781e+00 -7.8902e-01  8.7274e-01 -1.7985e+00
#> -1.5172e-01 -1.1929e+00  1.2157e+00  8.3334e-01 -6.1992e-01 -1.7877e+00
#> -7.8516e-01 -3.9467e-01  1.0526e-02  2.3291e-01  4.7377e-01  5.3410e-01
#>  9.4746e-01 -5.6066e-02  2.0054e+00  1.4613e+00  6.9870e-01  1.4481e+00
#>  7.0734e-01  9.7732e-01  1.5883e+00  1.0695e+00 -2.1042e-01  6.0703e-01
#>  3.9633e-01  1.2822e-01 -5.5730e-01  9.9984e-01  9.6743e-01 -3.6280e-01
#> -3.8136e-01 -9.4608e-01  6.8934e-01  1.2512e+00  8.7928e-01 -4.7629e-01
#> 
#> Columns 13 to 18 1.3158e-01 -1.1298e-01 -1.2962e-01 -1.4629e+00 -9.2829e-01 -3.8941e-01
#> -4.9421e-01 -1.7240e+00  9.6559e-01  5.1325e-01  1.6277e+00 -9.3538e-01
#>  1.5012e+00  7.8923e-01 -9.1468e-01  1.4702e+00  3.4762e-01  1.1778e+00
#>  7.7366e-01 -2.4343e-01 -1.3362e+00  7.3048e-01 -8.4169e-01 -3.7213e-01
#>  6.4685e-01 -1.1167e+00  1.4028e+00 -4.2949e-01  1.3971e+00 -3.0267e-01
#>  7.8132e-01 -6.9730e-01  1.8364e+00 -7.6433e-01 -7.9465e-01 -6.9485e-02
#>  2.9753e-01  8.2184e-01  1.6944e+00  5.3494e-01  3.9794e-02 -1.8379e+00
#>  2.5528e-01 -1.0589e+00  5.0422e-01  4.7731e-01 -6.7681e-01 -1.0455e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.7151
#>  0.2373
#>  0.2868
#> -0.3863
#>  1.0459
#>  0.6388
#> -1.5007
#>  0.7436
#>  0.4167
#> -2.0814
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.2868
#> -1.5007
#>  0.2373
#> -0.3863
#> -0.7151
#> -2.0814
#>  0.7436
#>  1.0459
#>  0.4167
#>  0.6388
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>