Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.8273  0.0571 -0.3213  0.5784  0.1440  0.4430  0.9046  1.7846  1.5831  0.7036
#>  1.9781  0.4071  0.7161 -0.6548  0.3627  0.8212 -0.3843  0.6426  0.6392  0.2243
#>  0.8404  0.5886  0.1668 -2.0226 -0.6176  0.3565 -0.5383 -0.6217 -1.5614  0.0821
#> -0.9341  0.3865  0.9210 -0.1879  1.0158  2.0206 -0.7679 -0.2754 -0.8144 -0.0773
#> -0.9457  0.6070 -0.9593 -1.3328  0.0438  0.4133  0.2663  0.7866  0.0819 -0.0308
#> -0.6080 -2.0321  0.2829  1.7557 -0.7224 -0.7393 -0.4020  0.2352 -1.9557 -0.1830
#> -2.2813 -0.0857  1.2160  0.3537 -0.4836  1.1058  0.5472  0.3500  0.4253  0.1395
#>  0.1238 -0.2896  0.5217 -0.4008 -1.4975 -0.8275 -0.5954  1.1322  1.4301  1.3216
#>  1.1937 -1.1930  2.0124  0.4288  0.2633  2.0156 -0.1138 -0.0658 -1.5572 -0.5499
#>  0.9041 -1.0852  0.4760 -1.1046  0.0356  3.4169  1.9241 -1.3368  1.3909 -0.0477
#> 
#> Columns 11 to 20-1.1420 -0.6599 -1.0575  1.2935  0.7459  0.4650  0.9350  0.8226 -1.1056  1.9284
#>  0.4399  0.1038 -1.1472 -0.3833  0.0745 -0.4118 -1.2375 -1.3876 -0.3021 -0.0288
#> -1.3161  0.2869 -0.7526  0.5355 -2.0783 -0.0996  0.1805 -1.3422 -0.2363  0.3470
#>  0.2876  0.3034 -0.2071  1.3265  0.5467 -0.9401 -0.6987 -0.2485  0.9747 -0.1976
#> -0.5900  1.1580  0.1996  0.4847 -1.0806 -0.0746 -0.0150  0.8929  0.4002  1.4377
#> -0.3611  0.1876 -0.4419 -1.5198  0.7202  1.0043 -0.3526 -0.8084 -0.8269  0.1212
#> -0.1371  0.4209  1.5901 -0.5928  1.1040  0.7934 -0.7963 -0.9653 -0.5557 -0.4700
#> -0.6318 -0.6720 -0.7113 -0.3740 -0.0522  0.2547 -1.0597 -0.7227  0.8852  0.1042
#> -0.0936  0.3424 -2.2357  0.5007 -0.2706  0.6685  0.0710  0.8367 -0.1000  0.2790
#>  0.3595  0.7304 -0.0548 -0.2028  0.4169 -0.7338 -0.3273 -0.0748 -0.3742  0.5335
#> 
#> Columns 21 to 30 0.3051 -0.5776  0.3256 -1.0190 -0.8833  1.7170 -1.0721 -0.5316 -1.4416 -0.6382
#>  0.3567 -0.0605  0.2809  1.0525 -1.3851  1.5754 -0.3100  1.4972  0.8010  0.6454
#> -1.2398  0.8913 -0.9779  1.0351 -1.4845  0.3461  0.5316 -1.1750  1.3388  0.3512
#> -1.6010  0.1772 -0.0325 -0.4559 -1.2073  0.0402 -1.3660  0.0380 -0.4814 -0.1313
#> -0.3426  0.7802  0.1280  0.6954  0.3707 -0.1758  0.3467 -0.3211 -0.1145 -0.5309
#>  0.0214  0.9714 -0.4331  2.1567  0.4394 -1.8465  1.2712 -1.4755  2.0639  0.3457
#>  0.1071  0.5178 -0.2385 -0.3187 -0.1217 -2.7440  1.5348  1.9482  1.1144  0.2804
#> -1.7394  1.2019  2.1047  0.0169  1.4554 -0.3917 -1.1985  0.5066 -0.3266  0.5469
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.0620
#>  0.6474
#> -0.2892
#>  1.1359
#>  1.3135
#>  0.1303
#>  1.3508
#>  0.1464
#>  0.0046
#>  0.9512
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.6474
#>  0.9512
#>  0.1303
#>  1.1359
#> -0.2892
#>  0.1464
#>  1.3508
#>  1.3135
#>  0.0620
#>  0.0046
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>