Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 1.1575  0.1111  1.8203 -0.6470 -0.8785 -0.4781 -0.1657  0.0182  0.6202 -0.4200
#> -0.5592 -0.7434 -0.5484  0.5466 -0.8324 -1.6446 -1.7667 -1.1305 -0.5777  0.4923
#>  0.3783 -0.0978  0.6096  0.2966  0.1128 -0.3571 -0.8631 -1.3816 -0.3737 -0.4465
#> -0.5979  0.4051 -1.1207  0.0689 -0.6801  1.2828  0.4348 -0.9480  0.3426  1.8156
#> -1.1636 -0.1564 -0.3526  0.1988 -1.1216  1.4364 -0.0549 -0.8559  0.0041  1.3436
#>  0.6835 -0.2074  0.5822  0.5354  1.1631 -1.2900  1.9968 -0.3602  0.0040  0.2644
#>  1.1855  0.0403 -1.4421  0.2717 -0.4165  0.2274  0.6243  0.5402  1.3707 -0.0769
#> -1.6488  0.8296  0.6132  0.4840 -1.4590  1.3786 -0.1781  0.3191 -0.6463 -0.7886
#>  0.3345  0.2646 -1.2302 -1.0884 -0.5907 -0.5499 -0.5615  0.3004 -0.5564  1.3791
#>  0.1546  0.6823 -0.0255  0.1437  0.3877 -0.1533  0.2561  0.6378 -0.8220  0.3995
#> 
#> Columns 11 to 20-0.7168  0.8363  0.2664 -0.3377 -0.2104 -0.4992 -0.0038 -0.0820 -0.2589 -0.3199
#>  2.4989  1.0256  1.3281  1.9660 -0.5315 -0.9485  1.9578 -0.8056  0.9521 -1.5046
#>  1.5499  0.0023  0.2858  1.7070  0.2429 -1.9741 -0.3719  0.3496 -1.0988 -0.6946
#> -0.5536  1.0864  0.7965  0.9263  0.5965  0.7251 -0.8391  0.2182 -0.7094  0.0935
#> -0.4905  1.6269  0.0457  2.1435  0.2960 -1.8516 -0.0943 -0.7899 -0.4691  0.8350
#>  1.0027 -0.0201  2.5889  1.4263  0.3536  0.0389 -1.1949  0.5468  0.6390  0.2286
#>  2.3953  0.7786 -1.8569  1.0702 -0.8011  1.9772 -1.1836 -0.8117  0.7706  0.0598
#> -1.7157 -0.8820 -0.2985 -0.7312 -1.0246  0.5344  0.7913  0.1666 -0.6745  1.1089
#>  0.3398  0.8799 -0.7290 -0.6038  1.1631 -0.3416  0.7315 -0.2854 -0.1607 -1.6364
#> -0.2691 -0.6841 -0.0869  0.3843 -0.3094 -0.4615  0.0997 -0.7157 -0.5833  0.1615
#> 
#> Columns 21 to 30-0.2720 -0.9863 -0.0985  0.0344  0.4070 -0.6575  2.2445  0.4863  0.2405  0.3913
#> -2.0836  0.0039 -0.4081 -0.3579 -0.1121  1.0974  1.6510 -0.0283  0.7891 -0.4643
#>  0.2399 -0.1064  1.0496  1.7162  0.4603  0.2397 -0.0784  0.3525 -0.8250  1.3141
#>  0.0515  0.0492 -1.6805  1.8101 -0.7048  0.9588 -0.5305 -0.2920  0.6982  1.6005
#>  0.2797 -1.7566 -0.2068 -1.2779  1.4311 -0.7483 -0.4684 -0.9601 -1.7259  1.5927
#> -1.3635 -1.5657  1.8584  0.8067  0.3834 -1.7981  1.0396 -1.2757 -1.3667 -0.7028
#>  1.7959 -1.1987 -1.1169  1.4883  0.9328  0.0658  0.5510  1.5268 -0.6078 -0.9026
#>  0.4682 -0.3698  0.6902  0.4461 -1.4926 -1.2321  1.3502 -0.4074 -1.5182 -1.1440
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.9769
#> -0.6819
#>  0.5129
#>  0.7862
#> -0.1732
#> -3.0986
#> -0.1333
#>  0.2260
#> -0.9839
#>  1.8287
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.5129
#> -0.6819
#>  1.8287
#> -0.1732
#> -0.9839
#> -3.0986
#> -0.1333
#> -0.9769
#>  0.7862
#>  0.2260
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>