Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 4.6553e-01 -1.5382e+00 -9.3089e-01 -1.0419e+00 -5.7085e-01 -3.0875e-01
#> -2.0248e+00  1.2412e+00 -1.3400e+00  2.8821e-01 -1.7968e-01 -1.9655e+00
#> -1.6373e+00  2.3650e+00 -1.3382e+00  1.3181e+00  1.8385e+00  5.9777e-01
#>  1.0072e+00  1.2267e+00 -6.8400e-01  8.7999e-01 -5.7025e-01  9.9041e-01
#>  1.3398e-01  1.8952e-01  3.1864e-02  5.6832e-01  8.8222e-01  6.5834e-01
#>  6.5179e-01  1.1162e+00 -8.8746e-01  3.9444e-01 -3.6023e-01  3.2097e-02
#> -2.7469e-01 -5.5779e-01  2.0271e-01  2.1686e+00 -2.3274e-01  9.9160e-01
#> -1.6114e+00 -3.4596e-01  1.4576e+00  1.1521e+00  7.5860e-01  2.1817e-02
#>  4.2340e-01  6.8507e-01  1.3806e+00  1.8427e+00 -1.6059e+00 -5.6978e-01
#> -6.7263e-01  2.9214e-01 -2.9361e-01 -7.4899e-01 -1.1188e+00  7.7401e-01
#> 
#> Columns 7 to 12-9.1320e-01 -2.3204e-01  1.2071e-01  1.9210e+00 -1.1923e+00  8.1648e-01
#>  2.6711e-01  2.8265e-01  1.7358e-01  4.1548e-01  1.7924e+00 -6.2452e-01
#>  8.9237e-01  1.2789e+00  2.6883e-01  4.0700e-01  1.2592e+00 -5.9237e-02
#> -6.8443e-01 -7.1875e-01 -4.1089e-01 -1.9616e+00  8.7137e-01 -1.2035e-01
#>  5.6542e-01 -7.9212e-01  7.8205e-01 -2.9163e-01  5.2225e-01  1.3447e+00
#>  1.0112e+00  1.9173e-01  1.2627e+00 -1.2990e+00  3.2190e-01 -3.2221e-01
#>  6.1233e-01  7.8692e-01 -9.5316e-02  3.9573e-01 -1.0905e+00  1.2457e-01
#> -6.5814e-01  2.0602e+00 -6.2451e-01 -7.3634e-02 -4.6624e-01 -7.2441e-01
#>  8.6225e-01 -4.9333e-01 -9.3667e-02  5.2299e-01  6.3985e-01 -3.8627e-01
#> -8.9374e-01  1.3698e-01 -2.1218e+00 -5.6842e-01 -6.9931e-01 -8.4365e-01
#> 
#> Columns 13 to 18-8.6252e-01  8.6374e-01  4.2332e-01 -1.2792e+00 -2.5980e-01  1.0100e+00
#> -6.1531e-01  4.8322e-01 -9.8655e-02  4.2422e-01  4.3240e-01 -1.3561e+00
#> -2.2428e+00  2.7655e+00 -6.6314e-02 -1.1566e+00  1.0175e+00  5.2958e-01
#>  2.3665e-02 -4.4631e-01 -5.9178e-01  1.3420e+00  5.9327e-01  9.7472e-02
#> -1.3890e+00 -1.1882e-02  2.4958e-01  1.2057e+00  6.9608e-01  4.2228e-02
#> -1.9020e-01 -7.8842e-02 -3.5478e-01 -7.7562e-01 -1.6760e+00 -9.3330e-02
#>  7.0193e-02  2.0363e-02 -1.6819e+00 -6.2997e-01  6.7333e-01 -1.4845e+00
#>  5.6803e-01 -7.0697e-01  1.5558e+00  5.9874e-01 -7.5909e-01 -1.6262e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.8484
#> -1.2321
#> -0.2846
#>  0.5583
#>  1.6198
#> -0.6709
#> -0.4613
#> -1.0001
#> -1.2502
#> -0.5724
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.2321
#> -0.2846
#> -0.5724
#> -0.6709
#> -1.2502
#> -0.4613
#> -1.0001
#>  0.8484
#>  0.5583
#>  1.6198
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>