Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-5.3456e-01 -1.4450e+00  8.9625e-01 -6.4364e-01  4.1345e-01 -2.5466e+00
#>  1.2728e+00 -1.3557e+00  2.4186e-02 -3.5864e-01 -3.4722e-01 -1.9357e+00
#>  1.0956e+00  1.6940e+00 -1.8388e-01 -5.6123e-01 -2.9419e-01  1.0483e+00
#> -1.0504e-01 -1.6140e+00  9.8666e-03 -1.0819e+00  1.3904e+00  2.0845e+00
#> -1.6174e-01 -2.9821e-01 -1.6010e+00  6.5642e-01 -8.2711e-01 -7.9869e-01
#> -4.1830e-01  4.5687e-01 -2.0209e-01 -1.5782e-01 -1.3419e+00 -1.4617e+00
#> -2.3499e-01  1.7621e+00 -1.4841e+00  5.8741e-01 -7.8399e-02 -2.3920e+00
#>  1.2652e+00 -5.2196e-01 -4.1250e-02 -1.1929e+00  4.2433e-01 -5.6939e-02
#>  1.7025e-01 -3.7597e-01  4.4852e-02  9.0993e-01  9.3743e-01  2.0446e+00
#> -4.4514e-02 -5.0278e-01  1.0789e+00  9.3975e-01 -7.0535e-01  7.6528e-01
#> 
#> Columns 7 to 12 1.2289e+00  1.2225e+00 -3.4811e-01 -2.2664e-01  1.2246e+00 -6.6083e-01
#> -2.2149e-01 -5.7439e-02  1.5405e+00 -4.8655e-01 -1.1728e+00  1.3320e+00
#>  1.4319e+00  1.4033e+00  1.4131e+00 -2.0447e-01  3.5387e-01  5.8541e-01
#> -3.4461e-01 -3.8464e-01 -8.0078e-02  3.8043e-01 -1.7441e-01 -4.3305e-02
#> -5.6694e-01 -3.7999e-01  4.4576e-01  9.3205e-01 -1.8594e-01 -2.3655e+00
#>  1.7122e-01  6.2313e-01  6.9086e-01  1.1921e+00  1.0605e+00  3.6915e-01
#>  1.0405e-01  9.0741e-01  6.5684e-01  4.3301e-01  2.6395e-01  1.2055e+00
#>  1.2880e+00 -3.2948e-01 -5.2643e-01 -7.7573e-01 -1.1666e+00 -9.7800e-01
#> -1.2089e+00  4.4039e-02  1.7079e+00  7.8570e-02 -5.3557e-01  1.6653e+00
#> -8.1268e-02 -1.5023e-01  1.1793e+00 -1.2404e-01  9.7404e-01  5.6291e-02
#> 
#> Columns 13 to 18 1.8962e+00  1.4346e+00  4.0056e-01  1.0567e+00  1.6413e+00  6.0146e-01
#> -2.2183e+00 -1.3634e+00 -2.0713e+00  4.4565e-01  3.1564e+00  3.1576e-01
#> -1.7583e-03 -8.9280e-01 -3.2169e-01 -8.3599e-02 -5.3020e-01  1.5926e-01
#> -5.2813e-01 -1.2391e-01  2.2479e-01  4.0400e-02 -7.4170e-01  1.4415e-01
#> -1.1331e-01  2.4895e-02 -2.9161e-01  1.6625e-01 -1.0451e-01 -2.8420e-01
#>  2.8684e-01  1.5793e+00 -1.5251e+00  5.8418e-01  5.6914e-01  5.0805e-01
#>  1.9003e+00 -1.0077e-01  5.4945e-02  6.9742e-01 -5.3316e-01  7.1520e-01
#>  2.5842e-01 -1.3646e+00 -6.7970e-01  2.4777e-01 -4.8084e-01  7.3287e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.3100
#> -0.6402
#> -0.6732
#> -0.6135
#> -0.1605
#> -0.6582
#> -0.4830
#>  1.2746
#> -0.9794
#> -1.4430
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.4830
#>  1.2746
#> -1.4430
#> -0.9794
#> -0.6732
#> -0.6402
#> -0.1605
#>  0.3100
#> -0.6582
#> -0.6135
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>