Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 1.0658 -1.2625 -0.4612  0.4538 -0.3531 -0.1959 -0.5922  1.0012  0.8720 -0.0578
#>  1.7048 -0.2544 -0.8494 -0.0598 -0.0321  0.0658  0.8653 -0.9474  1.3799  0.7770
#>  0.1153 -0.8589 -0.2221 -0.9530  0.1678  0.2011  0.0935  0.1928  0.7434 -0.0679
#> -0.3768  0.1735 -0.6343 -1.5370  0.0141  0.2310 -0.4702  1.2761  1.7573 -0.2411
#> -0.5045 -1.5536 -0.6181 -2.6573  0.6267  0.2510  1.3708  0.4377  1.2308 -1.1392
#>  1.6089  0.4027  1.0552 -0.1621 -1.6507 -0.3868 -0.0170  1.9605 -0.2142 -0.7693
#> -0.2910 -0.1738 -0.0335  0.0995 -2.3003  0.6118  0.3390 -0.4546  1.4593  0.1018
#> -0.3676  0.0138 -1.3697 -1.1444 -0.7095  0.5922  0.7534 -0.5540  0.7736 -0.2150
#>  0.2499  0.9617 -1.6267  0.2929 -1.1716 -0.9758  0.4222 -0.6319 -1.1609 -0.2574
#>  0.3276 -1.1107 -0.6140 -0.4257 -1.2838 -0.0116  1.2571  1.0902  0.4563 -0.4153
#> 
#> Columns 11 to 20-1.0155  1.5864 -0.8908 -1.6074 -0.7014  1.5423 -0.0751 -0.2681 -0.1065 -0.2062
#> -1.1322  0.4953  0.0577  0.4400  0.2413  0.9255  0.7985  0.0538 -1.3651 -2.7865
#>  1.0508 -0.8278  1.0673 -0.5859  0.6494  0.2712  0.1374  0.0937  0.8841  0.6344
#>  0.1010  1.0297  1.5642  0.9797  1.8710  0.2353 -0.2460 -0.7925 -0.1463  1.2131
#>  0.9535  0.0467 -1.0672  0.0464  0.2331  0.4126  0.6049 -0.2899  0.7816 -1.0205
#> -0.1011 -0.2483  1.0855 -0.2331 -0.0070  1.0708  0.0898 -1.1319  1.2278 -0.6891
#> -0.3922 -0.1875 -0.1145 -0.9011 -0.7627  1.2205 -0.3247  0.0951  0.9708  2.6170
#>  0.8865  0.6234 -1.3998 -0.7465 -0.5849  0.9682  1.2671  0.2402 -1.8944 -1.1290
#>  1.5808 -0.3071 -0.4171 -0.3478 -0.2756  0.2449 -0.2518 -0.6085  0.3399 -0.4044
#> -0.6940 -0.3352 -1.3707  0.3670 -0.5892  1.0240  0.9647 -0.0890 -0.5217 -0.1685
#> 
#> Columns 21 to 30-0.0186 -0.7457 -0.5479 -0.4717 -0.5320 -1.7221 -0.8034  0.1263 -0.2181  1.8101
#> -0.3502 -1.1755 -1.0534 -0.1402  0.8765  0.5138  0.4505 -0.1373  0.3056 -0.2962
#>  0.6986 -0.2280  0.1683 -0.4382 -0.2810 -1.1506  0.5960 -1.2865  0.9578  0.8227
#>  0.3893 -0.9086 -0.5103 -2.5906  0.7800 -0.2149 -1.6502  1.3815 -0.8581  0.5650
#> -0.8223  1.6706  0.3105 -2.2454 -1.7518 -1.4938  0.6919 -0.2966 -0.0530 -0.3589
#> -0.2983  1.0019 -1.6714  2.0625  0.5086 -0.7058 -0.3018  0.7571  0.4655  0.5607
#>  0.4279  0.4021 -0.1528  1.5835  0.5538 -0.3193 -0.8994  0.5747  0.4412 -1.1365
#>  1.0714  0.8497  0.4411  0.4590  0.5340  0.9242 -0.4615 -0.6006 -1.8453 -0.1928
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.5904
#> -0.4473
#>  0.5470
#> -0.5748
#>  0.5363
#>  0.1074
#>  0.3704
#>  0.0560
#>  1.4567
#> -0.3522
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.5748
#> -0.4473
#>  0.5363
#>  0.5470
#>  0.0560
#>  1.4567
#>  0.1074
#> -0.3522
#>  0.5904
#>  0.3704
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>