Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-7.9328e-01 -6.7437e-01 -4.8899e-01 -1.3583e+00  5.8313e-01  9.5022e-01
#>  1.2511e+00  4.3879e-01  6.5563e-01  4.5149e-03  5.9344e-01 -1.8823e+00
#>  1.6446e+00  2.1104e-02  9.9641e-01 -5.7555e-01 -2.7886e-01 -3.4155e-01
#>  1.2359e+00 -3.7456e-01 -7.1543e-01 -3.8370e-01  4.1308e-01 -1.2045e-01
#>  6.4095e-01 -1.0645e+00  1.7796e-01 -8.2596e-02  6.3521e-01 -1.3827e-01
#> -3.6074e-01 -7.9944e-01  1.2562e+00 -1.4274e-01 -1.3790e+00 -2.2785e+00
#>  7.4262e-01 -1.5697e+00 -2.3843e+00 -3.1425e-01  1.3835e+00 -2.2036e-01
#> -2.0340e-01 -7.6771e-01  3.7592e-01 -1.4050e+00 -4.3206e-01  1.5135e+00
#> -3.1351e-01 -8.9963e-01 -3.9650e-01  5.5626e-01  1.1194e+00 -3.4709e-01
#>  8.4185e-01  7.2285e-01  1.6173e-01  1.0286e-01 -1.7768e+00  3.8191e-01
#> 
#> Columns 7 to 12 1.8183e+00 -3.3634e-01 -1.1986e+00  2.7241e-01 -2.1282e-01 -8.9937e-01
#> -5.8510e-01 -9.8495e-01  1.7230e-01  8.5961e-01  1.7151e-01  1.1116e+00
#> -6.6192e-02 -1.0797e+00 -2.6986e-01 -9.3898e-01  2.1787e-01 -1.3035e+00
#>  4.0183e-01 -4.3330e-02  3.8852e-01  5.2936e-01 -4.4302e-01  2.5777e-01
#>  4.1617e-01  2.8438e-02 -1.2512e+00  1.3299e-01 -1.7544e-02 -4.7161e-01
#>  1.9984e-01 -1.1405e+00 -4.0076e-01 -2.7256e-01  5.4302e-01  2.4157e-01
#> -1.2131e+00 -1.0485e+00 -5.5903e-01  9.3402e-02 -1.3068e+00 -9.1295e-01
#>  4.0124e-01 -2.6031e-01 -1.1225e-01 -8.2090e-01  1.0345e+00  6.3361e-01
#> -3.2743e-01  1.1541e+00  1.2572e+00 -6.1162e-01  2.2665e-01  4.7835e-01
#>  1.6087e+00 -1.4117e+00 -1.7266e-02  3.6249e-01 -3.0607e-01 -5.1741e-02
#> 
#> Columns 13 to 18-1.4497e-01  8.2482e-01 -3.2696e-02 -1.1447e+00 -5.0698e-01 -3.8126e-01
#>  2.0055e+00 -2.7628e-01  3.2032e-01 -5.7423e-01 -6.9346e-01 -1.8053e-01
#>  2.5957e-01 -6.8758e-01  1.3921e+00  1.7112e-01  2.1416e+00  5.4556e-01
#>  1.5683e+00  7.8939e-01  2.4397e-01 -3.7335e-01 -1.0683e+00  3.2597e-01
#>  8.9697e-03  6.8443e-01  6.9859e-01 -4.3440e-01  4.0350e-01  1.4173e+00
#> -1.8401e+00 -9.5334e-02  1.5399e+00 -8.1816e-01 -1.4195e-01 -5.7639e-01
#>  2.3331e+00  3.2426e-01 -1.7371e+00  1.0170e+00 -1.3542e-01 -7.6868e-01
#>  9.1705e-01  9.8739e-01  9.1467e-01 -2.7484e+00  1.3522e+00  9.2406e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -1.5910
#>  0.8019
#> -0.0906
#> -1.1711
#>  0.6971
#> -1.5956
#>  0.2402
#> -0.0505
#> -0.2771
#>  0.5354
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.2402
#>  0.8019
#>  0.6971
#> -0.0505
#>  0.5354
#> -0.0906
#> -1.1711
#> -0.2771
#> -1.5956
#> -1.5910
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>