Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.4002 -0.0647 -0.8068  0.9849 -3.0993 -0.4292  0.2120 -0.7747 -0.3960 -1.1066
#> -0.9371 -1.5823 -0.9417 -1.5766  0.7337 -0.7633  0.2065  1.1898  0.6417 -1.1497
#> -0.1949 -0.0733  0.6719  0.2378 -0.1837  1.1590 -0.2091  0.8175  0.9110  1.5960
#>  0.0909 -0.2116  0.4304  0.0197 -2.7341 -0.5140 -0.2120 -0.5607 -0.9792 -0.0297
#>  0.8439 -0.9090  1.1350 -0.9158 -0.2079 -0.4669  0.7164  1.4545 -0.6323 -1.0695
#>  0.9712 -0.8560 -1.3389 -1.1367 -0.6287  1.2157 -0.3019 -0.6913 -1.6140 -2.3259
#> -0.3798 -1.1089 -0.1844 -0.5136  0.8011 -1.0699 -1.7227  0.0535 -0.1064 -0.2429
#>  1.0097  0.0849  1.0825 -0.6387 -0.1055 -0.3013 -1.7622 -0.3648  1.4341  0.2728
#>  0.4951  0.6987 -0.6488 -0.0179  0.0810 -1.1783  1.0502  1.1884 -0.7533 -2.6479
#> -0.2878  1.0131  1.2365  0.7619 -0.8866  1.5030  0.3221  0.1394 -0.9713  0.4660
#> 
#> Columns 11 to 20-0.5116 -0.8336  1.2970 -0.0911 -1.6951  0.2985  1.8865  0.6766 -0.7974 -0.9104
#> -1.6513 -1.4010 -1.5773  0.1478  0.1599 -0.2981  0.4644  0.3381 -0.1823  1.7833
#>  0.0055  0.4292 -0.3223 -0.3147  0.3293  0.9874 -0.0550 -0.3608 -0.0961  1.6745
#> -2.0833 -0.0131 -0.9856 -0.3092  0.0205  1.0638 -1.3145  0.4439 -0.1833 -0.9410
#> -0.4289 -0.7953  0.8711 -0.3653 -0.8828 -0.1752  0.8437 -0.1488  1.0179 -0.4533
#> -0.4861  1.1404 -0.2646  0.4412  0.5798  1.2303 -1.3025 -0.6695 -0.0702  1.1124
#> -0.0954 -0.0336  0.0027 -0.1370 -1.9371 -0.4813  0.6702 -0.4568 -0.7141  0.4395
#>  1.2241 -0.4283  1.4507  0.1234 -0.1869  0.1760  0.2899 -0.5474 -0.1339  0.4981
#>  0.8029  1.0569  1.4166 -0.1594 -0.6149 -1.1697 -1.1292  0.2404  0.4719  0.3245
#>  0.0861 -0.1767  0.0782 -1.9056 -0.0806 -0.7112  0.6501 -1.8648  0.9246  0.9723
#> 
#> Columns 21 to 30-1.2331 -0.1552  0.3923  0.9896  1.2032  2.1541  0.9625  0.6280  0.3178  0.1543
#>  1.1982  0.2084  0.8311  0.0670  0.4575  1.0782 -0.5031  1.4250  0.4221 -0.3877
#> -0.3386 -0.3064  0.1165  0.6834 -0.2089 -0.8959  1.3728  2.5332 -0.9211 -1.7850
#> -0.1248 -0.1803 -0.1026 -1.0172  1.6673 -0.8505 -1.4698  0.0675 -0.0826 -0.3541
#> -0.3393  0.7934  1.1436 -0.8274  0.8408  0.4536 -0.0664  0.8773  0.2465 -0.2063
#> -0.0360  0.3324 -0.0177  0.6512  1.2719 -0.8596 -1.6461 -0.4927  0.7789  1.4634
#> -0.9175 -0.5705 -0.1104  0.6383  0.1867  0.1354 -1.2939  0.1749  0.5937  0.8258
#> -0.0846 -1.0053  0.8671 -1.4870 -1.5808 -0.7406 -0.4416 -1.2295 -0.3298 -0.5764
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.9139
#> -1.4970
#> -1.6030
#> -0.4884
#> -0.1488
#>  1.4542
#> -1.6748
#>  0.3069
#> -0.3395
#> -0.8883
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.3069
#> -0.1488
#>  1.4542
#> -0.4884
#> -1.4970
#> -1.6748
#> -0.3395
#> -0.9139
#> -0.8883
#> -1.6030
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>