Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.9336 -0.5762 -0.0508 -0.4290  0.0396 -0.6127  0.2713 -0.0821 -0.8626 -1.1937
#>  0.3174 -0.3842  0.4011  0.2833 -0.7453  0.8586 -2.0872 -0.3144 -0.5095 -1.4684
#> -0.8757 -0.3141  0.2712 -2.1118 -1.1004  0.8590  0.5203 -1.3677 -0.2021 -0.0880
#> -0.2413  0.8399 -0.4629 -0.2091  0.0069 -0.6788 -0.7748 -2.4076  0.4357  0.9499
#>  0.3697 -0.5814  0.5856 -0.4755  0.0800 -0.6243 -0.9035 -0.2878  0.6051 -0.8996
#>  0.7938  0.9327  0.4838  0.8194  0.2048  1.2127 -1.6366  0.2175  0.9694  1.1130
#>  0.7652  0.3203 -1.0049  1.0681  0.0379  1.2486 -0.2673 -1.6343  0.2604 -1.1174
#> -0.2797 -0.1426 -0.4612 -0.5143  1.2875 -1.4431 -0.0953  0.8357  0.2235  0.0314
#> -1.3270 -0.9589 -1.2754 -0.1936  0.9512 -0.7926 -1.0133 -0.6350 -1.0719 -1.0746
#> -1.4047 -2.0370 -0.3287 -0.2773 -0.1519  0.0387  0.0711 -0.4913 -1.9657 -0.2846
#> 
#> Columns 11 to 20 1.0751  1.4242  1.3664 -0.1769 -1.2044  1.0322 -0.0777 -0.2278 -1.3463 -0.8984
#>  0.7920 -0.9466  0.3753 -0.4442  1.2894 -0.3972  0.0840 -0.9810 -1.1799 -0.1373
#>  0.8665  0.8330 -0.2593  0.0118  0.6029  0.0654  0.0256 -0.9896 -1.0696 -1.1030
#> -1.2729 -2.0952  0.2933 -1.4963 -0.5021 -0.7042  0.1205  1.0639  0.1440  0.2093
#> -0.0955  0.4263  0.4048 -1.6639 -2.1594 -0.2660  0.2632  0.2560  1.1842  0.3465
#> -0.7335 -0.3839 -0.1186 -0.0380 -2.5497 -0.2609 -0.5152  1.1851  0.0506 -0.8001
#> -0.9794  0.1021 -0.1039  1.0487  0.7160  0.3712  0.7418 -0.4997  0.8348 -0.7982
#> -0.6411  0.2986 -0.3054 -0.4264 -0.3258 -0.9211 -0.3380  0.0491  1.4460 -0.6014
#> -0.4811  0.3705 -1.6708 -1.1028  0.2963 -1.0650 -1.1942  0.2159  1.3411  0.0681
#>  1.9252 -0.2139  0.3092 -0.0844 -0.0052 -0.9096 -1.1540 -1.4128 -0.8194  1.0655
#> 
#> Columns 21 to 30-1.2425  0.2250 -0.7814  0.6811 -1.6900 -0.7874  0.7274  0.7014 -0.2806  0.8284
#> -0.6384  1.1431  0.8198  1.6099 -1.3111  0.7315  0.6229  0.0473  0.2733  2.4085
#> -0.7875 -0.3221  0.6590  1.2841 -0.5072 -1.1915 -0.4166  0.3377 -0.7260  0.1224
#>  0.0856  0.5885 -0.1768  0.0737 -0.5879  0.3851 -1.5277 -0.0403 -0.6467  1.3755
#> -0.5204  1.8531  0.7082  0.2655 -0.5669  0.8327 -0.4680  0.2964 -1.0295  0.6171
#> -0.6735  0.7316 -2.0044  0.0887  1.1166  0.7241 -1.8384  0.9268  1.6957 -0.8334
#> -0.8988  0.8506  0.7726  0.6846 -1.5397 -0.0269  1.0452 -0.3663  0.1253  1.4626
#> -1.4733  1.3926  0.8259 -0.4600 -1.6176 -0.3718 -0.0085  0.0204 -1.2469  0.8480
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.3826
#>  0.2951
#> -1.6163
#> -2.1386
#>  0.4364
#> -1.8863
#>  0.6801
#>  0.6847
#>  0.0607
#> -0.5399
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.5399
#>  0.0607
#>  0.4364
#> -1.6163
#>  0.6847
#>  0.2951
#> -2.1386
#>  0.6801
#> -1.8863
#>  0.3826
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>