Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.6019 -0.0377 -0.5464  0.2298  1.0200  1.4636 -0.7640 -0.7607 -0.8097  0.0776
#> -0.2244  1.3014 -0.4772  0.3010  0.1127  0.7498  1.4194 -0.5425  1.2037  0.4011
#> -0.0712  1.3585  0.0921 -1.0372  0.4986  1.0077  1.4374 -1.2140 -0.0216  1.4968
#> -0.0527 -1.0598 -0.9142  1.0826 -0.1211 -0.0160  0.3200  1.4719  1.0929 -0.2892
#>  0.6055 -0.0105 -0.5367  1.2236  0.7922 -1.1983 -1.2986 -0.0382 -0.1577 -0.7084
#> -0.0440 -1.3852 -0.8982  0.4873  0.5034  0.1605 -0.1740 -0.1067 -0.1452  1.1362
#> -1.4695 -0.5326 -0.3675  0.7327 -0.1234 -0.1638  0.4123  1.0332 -0.3110 -0.1781
#>  0.4162  1.4840 -1.4317 -1.3856  0.9821  1.2359  0.8844 -0.7280  1.3721 -0.5288
#> -0.7544  0.7407 -1.1504  1.0650  1.5800  0.2487 -1.3911 -1.2100 -1.9909 -0.0799
#> -0.8532  0.1586  0.4409  0.7752 -0.0108 -0.5707 -0.8383 -1.5805  1.0542  0.4215
#> 
#> Columns 11 to 20-0.7463  0.0100  1.0409 -0.8198 -0.1242 -0.2620 -1.1154  1.3574 -1.4487  0.7903
#>  0.8790  0.2857  1.2454  0.3067  0.3404  1.3226 -0.3949  0.4767 -2.1461  0.9117
#>  0.5422 -0.8015 -1.9631  0.8065  0.0003  0.1724  0.5819 -0.3766 -0.6770 -0.6339
#> -0.9881 -0.5825  2.0076 -0.1992  0.5089  1.4854 -1.3620 -0.4143  1.0253  2.5483
#> -0.5625 -0.8939  0.4545 -0.1395 -0.4711 -0.2594  0.4122 -0.1689  1.4730  0.5006
#> -1.3811  1.0694 -0.4410 -1.5315  0.7207  0.5405  0.3792  0.4362  1.4228  0.8585
#> -0.1557 -1.2291 -1.0688 -0.7593 -0.5617 -1.0373 -0.5436 -0.7923 -0.3029  0.0622
#> -0.7388 -1.3520  0.3007  0.5052 -0.2257  0.5326 -1.5443  0.4807 -0.7342 -0.3725
#> -1.0044 -1.8748 -0.4208  1.4131 -0.5686  0.9790  0.4023  0.3627 -0.8821  0.2857
#>  0.7068  1.0742 -1.1751  0.3672 -0.0556  0.8434 -0.5883 -1.0278  0.5817  1.1127
#> 
#> Columns 21 to 30-0.4209  1.3000 -0.0298 -0.4185  0.2327  1.4946 -0.7352 -1.1477  0.2643  0.4369
#>  0.0324 -1.3455  0.8792  0.1770  0.3614 -0.4318 -1.0823  0.3711 -1.0783 -1.4539
#> -0.5599 -0.3715  1.4327 -0.0116 -1.9137  0.1719  1.3373  1.5436  0.5363  0.9986
#> -1.4565 -0.2781  1.1770  0.0418 -0.8481  1.1749 -1.0513  0.7670  2.2985 -0.4835
#> -0.1976 -0.9548  0.8414  0.3688 -1.4354 -0.4858 -0.0590 -0.6920  1.0639  1.5640
#>  0.1364  0.4553 -0.8078 -0.9715  1.6628 -0.7778 -1.1092  0.9577  1.4546  0.0881
#>  0.6589  0.5921  0.0414 -0.6814  0.4268 -1.3155 -0.4947 -0.6647  0.2087 -0.3854
#>  0.6502 -0.8689 -0.5246 -1.2641  1.9915  0.7128 -0.2905 -0.0086  0.2938  0.9118
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.3031
#> -0.5453
#> -0.5365
#>  0.7263
#> -1.0509
#>  0.4077
#>  0.5983
#>  0.0812
#> -0.2549
#>  0.5213
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.5365
#>  0.3031
#>  0.5213
#> -1.0509
#> -0.2549
#>  0.5983
#>  0.4077
#>  0.7263
#>  0.0812
#> -0.5453
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>