Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.0207 -1.4970 -0.7902 -1.1999 -0.5511 -0.1128  0.6856 -0.1114 -0.8509 -0.5749
#>  1.0137  0.0114  0.5948 -0.9981  0.6469  1.4253  0.2731  1.5035 -0.7166  0.0433
#>  1.3996  0.2442  0.6808 -0.2517  1.3645  0.7165  1.1553 -1.0532  0.3703  0.5031
#>  0.6078  0.9070 -1.1991 -0.0873 -1.1705  0.2415  0.1115  1.6611  0.0843 -1.1565
#>  0.1616 -0.8762  1.0794 -0.6256 -0.5747 -0.3441 -0.9429 -0.0678 -0.5389 -0.6175
#>  0.1988 -0.1518 -0.4893  1.1478 -1.1167 -0.1408  0.8638  0.3424 -1.6832  0.0144
#>  0.3060 -2.0068 -0.6673 -1.6011 -0.5775 -0.4356 -0.9008  1.8656 -1.2029  1.5723
#> -0.6526  0.7542 -0.1608  0.4422 -2.2288  0.2385  0.3070  0.0338 -0.4825  0.0797
#> -0.9936  0.4829  1.8651  0.7029 -0.4321 -0.0080 -0.3270  0.2156  1.8367 -0.2403
#> -0.2450 -1.2459 -0.3257  0.5714 -2.1570 -0.2816 -0.4475  0.6120  0.6086 -0.6530
#> 
#> Columns 11 to 20-0.1319  0.5418 -0.8699  1.0465  0.5086 -0.6670 -1.0443 -0.7354 -0.1296 -0.0361
#> -1.0681  0.1410  1.2121  0.2284 -0.0222  1.4139 -0.9689  0.9617  1.1775  1.0906
#>  0.3614  1.3610 -0.0825 -0.0964  1.6217 -0.4633  0.4835 -0.6078  0.8146  0.4777
#> -0.2655  2.5777  0.8605  0.7121  0.7743 -1.5092 -2.1801  0.5046 -0.2247 -0.1157
#> -2.0807 -0.5112  0.0050  0.8691  1.4709  0.8610  0.9152  1.1102 -0.6382 -0.6203
#>  0.1835 -0.0202  0.3334  0.0346  0.3669 -0.0581  0.4908  0.5550 -0.2896  0.1424
#>  0.0292 -0.0059  0.2457 -0.0288 -1.1239  1.4896 -0.4881 -0.0037  0.3882  0.2994
#>  0.6593 -1.6460  0.3909  1.9617  1.1342 -0.1641  0.7088 -0.5004  1.1340 -0.5359
#> -0.9010  0.3742  0.2717 -0.4497 -0.6805 -0.1546  0.3877 -0.9824  0.8624  1.1182
#> -0.0562 -1.4688  0.3398  0.0721  0.2413 -0.8119  0.8714  0.3600  0.2658  1.4739
#> 
#> Columns 21 to 30-0.1024  0.4754 -0.7748 -1.0832  1.2931 -2.9065  0.2376 -0.0067 -0.4710  0.7936
#>  0.0130  0.6717  0.7918  2.0259  0.3264  0.2889 -0.4766 -1.0802 -1.3635  0.6753
#>  0.0225 -0.6655  1.0915  0.3121 -0.3403  0.6419 -0.1568  0.1600  0.3258 -0.6855
#> -0.8148  0.8160 -0.8004  0.0566  1.3005  0.5603  0.3222 -0.4541 -2.2469  1.0197
#> -0.6571 -0.4462  2.7068  0.0541 -0.6946 -1.3036  0.3749 -1.2376  0.1567 -0.9319
#> -0.7060 -0.4694  2.8087  2.1751  0.7635  0.5820  0.5918  0.4951  0.6652  0.7774
#>  0.8183 -0.7880 -0.0640 -1.5025 -0.1107  0.5087  0.0821 -0.2253  0.5944  0.2676
#>  1.0278 -0.8210 -0.0038  0.0720  0.3526 -3.5415 -0.8571  1.8328  0.2513 -1.7658
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.4679
#> -0.4461
#> -2.2575
#>  0.8181
#>  1.2766
#> -1.4176
#>  0.2632
#> -0.3171
#>  0.5289
#> -1.0294
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.5289
#> -1.0294
#> -2.2575
#> -0.4461
#>  1.2766
#> -1.4176
#>  0.2632
#>  0.4679
#>  0.8181
#> -0.3171
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>