Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 2.1105e-01 -2.5707e-01  5.0293e-01 -2.6365e-01  2.9616e-01  4.3874e-01
#> -5.3747e-01 -5.0843e-01  2.0182e-01  8.5945e-01 -7.9758e-01 -1.6854e-01
#> -1.8622e+00 -7.2628e-01 -1.2179e-01 -5.3673e-01  8.8290e-01 -7.7352e-02
#>  6.0087e-01  7.7442e-01 -1.8468e+00  4.8284e-01  1.4391e+00  2.0366e-01
#> -1.0608e+00 -4.1275e-01 -9.6645e-01 -5.1798e-01  2.5813e-01  1.7352e-01
#> -8.6037e-01  1.4365e-01  6.6950e-01 -4.4121e-01  4.2209e-01 -3.8243e-01
#>  4.5700e-01  7.9541e-01  4.9467e-01  1.3577e+00 -5.6978e-01 -1.1119e+00
#>  7.8966e-01  4.9365e-01  1.0959e+00  6.6656e-01  2.4713e-01  2.4156e-01
#>  6.8629e-01  4.3494e-01  1.5368e+00  4.5424e-01 -3.3821e-01 -6.9955e-01
#> -2.9487e-01 -2.7045e-01 -9.3513e-01 -3.1766e-01  7.1092e-01 -8.8386e-01
#> 
#> Columns 7 to 12-5.8458e-01 -7.6261e-01 -9.8281e-01  1.0952e-01 -3.8169e-01  4.4187e-01
#>  1.5732e+00 -4.8694e-01  1.8215e-01  2.4406e-01  4.9622e-01  4.1927e-01
#> -2.1962e-01 -1.5063e-01 -4.3045e-01  6.1290e-01  1.3646e+00 -7.8468e-02
#> -2.7856e-01 -1.4861e+00  5.4135e-01  2.5380e-01 -2.1084e+00 -6.7824e-01
#> -7.2378e-01  6.6451e-01 -5.6135e-02 -2.5516e-02 -1.6625e+00 -1.2545e+00
#>  4.8056e-01  1.1037e+00  1.6371e+00  1.1139e-01 -6.4466e-01 -1.5184e+00
#> -1.2516e+00 -1.3917e-01  1.8302e-01 -7.0514e-01 -2.0332e+00  5.3306e-01
#> -5.2930e-01 -3.2553e-01  7.5119e-01 -8.0412e-01  1.3013e+00  1.3164e+00
#>  7.5627e-01  5.4333e-03 -1.1749e+00  1.0025e+00 -1.3122e-01  7.3929e-01
#> -3.6906e-01  7.2472e-01  6.5900e-01  1.6670e-01 -1.2228e-02  6.5552e-01
#> 
#> Columns 13 to 18 3.4535e-01 -1.2002e+00 -8.3307e-01 -1.6820e+00 -5.6943e-01 -1.2224e+00
#>  1.0747e-01  4.3129e-01  1.0875e+00  4.7297e-01 -5.5352e-01 -6.9736e-01
#> -5.6236e-01  5.3038e-01 -4.8145e-01  9.4094e-01  2.5152e+00 -8.0532e-01
#>  1.2296e+00 -5.9918e-01  9.1384e-01  5.5982e-02  1.0325e+00  9.0756e-01
#> -5.5983e-01  9.8870e-01  2.4292e-01 -2.4190e-01  4.6381e-01  6.5734e-01
#> -5.8572e-01 -5.4169e-01  4.0119e-01  5.8703e-01 -4.3276e-01  9.2243e-01
#> -3.0966e-01  9.1974e-02  1.8338e-01  1.0977e+00  9.2757e-01  1.5192e+00
#> -4.0285e-01 -1.2765e+00  3.3926e-01 -1.7810e-02 -7.1996e-01 -1.3532e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.9905
#> -0.3795
#>  1.2743
#> -0.5082
#> -1.0673
#> -0.4616
#>  0.5942
#>  1.0820
#> -0.3193
#>  0.2713
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.3193
#> -0.3795
#>  1.0820
#>  0.5942
#> -0.4616
#>  1.2743
#> -0.5082
#>  0.2713
#> -0.9905
#> -1.0673
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>