Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.5784  0.1256  2.7669 -0.8787 -1.0363  0.2277 -0.0842 -1.2338  0.3572  1.6562
#> -0.0255 -0.6929  0.3624 -0.8134 -2.1113  0.1022 -1.8654 -1.0516  0.5067  0.1946
#>  0.2924 -0.6669 -0.2758 -0.1485  0.1285  0.1774 -0.7725 -1.2894 -0.1815 -0.1514
#>  0.0342  1.0602  1.0330  1.3827  1.7186  0.6163  0.9165 -0.7540  1.1619  1.6370
#>  0.2601  0.7321  0.6147 -0.8213 -0.7620 -3.8616  0.6805 -0.4993  0.1862 -0.2783
#>  0.2022 -0.6524 -0.1935  0.5775 -0.3151 -0.4271 -0.3142  0.2354  1.5357 -0.3387
#>  0.3288  0.7004  1.1206  0.1119 -0.9774  1.3771  0.7428  1.5765  0.6089  0.1764
#> -0.1584  0.7303 -1.2253  0.0707 -0.5118 -0.0411  1.8664 -1.8185  0.1419 -0.8622
#>  0.8311  0.8525 -0.8358 -0.4720  0.3427  0.0698  0.8829 -0.2850 -0.2289  1.2680
#>  1.4102 -0.3677  0.9065 -0.3229 -0.3239 -1.4188 -1.6949  0.4804  0.3961  1.2874
#> 
#> Columns 11 to 20 0.4993  0.8466  0.1906  0.7285 -0.5045  1.2703 -0.2362 -0.4563  0.5131 -0.8054
#>  1.9250 -0.3826 -0.5222  0.4288 -0.0421  0.2385  0.4663  0.4890  0.9246 -0.6614
#> -0.1677 -0.1219  0.3807  2.2093 -0.7449  1.2687 -1.4790 -1.2011  0.9731 -0.0240
#>  0.8507  0.6611  0.5638  0.8936 -0.9718 -0.0339 -0.1719  0.4446  0.9031 -0.4484
#>  1.7293  2.2188  0.4373  0.7118 -0.4951 -0.6846  0.2403  0.2399 -0.5375  1.0656
#>  0.0095 -0.3087 -0.4823 -0.0616 -0.2378 -1.3948 -0.6557 -0.4627  0.5015  1.8359
#> -0.1801  1.5821  0.1438  0.4711  0.6781  0.6423  0.5321  0.1058 -0.0438 -0.8080
#>  0.8752  0.4676  0.8024  1.7217  0.9499 -1.0777 -0.5257 -0.4778  0.0743  0.7788
#> -0.7845  0.9767 -0.1071  0.0288 -1.2368 -0.7145 -0.2089  0.4135 -0.3296 -0.5364
#> -0.5917  0.1031  1.2228  0.4499 -0.4252 -1.9250 -1.2984  0.5265  1.3286  0.5279
#> 
#> Columns 21 to 30-0.8226 -0.3894 -1.4672  1.1626  0.7071 -0.2007 -0.5744  0.9055 -0.8332 -0.6059
#>  0.2436  1.1839  0.8900 -2.0819 -0.6666  1.3076 -0.1313 -1.7200  0.6052  0.8579
#> -0.3147 -0.1000 -0.0704 -0.6582 -0.1198 -1.0539  0.1576 -1.7467  0.4153 -0.2959
#>  2.1302 -0.8073 -0.2897  0.6657  0.3046  0.6057 -0.3742 -1.2705 -1.0509  0.3213
#>  0.1802 -0.2364  2.8658 -0.0580 -2.0249  0.7019 -0.7383 -0.6310  1.1105 -0.6888
#>  0.8356  1.1366 -0.8227 -0.7840  0.8244 -0.3580  0.2367  0.7684 -0.8618  0.3347
#> -0.0495  0.6047  0.6770 -0.1279 -0.1797 -0.5429 -0.2166  0.9642  1.5412 -0.8769
#>  0.3855 -2.2277  1.0243 -0.4808 -1.5531 -0.3979  1.9667  0.1652 -0.4485 -0.8676
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.2975
#> -1.2352
#> -0.7823
#>  1.2107
#> -0.8555
#>  0.6695
#>  0.4547
#>  0.8248
#>  0.0665
#> -0.9901
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.2352
#>  0.2975
#>  0.8248
#>  0.4547
#> -0.9901
#> -0.7823
#>  0.0665
#> -0.8555
#>  1.2107
#>  0.6695
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>