Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.8984 -0.0519  0.5438  0.0136  1.2412 -0.1531 -1.0209 -0.1653  0.2467  0.0396
#>  1.0295  0.3453  0.4485  0.4768 -0.5541  1.2317  2.2698 -0.8272  0.2117 -0.6338
#> -0.4344  0.5822 -0.5548 -0.5387  0.5549 -0.6933  0.0099 -0.6326  0.0828  0.0328
#>  0.1569 -0.0574 -0.2461 -0.6213  0.2639  1.4946  0.6556  0.7999  0.8782  0.4731
#>  0.8639  0.1609  0.1606 -0.7057  1.1209 -0.0067 -0.5322 -0.6065  0.5669  0.3807
#> -1.8977  0.0790 -2.1381  0.1841 -0.5107 -2.4149 -0.1003 -0.7090  0.6851  0.1775
#> -0.1107  0.9294 -0.7672 -0.4803 -0.1325  1.1327 -0.4175 -0.1723  0.7057  0.4801
#> -0.5729  0.4628 -0.0704 -1.7269  0.5750 -2.2681 -0.2691 -0.1589 -0.8251 -0.9699
#>  1.1397 -0.4591  0.5256  1.2126  1.0523  0.4265  3.0015  0.2574 -0.7142  0.8203
#> -0.2940 -0.5284 -0.7825  0.6620 -1.0078  0.2310 -1.2281  0.7845 -1.9814  1.0215
#> 
#> Columns 11 to 20 0.3226  0.0225 -0.7209  0.9639 -1.3865 -1.5149  0.0103 -0.1244 -0.7510 -0.0137
#>  0.0874 -1.1418 -0.4568 -0.5523 -0.4370 -1.6581 -0.2465 -0.3704 -1.1358  0.4864
#> -0.8838  0.0498  0.5567  0.2715 -0.5829  1.3657  0.3016  0.0373  0.0613 -2.3655
#> -1.1645  0.4089  0.6694  0.4951  1.6362  0.4101 -0.5449  0.6670 -2.0817  0.9210
#> -0.4449 -1.5282 -0.1203 -0.0570 -0.6974  0.8025  2.7745  1.0724  0.2179 -1.2572
#> -0.6437 -0.0958 -1.2028  0.0158 -0.6817  2.6280 -0.6981  0.4583 -0.3455 -1.0832
#> -0.0256 -0.3991  1.6293  1.2198  0.5040  0.2076  0.9826 -0.3935  0.3038  0.0349
#> -1.0276  0.6786  0.9907  1.9587  0.5853 -1.4830 -0.0205 -0.1561  1.2248 -0.6970
#>  0.0112 -0.4066 -0.5502  0.3457  0.4670 -0.4770  0.3044 -0.8642  1.1055  1.1117
#>  0.3454  0.5501 -1.0415  0.1650 -0.2911  0.6314 -0.7404 -0.8344  1.0534 -0.6749
#> 
#> Columns 21 to 30 0.8957  0.5638  0.2480  0.7148 -1.1898  1.0906 -0.7063 -0.2669  1.0438  0.4590
#>  1.2894 -1.1173  0.8435 -0.1172 -1.7420 -1.3666 -0.3445 -0.7693  0.3464  0.5145
#> -1.2438  0.4954 -1.2721  1.7564  0.6719  0.6354  0.5331 -0.4133  0.8392  0.7158
#>  0.4020 -0.4855 -0.6413 -1.3033  0.2599 -0.1287 -1.0314 -1.0610 -1.1316  0.0464
#>  1.7335 -1.3391  0.3136  0.1367  0.2586 -0.0726 -0.3817 -0.6940  1.1917  0.2321
#>  0.2303 -0.4869  0.3606 -1.2109  0.6665  1.1447  0.3646  0.1170  0.4694  1.3784
#> -0.0215  1.8020 -1.2999  0.2856  0.6157  0.4418  0.1779 -1.3285  0.6829 -0.1824
#>  0.4235  1.1438 -1.2808  2.5665 -1.0026 -0.0008 -0.3854 -1.1176 -1.0879 -0.7756
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -1.8915
#> -0.3989
#>  0.2760
#> -0.8259
#>  1.4815
#>  0.9324
#> -0.0843
#>  1.2259
#> -0.3141
#> -0.8339
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.8339
#>  1.4815
#>  1.2259
#>  0.9324
#> -0.3141
#> -0.3989
#>  0.2760
#> -1.8915
#> -0.8259
#> -0.0843
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>