Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-1.2892e+00 -3.7406e-01  1.0721e+00 -1.3223e+00  8.9182e-01 -9.3185e-01
#> -3.1140e-01 -6.0374e-01 -1.3507e-01 -1.5700e-01 -5.0236e-01 -4.7386e-02
#> -3.8674e-01 -1.2485e+00  7.0795e-01  3.8612e-01  2.5393e+00 -1.3086e+00
#>  6.7202e-01 -1.9742e+00 -1.6796e+00 -1.5665e+00  1.9840e+00  5.2015e-01
#> -1.0374e+00 -8.6956e-01 -6.3166e-01  1.5846e+00  8.1050e-01 -9.6061e-01
#>  8.4170e-01  6.0600e-01 -1.2998e+00  1.6430e-01 -2.4806e-02 -1.7081e+00
#>  2.8695e-01 -9.0040e-01  8.6067e-01  1.1225e+00  1.2038e+00  5.6412e-01
#> -1.5864e-01 -1.5183e-01 -9.8403e-01  1.7026e-01 -6.5560e-01  1.7114e+00
#>  1.1345e+00  9.0294e-01 -7.8902e-01  1.2716e+00  1.1344e+00  6.3781e-01
#>  8.5180e-01  5.7978e-01  1.1254e+00 -6.5050e-01 -3.5187e-01  4.3612e-02
#> 
#> Columns 7 to 12 3.4228e-01 -7.8487e-01  6.1008e-02 -1.5287e+00  5.8079e-01 -6.9210e-02
#> -3.8952e-01  6.8729e-01 -1.7699e+00 -4.6040e-01 -6.9762e-01  1.1207e+00
#> -6.9425e-01  5.7049e-01  1.7133e+00  7.5440e-01 -4.8412e-01  1.7551e-01
#>  1.5990e+00 -1.4452e+00  7.3146e-01  6.0250e-01  6.8315e-01 -6.7243e-01
#> -6.8818e-01  2.5067e-01  5.5305e-01  6.6830e-01 -3.4310e-01  1.3696e-01
#>  2.2898e-01  7.2239e-02 -6.4519e-01  1.3310e+00  5.5591e-01 -8.8467e-01
#>  1.5759e+00  1.6688e+00 -1.6279e+00  6.4913e-02 -7.8527e-01  2.9929e-01
#>  1.3342e-01 -1.4369e+00  1.6998e+00  1.7123e+00  2.8178e-01 -8.1133e-01
#> -3.1778e-01  5.9827e-01  1.1015e-01  1.5857e+00  9.6009e-01  1.1583e+00
#>  2.4343e-01 -9.2633e-01 -6.5510e-01 -1.4755e+00 -1.6901e-01 -4.0132e-01
#> 
#> Columns 13 to 18-1.5746e+00 -3.0973e-01  9.9613e-01 -3.5758e-01  1.8311e-01  1.3366e+00
#>  7.9706e-01 -5.8312e-01 -5.4384e-01 -6.7144e-01 -6.0226e-01 -1.1756e+00
#>  1.0798e-01 -4.3339e-01 -1.8277e+00  9.6369e-01 -2.1251e-01  5.1895e-01
#> -1.7549e+00 -4.4111e-01 -2.8595e-01 -6.4313e-01 -1.0615e+00 -8.2749e-01
#> -7.6301e-01  2.5666e-01  4.4612e-01  1.2685e+00 -1.3121e+00  8.5487e-01
#> -1.0243e+00  1.0353e+00  1.4352e+00  1.0547e+00 -4.6886e-01 -7.1846e-01
#>  1.5331e-01  4.5968e-01 -1.2521e-01  1.8781e-01 -1.0754e+00  1.3306e+00
#> -1.4756e+00 -5.2123e-01 -5.3256e-01  4.9838e-01 -7.6477e-01  5.9786e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -1.6724
#>  0.9660
#>  0.9848
#> -0.3082
#> -1.9499
#> -1.2669
#> -0.3570
#>  0.3572
#> -0.7248
#> -0.1922
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.3570
#> -0.1922
#> -0.3082
#>  0.9660
#>  0.9848
#> -1.2669
#> -1.9499
#>  0.3572
#> -0.7248
#> -1.6724
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>