Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 1.3171e-01 -6.2645e-02 -3.1940e-01  1.2441e+00  9.3343e-01 -9.5996e-02
#>  9.0365e-01 -1.3481e+00  1.1036e-02 -1.1757e-01  4.4134e-01 -1.3683e+00
#> -5.2395e-01  2.5274e-02 -5.1434e-01 -5.8587e-02 -3.5227e-01 -3.3220e-01
#>  4.9192e-01  9.0113e-01 -1.1884e+00  7.1272e-01  5.3986e-01  5.0472e-01
#>  1.8478e-01 -1.1551e+00  6.1390e-01 -7.8220e-02  1.3289e-03  3.9134e-01
#> -1.1879e+00 -1.1390e+00 -5.3588e-01  1.1257e+00 -6.4308e-01  2.9605e-01
#>  5.5000e-02  2.8793e+00 -3.9118e-01  3.0320e-01  1.2089e+00 -6.1567e-01
#>  8.1651e-01  1.2308e+00 -4.5822e-02  6.9543e-01  8.9530e-01  1.3605e+00
#> -8.7408e-01  7.1683e-01  2.5113e-01 -2.1254e-01 -2.4471e-01 -1.0405e-01
#> -6.3985e-01 -2.0271e-01 -7.9037e-01  1.7744e+00 -1.0713e+00 -7.0036e-01
#> 
#> Columns 7 to 12-6.0431e-01  5.6093e-01 -4.2688e-01 -8.9444e-01 -1.0645e-01 -2.0527e+00
#>  2.0586e+00  8.3779e-01 -2.9080e-01  1.6026e+00 -7.6857e-01 -2.7069e-01
#>  4.1839e-01 -1.0201e-01  1.2734e+00 -1.9596e-01 -4.7316e-01 -7.5437e-01
#>  5.4476e-01  2.0667e-02 -3.9706e-01  1.5557e+00 -7.7675e-01  2.7717e-01
#>  2.0538e-01  2.4643e+00 -1.1382e+00  2.7682e-01  2.3236e+00 -1.5191e+00
#>  4.5738e-01  5.2883e-01  4.9306e-01  1.7365e+00 -5.2194e-01  2.5967e-01
#> -5.5450e-01 -4.7771e-01  1.4190e-01 -1.1102e+00  4.1796e-01 -1.7872e+00
#>  1.0945e+00  2.2565e-01 -4.3070e-01  1.0024e+00  2.3808e+00 -1.5614e+00
#> -2.5042e-01  1.0823e+00  7.2643e-01 -2.0478e+00  2.3171e+00 -8.3517e-01
#>  1.9777e-01 -1.5792e-02 -3.9300e-01 -2.8591e-01  2.7570e+00  3.6250e-01
#> 
#> Columns 13 to 18-3.3362e-05  4.8682e-01  9.6906e-01  1.0674e+00  7.1532e-01  8.8217e-01
#>  2.4847e-01  4.0733e-02  1.4986e-03 -9.6304e-01  2.3566e-01  2.8517e-01
#>  1.9782e-03 -6.4348e-01 -1.0328e+00  4.1730e-01 -1.0174e+00 -3.3413e-01
#> -1.1701e+00 -1.0720e+00  1.6867e+00 -1.3460e+00 -4.5331e-01 -2.3605e-01
#>  4.9774e-01  7.8906e-01  3.4680e-01 -9.7843e-01  6.1851e-01  1.5141e+00
#> -1.9873e-02  2.0196e-01 -1.6510e-02 -3.3498e-01  8.1869e-01  7.9583e-01
#>  1.0052e+00 -2.4889e+00 -7.2288e-01 -6.5719e-02 -3.0481e-01 -2.9556e-01
#> -9.7800e-01 -8.8058e-01 -8.5495e-02  1.0457e+00  2.1942e-01  1.1591e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.0198
#> -0.1179
#> -1.5478
#>  2.7005
#> -0.3450
#> -0.3156
#>  0.6492
#>  1.5052
#> -2.0257
#> -1.0343
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  1.5052
#>  0.6492
#> -0.3156
#> -2.0257
#> -1.0343
#> -0.1179
#>  0.0198
#> -0.3450
#> -1.5478
#>  2.7005
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>