Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.4862  0.5335 -0.8930 -0.4291  0.5676  1.0684  2.1062  0.8249 -0.0454 -0.4483
#>  2.4481 -0.6106 -0.3430 -0.9247  0.5252 -0.7728 -0.0346 -1.3941 -1.5802 -0.5473
#> -0.3049 -0.9998  1.8993  0.3333  1.6576  1.1098 -1.0208 -1.4057 -0.8819 -0.4907
#> -0.7068  0.3127 -0.7202  1.0287  1.4690 -0.8730 -0.4854 -1.1737 -2.3097  2.1189
#>  0.1572 -0.8219 -1.5502  0.1801  1.4193 -1.3927 -0.5114  0.7399 -0.3769 -0.0050
#>  1.6357  0.0160  1.7546 -0.3200  0.2524 -0.3104 -0.5832  0.7543 -0.7222 -0.5805
#> -1.4704  1.4797  0.0581 -0.9328 -2.2983 -2.7040  0.1261 -1.6591 -0.3998  1.5030
#>  0.6897 -0.5375  0.4247 -0.2241 -0.9478  2.5041 -1.6991  2.2312 -1.0099  0.9687
#> -0.8045 -0.6929 -1.7951 -2.2976  1.7419  1.3875 -0.0410 -0.4612  0.0686  1.2517
#>  1.0105 -0.5842  0.5416 -0.5666  0.2973 -1.8300  0.3722  1.1527  0.1469  0.7629
#> 
#> Columns 11 to 20-0.1566  0.4766 -0.2353 -0.1444  0.1205  0.6502 -0.0626  0.1137  0.3549 -0.1642
#>  0.6168  0.0613  0.3948  1.0506  0.6168 -0.3448  1.2258  0.6564 -0.1607  1.5295
#>  0.8154  0.2485 -0.2807  0.5762  0.1709 -0.1973  1.1330 -0.5988 -0.9783  0.0454
#> -0.2332  0.0917  0.9134 -0.0378  0.3115 -1.1118  0.0402 -1.0756 -0.6789  0.5627
#> -1.0578 -0.9544 -0.0922 -1.7467  0.6555  0.6639 -0.1693  0.5077 -1.3808  0.4596
#> -0.4308  0.9553  1.2562  0.8002  2.1814  1.5209  2.0914  0.7831 -0.6554 -0.4518
#> -0.2078 -0.4774 -2.7402 -0.2524 -1.0273 -0.1029 -1.3828  0.5283  0.2148  0.5632
#>  1.2690 -0.1852  0.7550  0.7936  0.2398 -0.6197 -0.1021  0.0364  0.3510  0.5944
#>  1.2631  0.2873  0.2027 -1.1663 -0.2797 -0.8279  1.5672 -0.6143  0.4031  0.3029
#> -1.6294 -0.0279 -0.4106  0.4392 -1.3886  1.2767 -1.3705 -0.1728 -1.0858 -0.7084
#> 
#> Columns 21 to 30-0.6839 -0.7974  0.6561  0.7106  0.1771  0.3903 -0.9200  0.7772 -0.2715 -0.3793
#> -1.0761  0.3521 -0.1190 -0.5964 -2.4916 -0.4049  0.5951  0.6317  0.5790 -0.5773
#> -0.1199 -0.1867 -1.5099  1.5440  0.1155 -1.5644  0.8282  0.0063  0.3022 -0.8015
#> -0.9577 -1.6795 -0.0272  0.6646  1.0159 -1.4779  0.1126 -0.0162  1.4313 -1.1682
#> -1.7316 -0.8811 -1.0349 -1.3285 -1.1731 -1.7170 -0.6233 -0.6333 -0.6821 -0.9835
#> -1.1556 -0.2049 -0.7995 -0.1119 -0.0157 -0.1796  1.0351  0.6515  0.2191  0.1243
#>  0.7838 -1.7163 -0.0437  0.1921  0.3799  0.6260  0.8231 -0.8181  1.4456  0.0897
#> -0.9616  1.0578  2.0205 -1.2846 -0.0625 -0.1330  1.0035 -0.3131  0.9971 -0.2451
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.4739
#> -0.6974
#>  0.7513
#>  1.2651
#> -0.6937
#>  0.7194
#>  0.0814
#> -1.1922
#> -0.6758
#> -0.8017
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.8017
#>  0.4739
#>  0.0814
#> -0.6758
#> -0.6974
#> -1.1922
#>  1.2651
#>  0.7194
#>  0.7513
#> -0.6937
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>