Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 3.0748e-01  1.6858e-01  1.2128e+00  9.4315e-01 -5.7265e-01 -1.0609e+00
#>  5.0393e-01 -3.3849e-01  4.8823e-01  2.1474e-01 -1.2466e+00  6.1013e-01
#> -1.9351e-01 -8.8629e-01 -7.2614e-03 -1.0559e+00 -9.4638e-01 -1.4307e+00
#>  1.8761e+00  1.0782e+00 -5.4737e-01 -6.8296e-01 -8.0563e-01 -1.5617e+00
#> -6.4056e-01  6.2893e-01 -6.5916e-01 -1.4269e+00 -1.7959e+00 -6.3633e-01
#>  4.4965e-01 -7.5968e-01  6.8420e-01  8.0451e-02  4.3534e-02 -2.0798e-01
#>  2.1637e-01 -1.0343e+00  3.9201e-02 -1.3796e+00  2.0072e-02 -4.3476e-01
#> -2.0475e+00 -4.0169e-01 -1.1596e+00 -3.8746e-02 -1.5975e-02 -5.4923e-01
#>  4.0961e-01 -1.9766e+00 -1.4672e+00  6.6014e-01  7.4526e-01  3.6126e-01
#>  3.5571e-01 -1.4275e+00  5.7174e-01 -1.8876e-01 -2.5416e-01  4.8208e-01
#> 
#> Columns 7 to 12 8.5008e-01 -6.9802e-01 -1.5131e+00 -8.6559e-01  5.9199e-01 -7.6003e-01
#> -8.8612e-01 -6.4573e-01  1.2966e-01  7.0596e-02 -8.7576e-01  8.3553e-01
#> -1.8115e+00  4.1528e-01  5.0383e-01  8.4235e-01 -1.3107e+00  1.0996e+00
#>  1.4609e+00  2.8475e-01 -1.9604e-01  2.9058e-01  2.7315e+00  5.0494e-01
#>  6.6458e-01 -4.8503e-02  3.3604e-01  6.9890e-01 -1.3226e-02 -7.9190e-01
#>  7.2485e-01 -2.6705e-02 -7.9356e-02  2.8763e-01  4.3002e-01  5.2453e-01
#>  8.3123e-01 -1.9018e+00  6.1823e-01 -1.5202e+00  5.2050e-01  4.6747e-01
#>  1.7242e+00  2.2735e-02 -6.7380e-01  6.9343e-01 -8.3388e-01 -1.1823e-01
#>  4.7675e-01 -4.1585e-01  1.0665e-01  1.3598e-02 -2.2972e-01  1.6412e-01
#> -4.3965e-01 -1.0611e+00 -5.8381e-01 -4.0831e-01  3.1632e-01 -4.9778e-01
#> 
#> Columns 13 to 18 3.9744e-01 -7.5651e-01 -3.5248e-01 -2.9740e-01 -9.4752e-01  4.6324e-02
#> -1.7270e+00 -6.2087e-01 -5.5038e-01  7.8012e-02 -2.2468e-01 -4.9784e-01
#>  8.1708e-01  3.1757e-02 -4.6769e-01  5.5856e-01  1.2576e+00  3.9381e-01
#>  1.3226e+00  3.5121e-01  5.5929e-01  1.8354e-01 -1.1558e+00 -3.1910e-01
#> -3.0338e-01  3.7663e-02  7.2175e-01 -1.1136e+00  8.4754e-01 -1.7296e+00
#>  1.2045e+00  4.3666e-01  5.3179e-01 -3.2013e-01 -8.2787e-01  1.5745e-01
#>  3.5553e-01 -2.9251e-01 -1.9605e-01 -1.5411e+00 -4.2136e-02 -1.4936e+00
#> -4.7588e-04 -1.1857e-01 -1.5532e-01 -2.2122e+00  1.4031e-01 -6.2131e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -1.0663
#> -0.9743
#> -1.4829
#>  0.1270
#>  0.0065
#> -0.6996
#> -2.5692
#>  0.4456
#> -0.1930
#> -0.0268
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.6996
#> -0.9743
#>  0.1270
#> -0.1930
#> -2.5692
#> -0.0268
#> -1.0663
#>  0.0065
#> -1.4829
#>  0.4456
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>