Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-1.6129 -0.6514 -0.0005 -0.1299 -0.0186  0.7856 -0.2318 -0.5054 -1.1800 -0.4573
#> -0.3187  0.2511 -0.5413  0.5586 -0.3886  1.2698 -0.5388 -0.4903 -0.3484  0.1465
#>  0.4670 -0.4047  0.5951  1.7964 -0.1125 -0.1080  0.0275 -0.1578  0.4809 -1.3088
#> -1.2264 -0.5856  0.8783  0.1327 -0.6009 -0.2881  1.3573 -1.4678  0.5514  0.7259
#> -1.3347 -0.1024  1.3148 -0.4473  0.5465 -0.5149  0.9301 -0.0461  0.3143  0.5608
#> -0.8352 -0.4503 -0.6097  0.0943  0.8621 -0.3653 -0.0237  0.9062 -1.2661  1.7011
#>  0.0790 -2.2431  1.7278  1.5395 -0.5357  0.3805  1.7119  0.1466 -0.1981  1.3738
#>  0.5916 -0.0441 -0.1281  0.3864 -0.4095 -0.8772 -0.4108  1.1746 -1.2667 -0.9288
#> -0.5432 -0.8673  0.3779  1.4750  0.9157 -0.8307  1.0736 -0.2050 -1.2962  0.7474
#> -0.1829  0.7280 -1.1095  0.3540  0.9854 -0.2013 -0.1124  1.8542 -0.3396 -1.2286
#> 
#> Columns 11 to 20-0.4164 -1.8517  0.8905 -0.7747 -0.6324  0.4760 -0.7748 -0.0742 -1.0611  0.7593
#>  1.3576  0.3285  0.2676 -0.0533  0.2062 -0.0335  0.3198  0.3276 -1.2097  0.0647
#> -0.2135 -0.0988  0.3074 -0.4857 -1.5481  0.5156 -0.0364 -0.6499 -1.0595 -1.4608
#>  0.0545 -0.1533  1.0694 -1.9981 -0.8471 -0.7479 -0.4441  0.3173  0.9581 -0.1928
#> -0.1570  0.1221 -0.0325  0.5952 -0.7952  0.9146 -0.6144  0.1231 -0.3203 -1.2061
#> -0.0764  1.8272  0.3365 -0.1222 -0.2703  0.5525  1.1011 -1.0144  0.5019  0.7357
#> -0.5516  0.5019  0.3448  0.1197 -0.4418  1.5774  0.5755  0.0448  0.7243 -1.5652
#>  0.1351 -0.4297 -0.2023 -1.3988 -1.3668 -0.4454 -0.5770  0.3981  0.2843  0.3258
#>  0.2701  0.4466  0.5089 -1.1313  0.4318  0.9925  0.6326 -1.3562 -0.3284  0.3340
#>  0.2723  0.8889 -1.0425 -0.6844 -0.2525 -0.4499  0.3906 -0.5498 -0.2378 -0.8349
#> 
#> Columns 21 to 30 0.4978  0.0260  0.1258 -0.8327 -0.7895  0.1126 -0.2001  0.5339  1.1938  0.4961
#> -0.1211 -0.9407  1.1080  1.0563  0.6874 -0.8214  0.4860 -0.4867 -0.2094 -1.9037
#>  0.0040  0.6689 -1.3853  0.1822  0.7930 -0.8359  0.6396 -0.0760  1.2617  0.2426
#>  0.2137  0.6507  0.2203 -2.1912 -0.4301 -0.9777  0.4276 -1.0610  0.7440 -1.2844
#>  0.4721  1.3801  0.1906  0.0381 -1.8842 -0.2035  1.1486 -0.4319  1.2018 -1.0576
#> -0.1008  1.8099  0.1133 -0.3625 -0.8228  0.3376 -1.2784  0.4270 -1.8851 -0.7659
#> -0.1893  0.3863  0.0251 -0.0711  0.2325 -0.3685 -1.2638  0.1694  1.3024 -0.4186
#>  0.6907  1.8215 -0.5999  0.0687 -0.2703 -1.2546 -0.7732  1.1821  0.6891  0.7568
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.8530
#> -0.3463
#> -2.1121
#> -0.0545
#>  0.7580
#> -0.9011
#> -0.2182
#>  0.6930
#> -1.0155
#>  1.3686
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.7580
#> -0.9011
#> -0.0545
#> -0.2182
#>  0.6930
#>  1.3686
#> -1.0155
#>  0.8530
#> -0.3463
#> -2.1121
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>