Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.0004 -0.0581 -0.2522 -1.1770  1.1752  0.3230  1.9455 -0.4825  0.3313  0.0377
#> -0.4630 -0.6369 -0.9770  0.4741 -0.1994 -1.0416 -0.7298  2.2623  0.5925 -0.4710
#> -0.5418 -0.7545  1.0482  0.0115 -0.4686 -1.2242 -0.0140  0.2020  0.9118  0.9418
#> -0.3399 -1.6543  0.2101  0.8330  1.0630 -0.1271  0.4573 -0.2291 -0.5472 -1.4826
#> -1.3652 -0.5669  0.1556 -0.3002  0.2660 -0.1710 -0.3806  0.1013  0.3214  1.2227
#>  0.0437  0.7181  0.7429 -0.2970  0.5015 -0.6572  1.2906 -0.3960  0.1885  0.5447
#>  0.0664 -0.1790 -0.4453 -0.7032 -0.1673  1.7421  1.5582  1.4835  0.4789  0.9693
#>  1.3524 -1.3462 -0.4048 -0.9331 -0.7311  2.8241  1.0077  0.3978 -0.8472 -0.2863
#>  0.2998 -0.5028 -2.6757 -0.7908  0.3351 -0.1683  0.1969  1.8166  0.8486  0.7744
#>  1.5838  0.3304  1.4461 -0.9935  1.3840 -1.7817  0.3004 -0.8359 -1.1951 -1.0117
#> 
#> Columns 11 to 20-0.2292  0.5403 -1.2814  0.3675  0.6973 -0.2334  0.3517 -0.6564  1.0686 -0.8603
#>  0.5977 -1.5267  0.7682 -0.8353  0.1808 -0.6257  0.3589  1.5547  0.2499  0.0471
#>  0.3927  0.3281  1.4350 -1.9260  1.1660 -1.4453 -0.1628  0.2940  0.0224 -0.0234
#> -0.9575 -0.0255  0.6107 -0.1874  0.9317 -1.1163  0.9639  0.3211  0.8382 -0.4205
#> -0.9301  1.1426  0.3450 -0.3844  0.4165  1.0528  1.0641 -1.5109  0.9558  0.5572
#> -0.3870 -0.0404  0.5184 -0.5838  0.2354  1.0801  1.1920  0.3770  0.6183  1.2887
#>  0.3368  1.0177  0.7442 -0.0973  0.5040 -0.8950 -0.4387  0.4935  0.4507  0.4755
#>  2.0163 -0.2772  0.5313  2.1813  0.5181 -0.4688  0.9238  0.4341  0.0892 -0.8280
#> -0.3414 -1.4300  0.1428 -0.0400 -0.2629  1.1402 -0.9371 -1.4509 -1.1357 -0.0609
#> -2.3888 -1.3117 -2.3247  0.9426  0.5715  0.7960 -0.2429 -0.9162 -0.9827 -0.3814
#> 
#> Columns 21 to 30-1.6154 -0.0109 -1.2916 -0.3370 -0.5325 -0.0944 -1.1923 -0.6115  0.5266 -0.7132
#> -1.1426  0.7583 -0.7961 -1.6647  0.8479 -1.0770 -0.7589  0.5570  0.4364  0.9587
#> -0.3297 -1.5225  0.6944 -0.6960 -0.6437  0.9605 -0.5566 -0.7059  0.3311  1.1928
#> -0.9072 -1.2403  0.5332 -0.4815  1.6264 -0.6378  0.7928  0.5024 -0.6220 -0.1843
#>  0.0800 -1.1234 -0.7241 -0.5179 -0.9237  0.7728  1.0668 -0.4085  1.5933 -0.1047
#> -0.0085  0.7901  0.4171  1.2202  0.9613 -0.6494 -0.1147 -0.8545 -0.5763  0.0497
#> -0.1280  1.2097  1.1551  1.3843 -0.6525 -0.7979  1.9651 -0.0244  0.2072  2.3973
#> -0.1478 -0.8050 -1.1382 -0.0669  0.1266  0.5871  0.5584 -1.5507 -0.0574 -0.1546
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.3808
#> -0.3507
#>  0.0755
#> -2.5977
#> -0.8459
#> -2.1229
#>  1.8828
#> -0.3561
#> -0.9508
#>  1.4956
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -2.1229
#>  1.3808
#> -0.9508
#> -0.8459
#>  0.0755
#> -2.5977
#> -0.3561
#>  1.8828
#> -0.3507
#>  1.4956
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>