Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-1.2294 -0.5132 -0.1371 -0.5007 -1.6178 -0.6066 -0.4749 -2.0377  0.0457 -1.9919
#>  0.9851  0.3957 -2.1806 -0.7635 -0.4563 -0.9382  0.8301 -0.3194  0.1303  0.3376
#> -1.8105  0.5248  0.9909  1.8606 -0.0098  1.3756  0.0726 -1.5724  0.0816 -0.0728
#> -0.3745  0.1360  1.5250  1.5023  0.2220 -1.7721 -0.7426 -0.1306 -0.5083  0.2385
#>  1.0519  0.1950 -1.3550  0.3178  0.3270 -0.9548  1.1170 -0.9675  0.4789  0.9390
#> -1.1070 -0.8683 -0.0608  1.0128  0.8061 -0.1579  0.3756 -0.5550 -0.6828 -0.1765
#> -0.4448  0.4053  0.1443 -1.1685 -2.5440 -0.6744  2.1845 -1.1200  1.4200 -0.9996
#> -1.6178  0.8943 -0.2837 -1.1797 -2.8909 -1.1078 -0.5841 -0.8902  1.4176 -0.3722
#>  0.1417  0.0193 -0.1609  0.2882 -0.3423 -0.5114  0.1588 -1.8464  0.5999  0.2071
#>  0.0632  0.2612  0.4949  1.3183  0.0334  0.4211 -0.5556 -0.6531 -0.6299 -1.5986
#> 
#> Columns 11 to 20 1.0866  2.6490 -0.9250 -0.2238 -0.6456  0.1198 -0.0954 -0.7459  0.4091 -0.9561
#> -0.8790  0.6274  0.9348 -0.6659 -0.7148  0.9423 -0.4910 -0.1694  0.3337 -0.1521
#>  1.4098 -0.5138 -1.5744  0.7796  1.1792 -0.2674 -0.6903 -0.6442 -0.4772  1.8854
#> -0.3097  1.3827  1.0782  0.2193  0.7456 -0.0219  0.0919 -2.1108  0.1109 -0.9817
#>  0.0505 -1.2766 -0.4844 -0.3800 -0.1656  0.1451  0.7822  0.2662  0.6572  0.2816
#>  0.3741  0.5895 -0.1735 -1.6264 -1.3444 -0.6997 -1.8576  0.2514 -0.4879  0.6693
#> -0.9242 -1.0283 -0.4673  1.3811 -0.7193  0.1192 -0.3947  1.6920  0.0373  1.1298
#>  0.7921  0.6742 -0.2010 -0.0076  0.3448 -0.3641  0.1511  0.6595  0.1363  0.6146
#> -1.1040  1.6336  0.5738  0.0757 -0.6194  0.7564  0.6033 -0.2902 -0.3328 -0.3274
#> -0.3517  0.3003  1.0147  0.1556 -1.1982  0.1472 -0.7526  0.2092  0.5126 -0.4501
#> 
#> Columns 21 to 30 0.4783  0.0439  0.6656 -0.4111  1.0481 -0.6630  0.3260  0.8199  1.3125  0.4379
#> -1.2345 -0.2002  0.5447  2.7696  0.3167  0.8990 -2.2004  0.3226  0.5013 -1.0433
#>  0.3547  0.1691  0.6389  1.1319 -1.6172 -0.5203  0.1772  0.4372 -0.9012  1.6967
#> -0.7742  1.0542 -1.1327  0.9020 -0.6915 -0.4743 -0.2693  0.6364 -0.8937  0.4896
#>  0.9427 -0.7919  1.1850  1.3255 -0.8734 -0.3704  0.2721  1.4249 -0.4594 -0.0593
#>  2.1346 -1.1454  0.0780  0.6560  0.5711 -0.4271 -1.0858  0.7412 -0.3693 -1.3167
#>  0.7069 -0.9942 -0.7669  0.1804 -0.5955 -1.0223 -0.3781 -0.1050  0.3174 -1.2200
#>  0.2235 -1.4173  0.2072 -0.6505  0.7002 -0.1939 -1.8817  0.4809  0.3920 -1.0098
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.1213
#> -0.9386
#>  0.7098
#>  0.0554
#> -0.6884
#>  0.7028
#> -1.2193
#> -2.1738
#> -0.4478
#> -1.5632
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.4478
#>  0.7028
#> -2.1738
#>  0.7098
#>  0.0554
#> -0.6884
#> -1.2193
#> -1.5632
#> -0.9386
#>  1.1213
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>