Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.0742  0.3700 -0.1372 -0.6220  0.0750  1.0885 -1.2682  0.8924 -0.1993 -1.1789
#>  2.2188  1.2308 -1.1576 -0.1035 -0.3987 -0.3057 -0.1582 -0.4874  0.8714 -0.0081
#>  0.3735 -1.0068 -0.6934 -0.0789  0.0992 -1.4655 -0.7927 -0.0079  1.7898  0.1435
#> -0.0981 -0.0961 -0.4252  0.6895  0.6213 -1.3515  0.4495  0.6057  1.8272  0.6800
#>  0.1634 -0.9266  0.3159 -0.2352 -0.8439 -0.6814 -0.8272 -1.7601  0.5803 -1.3248
#> -0.0420  0.4949  0.8942  1.3191 -0.7448 -0.6523 -0.0729  0.4960 -0.4752  0.5784
#> -0.7067  1.4046 -1.9320  0.0205  0.2507 -2.2144 -0.2472 -0.2498  0.4818  1.7600
#>  0.2105 -0.9131 -0.3275  0.2050  0.4337  0.8932  0.0603  0.6820 -0.9313  0.7340
#>  0.5530  1.1101  1.1118 -0.7895 -1.3060 -0.6802  0.5816  0.9675  0.0744 -0.6636
#> -0.8611  1.2745  0.2255  0.9105  0.1224 -2.2285  1.0099 -1.3841 -0.6476  0.3910
#> 
#> Columns 11 to 20-0.2509 -0.2309 -1.1556  0.1325 -0.0572 -0.3381  0.8938  0.3739 -0.2229  0.6914
#>  1.8343 -0.6911  0.9334  0.9031  0.1438 -0.2316 -0.6885  0.1173  0.3405 -1.4315
#> -0.6330  0.4699 -1.0868 -1.4603 -0.3643 -0.1146  0.1069  1.5978  0.9974 -2.2503
#> -2.1416 -0.6306  1.0711  0.6050 -1.2523  0.4461  1.4732 -0.0656 -1.2244 -1.0879
#> -1.2493 -0.8725  2.4980 -1.0458 -0.7152  0.0491 -1.6703  0.3092 -2.9063 -0.8616
#> -0.3506  0.2901  1.2358  0.7944 -0.4259  1.2000 -0.8581  0.0131  1.1369 -0.5624
#>  0.8400  1.6105  1.6205  2.5378 -0.6199 -1.2931 -0.5856 -0.4835 -0.1230 -0.5077
#>  0.2242 -1.3489  0.0682  0.9408 -0.1039 -0.0105  0.6485  0.1984 -1.1310  1.4885
#> -0.1276  0.7037 -1.0488 -1.4227  0.4258 -1.5675  0.8330  0.6503  0.9547 -0.0265
#>  0.4823  0.6676 -0.6516 -0.4937  0.9106 -0.0311  1.5447 -0.4456  2.5214  1.4131
#> 
#> Columns 21 to 30 0.2158 -0.8391 -2.0863 -0.0043  0.4311 -1.3589 -0.6297 -0.8821 -0.3257  0.2927
#>  0.1703  1.5758  0.6161 -0.3832 -1.5546 -0.6683 -1.7890 -0.6580  0.4275  1.0133
#> -1.4220 -0.4198 -0.7035  0.1331 -0.1088  0.6613 -1.3407  0.4900 -0.0684  1.6255
#> -1.8585 -0.2960 -0.0858 -0.7550  0.6370 -1.6701  0.1480 -0.9206 -0.4682 -1.0017
#>  1.3392 -0.1663 -0.0941  0.7230 -1.1687 -1.6970  0.3282  1.9184 -0.2261  0.3483
#> -1.0496 -0.0553  1.0414 -0.5814 -0.7746 -0.1143 -0.2639  0.2391  0.2449 -0.3822
#> -0.8681  0.9886 -1.9965 -1.0086 -1.0212 -1.0975  1.1562 -1.0384  0.7597 -0.4456
#>  0.5359 -0.9432  1.9372 -0.0553 -2.0555 -1.0428  0.1860 -0.4698 -0.2791 -1.4313
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.9229
#> -0.1974
#> -0.3262
#> -1.3446
#> -0.9261
#> -0.2831
#> -0.5694
#> -0.3834
#>  0.7099
#>  0.6746
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.3446
#>  0.9229
#> -0.9261
#> -0.3834
#>  0.7099
#> -0.1974
#>  0.6746
#> -0.2831
#> -0.3262
#> -0.5694
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>