Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.4737 -0.9475  0.5717  0.6105 -1.8231  0.6401  0.7917 -1.0524 -2.3514  0.5166
#>  1.0265 -1.8958 -1.0094 -0.5211  0.1229  0.0664  0.1281  0.1663 -0.6431 -0.3661
#> -1.3204 -0.2253  1.0408  0.2408 -0.1659  1.7502  1.1051  0.3249  1.3002  0.9335
#> -0.0585  0.0352 -1.2057 -1.1005 -0.1192  1.6435  0.2539  0.9209 -0.0800 -0.8892
#> -0.6812 -0.0250  1.0222 -1.9183  0.3683 -1.0767 -1.1276 -0.8123 -0.0579 -1.3691
#> -1.1961 -0.7931 -0.2370 -0.3047 -0.3914  0.3205  0.8859  0.3518  0.8394 -0.3357
#>  0.0033 -0.2383 -0.6367  0.5887 -0.7627  1.3198  0.2856 -0.5249  1.6031 -0.5819
#> -0.9429 -1.2277 -0.0455 -1.2789  0.7142 -0.8262  0.6341 -1.4360 -0.6307 -0.3521
#>  1.4957  0.7473 -1.8578  0.3587 -1.0331 -0.5529 -0.1712  0.2042 -0.9766 -0.9836
#> -0.5175 -1.4968  0.9594 -0.3858  1.2911 -1.0799  1.0151  0.5995 -0.3682  0.5551
#> 
#> Columns 11 to 20-0.0814 -0.4368 -0.4486 -0.2887  0.2278 -0.9203  1.2500  2.2513 -0.4815  0.3933
#>  1.5942 -0.8265  0.3669 -0.0531 -1.3276  2.1081 -1.1236  0.5472 -0.9119  0.5426
#>  1.0736 -0.0169 -1.1838  0.8925  0.1242 -0.7911 -1.1350 -0.8797 -0.8576  0.3741
#> -0.3385  0.1746 -0.1869  0.3035  0.5458  0.4248  0.1019 -0.3436  0.9279  1.4648
#> -0.4602 -0.1323 -1.1093 -1.5608  1.2885 -0.0601 -1.8707 -0.8152 -0.4319 -1.2978
#> -0.7625  0.5522 -0.1738 -1.3562  0.3954  1.1262 -0.3996 -0.5445 -0.1108 -0.0714
#> -0.3304  0.9058 -1.6730 -0.2314  0.3940  0.1389 -0.4024  1.7844  0.1440  0.3027
#>  0.0193  1.1099  0.2797 -1.1272 -0.1230  0.4144  1.2285  0.1526 -0.3064  0.3819
#>  1.3467  0.4632  0.8238 -0.0738 -1.6315 -0.8287  0.4870  0.9209  0.9812  1.1381
#> -0.5311 -0.3513  0.0834 -0.3497  0.6810  0.2547 -0.6026 -1.1071  0.3358 -0.1280
#> 
#> Columns 21 to 30-0.4725  1.2155  0.6184 -0.0546 -1.4183 -1.2814  0.9331 -0.2697 -1.5924 -0.9243
#>  0.4913  0.6936  0.3167  0.2052  0.9222 -0.6686 -0.6525 -0.4246  0.3763  1.8761
#> -0.6670 -0.3876  0.0881  1.1242 -0.6769  0.9645 -1.4162 -0.6961  1.3841 -0.8016
#>  1.3217 -0.2373  0.3507 -2.3535 -0.2734 -1.9722  0.8201 -1.5131 -0.8892  0.9861
#>  0.7116  0.2615 -0.6090  1.2780  0.1314 -1.2318 -0.1865  1.4620  0.4487  1.2241
#>  0.3671  0.6427 -0.5893 -0.7547  0.1756 -0.2957 -0.2219  1.1712 -1.4518  0.9147
#>  0.5287 -1.1822  0.8913 -0.4258 -0.1965 -0.4794  0.4075  0.1356  1.1217  0.2529
#>  1.2652 -0.9049  0.9346  1.1094 -0.3243  0.4321 -1.0953 -1.7112 -0.6982 -0.3406
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.1839
#>  0.0828
#> -1.7097
#> -0.2802
#>  0.2763
#> -1.3682
#> -1.2963
#>  0.0225
#>  0.3237
#>  1.8729
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.2963
#>  0.0828
#>  1.8729
#>  0.3237
#> -1.7097
#> -0.2802
#> -1.3682
#> -0.1839
#>  0.0225
#>  0.2763
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>