Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10 0.2009 -0.9545 -1.3849  1.0194  0.3415  0.9052  0.5526  0.7816  1.4555  0.3289
#> -0.2874 -1.2782 -0.8443 -0.0356 -0.0370 -1.4046 -1.1495 -0.1270 -1.2831  0.7468
#>  0.0037 -0.7960 -1.1937 -0.9158 -0.4041  1.0313 -1.0585  0.2613  1.3288  0.5327
#> -0.5092  1.2128  0.4580 -0.2132 -0.2307  0.5736  0.9111  1.4031  0.4946  0.7260
#> -1.1118 -0.5526 -0.9124 -0.3758  1.5000 -0.4522 -0.1315  1.5198 -0.3757  0.4687
#>  0.1366 -0.0877  1.2373  0.0315 -0.4538  0.2690  1.0834  0.6098 -0.3818 -0.3291
#>  0.5144 -1.2050 -1.6644  3.2638 -1.2124 -0.7250  0.3797  0.4458  0.7784  0.1526
#> -0.4244  0.3324 -1.3047  0.0964 -0.1385 -0.1504  2.7950 -0.5229 -0.6830  0.0771
#>  0.4825  0.5828 -1.9790 -1.1855  1.1016 -0.9237  1.3766  0.4951 -1.3019 -0.2552
#>  1.5214 -0.8319  0.3933  0.8349  0.1119 -1.4049  1.5020 -0.8086  1.6649  0.3548
#> 
#> Columns 11 to 20 0.4462  0.8233 -1.9754  0.8314 -1.8595 -1.7454 -0.4037 -1.5225 -1.0283 -1.3386
#>  0.0518  0.0312  1.1977  0.6642 -1.0369  1.3199 -1.1696  0.8921 -0.1350 -0.9876
#>  1.1213  0.2303  1.4798  1.9180 -0.0758  1.2127 -0.2078  0.6198 -1.3082  2.5612
#> -0.2098 -0.6728 -0.0451 -0.8406 -0.5670 -0.4451 -1.2593  0.2562 -0.6416 -0.4861
#>  0.9490  0.2496 -2.2769 -1.0274  0.5511 -1.2951 -0.2119  0.1795 -1.5222 -0.0721
#>  0.8009 -0.2438  0.0716 -1.0288 -0.2056 -0.2652  1.0179 -1.1131 -0.5345 -0.5893
#> -1.2315  0.5533 -0.7633  0.5350 -1.1293  0.3977 -1.0216 -1.3518 -0.4681 -0.8050
#>  1.1656 -0.2443 -0.6976 -0.1918  0.7958 -0.5640  0.6199  0.1247  0.5789 -0.8596
#> -0.1997  0.0579  1.1454 -1.1151  1.6974  0.7738  1.5345  1.4364 -0.2408  0.5792
#>  1.7166  0.3712  1.0473 -1.6602  0.0936 -0.1164  0.1792  0.4706  0.1018 -1.1539
#> 
#> Columns 21 to 30 0.9427  0.9103 -0.0033 -1.0835  0.9578  0.4596 -0.7024  0.1062 -1.6907  0.1892
#>  1.0588 -0.2378 -1.1735 -0.7730 -1.6897 -0.0996 -1.2376 -0.8273  0.9368 -0.3387
#> -0.7991 -0.7504  0.0362  0.9570 -1.0246 -1.0452 -1.4752  0.2127 -0.3660 -0.3898
#> -0.6335  0.3536  0.5553 -0.9056  1.1837 -0.2108 -1.0520  0.2394  0.5090 -0.2711
#> -0.2681 -1.0333  0.6848  0.5663 -0.6867  0.1099 -0.0250  0.2209 -0.3605  0.1902
#>  0.2736 -1.4072 -0.3393  0.3292 -0.2936 -1.1245 -1.3581 -1.9757 -0.4382 -1.2293
#>  0.1253 -1.5296  0.4535  0.1407 -0.0399  1.4600  0.4521  0.7591  0.0292 -0.6841
#> -0.2767  0.9681 -1.6188 -1.3841 -0.9664 -1.1512  0.6265 -0.1666  0.4086  0.4228
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.7075
#>  0.7935
#> -0.6895
#>  0.0401
#>  0.4780
#>  2.0369
#> -0.2017
#> -1.9061
#>  2.2110
#>  1.1449
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.2017
#> -0.6895
#>  2.0369
#>  0.7935
#>  2.2110
#>  0.4780
#>  0.7075
#>  0.0401
#> -1.9061
#>  1.1449
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>