Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-5.5830e-01 -3.0965e-01 -2.6381e-02  1.1970e+00 -3.4009e-01  7.2677e-01
#>  1.6674e-01  1.0774e-01  2.0279e+00 -9.1778e-01  6.3297e-01 -5.1154e-01
#>  4.9640e-01  5.1064e-01 -2.6033e+00 -3.9300e-01  1.7351e-01 -6.6746e-02
#>  1.6843e+00  9.3860e-01  3.0977e-01  4.3805e-01 -1.4713e-02  3.2065e-01
#>  1.3411e+00 -3.7351e-01 -3.0024e+00 -4.3538e-01  6.5472e-01 -3.1228e-02
#> -8.6733e-01  1.7692e-01  1.2293e-01 -1.1836e+00 -7.9885e-01  6.3016e-01
#> -1.2761e+00  1.0829e+00  9.2411e-01 -1.4541e+00  3.5325e-01 -4.6782e-01
#>  4.3901e-01 -2.4225e+00 -1.6833e+00 -1.0478e+00 -2.9639e-01  1.3776e+00
#> -6.9241e-01 -6.0549e-01 -2.2570e-01 -5.4678e-01  5.3306e-01 -7.8562e-01
#> -4.4314e-01 -7.4439e-01  3.5405e-01 -2.1489e+00 -1.3899e+00  4.7245e-01
#> 
#> Columns 7 to 12 2.8269e-01 -1.2891e+00 -6.6531e-01  2.0707e-01 -1.1068e+00  1.2013e+00
#>  3.6986e-01  6.0711e-01  1.2201e+00  1.2336e+00  2.4728e-02 -4.2812e-01
#> -4.4649e-01  2.0059e-01 -8.4152e-01 -5.5776e-01 -1.1346e-01 -4.2508e-01
#>  3.3213e-01 -8.3318e-01 -7.7902e-01 -1.6778e+00 -1.0445e-01 -2.2878e+00
#> -7.3803e-01 -1.9002e-01 -4.3996e-01  1.0452e+00 -1.1564e-01  2.1468e-01
#> -3.5342e-01  7.0869e-01  2.0442e+00 -2.7595e-01 -1.1578e+00 -7.0866e-01
#>  6.0811e-01 -5.0794e-01  2.3185e-01  1.7101e-01  7.7679e-01  2.2783e-01
#> -7.6849e-01  5.3851e-01 -1.1585e+00 -2.7331e-01 -9.2852e-01  1.0918e+00
#>  2.6823e-01 -2.0164e-01 -1.4444e+00 -5.8237e-01 -2.3296e-01 -8.0822e-01
#>  3.7141e-01  6.3507e-01  1.0910e+00  1.3953e+00  3.1936e-01  1.4726e-01
#> 
#> Columns 13 to 18-8.8508e-01  1.6997e+00 -1.5091e-01  1.2378e+00  1.6599e-01  1.9557e-01
#> -4.2366e-01  2.0408e+00 -2.5977e-01 -1.7599e+00 -1.1453e+00  7.0240e-01
#>  1.2567e+00 -4.3603e-01  3.3340e-01 -3.6641e-01  4.0088e-01  1.2073e+00
#> -9.8044e-01 -5.0331e-01  4.1651e-01  3.1006e-02  8.5402e-01  1.3461e+00
#> -2.5622e-02  3.3389e-01  5.1313e-01 -1.3421e+00  7.9373e-01  2.1274e-01
#>  1.0789e-01  1.2168e+00  1.0978e+00 -1.0832e+00  6.6778e-01  5.0774e-01
#> -8.7930e-01  2.6011e+00  7.8948e-01 -9.6765e-02  7.7786e-01 -6.7945e-01
#>  3.6968e-02 -2.7445e+00  5.2125e-01 -8.1687e-01  1.3473e+00  2.0523e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.7952
#> -1.2202
#>  0.3208
#> -1.8271
#>  0.7674
#>  1.5034
#> -0.8292
#>  2.1667
#>  0.1326
#> -0.4100
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  1.5034
#> -0.4100
#> -1.8271
#>  0.1326
#>  0.3208
#> -0.8292
#> -1.2202
#>  0.7674
#>  2.1667
#>  0.7952
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>