Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6-8.5386e-01 -6.2999e-01  3.0612e-01  4.9400e-01 -9.3538e-01 -9.9691e-02
#>  1.3554e+00 -2.4328e-01 -2.4640e-01  7.5403e-01 -1.5416e-01 -3.1540e-02
#>  8.5094e-01  1.9844e+00 -6.2429e-01  1.5161e+00 -1.6170e+00 -4.6700e-01
#> -2.9155e+00  3.7638e-01 -9.2374e-02  1.2877e+00 -1.0598e-01  3.3553e-02
#>  1.1778e+00  9.3488e-01 -2.8174e-01  3.4695e-01 -1.3762e-01 -7.4141e-01
#> -7.2056e-01  1.2165e+00 -1.0830e+00 -2.3249e-01 -1.9766e+00 -1.8071e+00
#> -1.4609e+00  9.5411e-01 -1.3297e-01 -7.3411e-01 -6.0603e-02 -1.3018e-01
#>  1.5854e+00  1.0517e+00  9.7937e-01 -4.9912e-01  1.2226e+00 -6.0737e-01
#>  1.8425e-01  2.8485e-01  3.7555e-01  6.7545e-01 -1.4538e+00  1.2723e+00
#>  3.0501e-01  7.2141e-01  1.3818e+00  4.0092e-01 -1.3880e+00 -2.0183e-01
#> 
#> Columns 7 to 12-1.0084e+00 -4.3320e-01 -3.7767e-01  1.3003e+00  2.8695e-02  1.4378e+00
#>  4.0446e-01 -8.6805e-01  9.1372e-01  1.0523e-01 -1.8026e-01  7.3104e-01
#>  4.2485e-01 -3.2090e-01 -1.9197e-02  1.6314e+00  1.6252e+00 -1.7347e-01
#> -3.1192e-01  8.1360e-02  8.9430e-01 -2.1334e-01 -5.4403e-01 -1.1842e-01
#> -6.7364e-01 -3.8464e-01  3.9851e-01 -2.0339e+00  6.6887e-01 -5.4997e-02
#>  6.9269e-02 -1.6938e+00 -2.7225e-01  1.1381e+00 -1.2553e+00  1.4449e-01
#>  9.5369e-01 -1.4501e+00  8.8271e-01  8.9186e-01  2.6747e-01 -5.2237e-01
#>  9.2069e-01 -1.2520e+00 -1.4568e-01  8.9595e-01 -1.5875e+00 -8.4929e-01
#>  8.8352e-02 -6.0917e-01  1.4501e-01  9.6566e-02 -1.3398e-02 -3.5771e-02
#>  3.8588e-01  5.8070e-01 -1.1838e+00 -1.0464e+00  3.3392e-01  8.9261e-01
#> 
#> Columns 13 to 18-3.7316e-02 -5.7360e-01 -1.5084e-01 -8.5090e-01 -7.6047e-01  1.2279e+00
#>  6.8889e-01  8.7286e-01 -6.4122e-01  2.2397e-01 -2.7006e-01  2.6966e-01
#> -1.8193e+00 -5.5417e-01 -5.9637e-01  5.4936e-01 -2.6145e-01 -2.4454e-01
#> -3.4884e-02 -3.2046e-01 -1.2735e+00 -1.3523e+00  8.3328e-01  1.0688e+00
#> -3.1033e-01 -8.2826e-01 -1.0114e+00 -4.2948e-01  5.2194e-01 -5.8471e-01
#>  1.2413e+00  7.7682e-01  5.0990e-01  1.5227e+00  1.5158e+00  1.1049e+00
#> -1.4246e+00 -6.3776e-03  5.9704e-01  7.5744e-01 -1.0835e+00  4.9948e-01
#> -4.6140e-01  2.0223e+00  3.6875e-01  7.1820e-01  9.9486e-02 -4.4240e-01
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#> -0.0278
#>  2.0612
#> -0.1665
#> -0.5123
#>  0.6060
#> -0.8126
#>  1.0693
#>  0.7559
#> -0.3528
#> -0.7005
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -0.0278
#> -0.1665
#>  0.6060
#>  1.0693
#> -0.3528
#>  0.7559
#>  2.0612
#> -0.7005
#> -0.5123
#> -0.8126
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>