Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.3515  0.2237  0.8252 -0.6743 -0.7667 -0.5206  0.5267  0.8325 -1.5640 -1.7835
#>  1.1790  1.1046 -0.1922 -1.0800  1.6265 -0.4668 -0.5370  1.3464  0.5196  1.3288
#> -1.1571 -1.3863  1.8116  1.0386  0.4519  1.0031  0.1061 -0.0199  0.2892 -0.1698
#> -0.3628 -0.7521  0.7566  0.0720  0.5876  0.0298  0.5511  0.0192  2.0740 -1.2214
#>  1.0607  0.0922  0.9293 -0.8169  0.3020  0.5303 -0.3258 -1.1513  0.3578 -0.2395
#>  0.5558 -0.3523 -0.2932  2.9124  0.3865  0.1034  1.6208 -0.4582  1.2536 -0.6683
#> -2.0586 -0.7486 -0.6011  0.1539  0.9969 -1.4125  0.7834 -0.2458 -0.0971  2.4322
#>  0.6913  0.7955 -1.2933 -2.7692  0.7501 -0.4206 -0.6376  0.4089  1.2332 -0.8557
#>  0.3474 -0.0653  2.7501  1.5469 -1.3042 -0.9557  1.4748  0.2263 -0.0659 -0.0002
#>  0.4949 -1.2402  1.5710 -0.4868  0.7104  0.0702  0.6720  0.4283 -1.5196  0.7143
#> 
#> Columns 11 to 20-0.3943  0.2736 -1.8928 -0.0975 -0.8966  1.3438  0.4807  0.6675 -0.0200 -1.1050
#> -0.0007 -0.3385  1.0831  1.3470 -2.7735  0.1512  1.0784  1.2258  0.4912  0.1335
#>  0.4593 -0.6778 -0.2500  1.0829 -1.8914  0.6580  0.9512  0.1963 -0.5621 -0.5459
#> -0.6626  0.2859  0.6352 -0.5983  0.0588  0.6002 -0.5043 -0.1095 -0.1583  2.0797
#>  0.3567  0.1052 -0.9874  0.2289  0.6217  1.5982 -0.3134  1.8260 -0.3720 -0.7847
#>  0.4013 -1.3458 -0.5956 -0.2773 -0.7951  0.5613  1.7531 -1.5574  1.5400 -0.7739
#> -0.8732 -0.1187 -1.3919  0.0857 -0.1194  0.8179  0.7926 -0.3138  0.0911 -0.1175
#> -1.1124 -0.6831 -0.3295  0.0381  1.3450  1.4559  0.6812  0.0338 -0.7723 -0.9264
#>  0.7834  0.4379  1.2595  0.8158  0.5620 -0.4209 -0.8636  0.2861 -3.6178 -2.1891
#>  0.2655  0.6588 -0.5508  0.1195  0.1002  0.5999 -1.1708  0.4720  0.4365 -1.9289
#> 
#> Columns 21 to 30 0.5238  0.1863 -0.0564 -0.0639 -0.1825 -0.1804 -0.6057  0.0859 -1.5076 -0.4590
#>  0.3303  0.7518 -0.4857 -0.2795  0.6588  0.5853  0.3666 -0.9074 -0.3889  1.1277
#>  0.4153 -0.1587 -0.3121  1.0117 -0.4111 -0.0280 -1.0890  0.7589  0.1877  0.2210
#>  2.3449  0.2551  0.0166 -0.0116  0.5065 -0.1089  1.0416 -0.1064  0.2381  0.2646
#> -0.1302 -0.0844  0.2862  1.7499  1.3434 -1.1920  0.8152  0.0342 -0.8548  0.6045
#> -1.9840 -0.3310 -0.6227 -0.8303 -0.2535 -0.0184  1.8156  1.0807 -0.2392  0.1027
#>  0.2292 -0.6144 -1.0009 -0.3437  0.2001  0.7423 -1.5680 -1.2999 -0.9178  0.7991
#>  0.0290  0.1510 -0.5770  0.3687  1.1524 -0.8952  0.5175  0.6724  0.5539  0.2574
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  1.3469
#> -1.8503
#>  0.5203
#>  0.1844
#>  2.3800
#>  1.0969
#>  0.8616
#>  0.3207
#>  0.4519
#>  0.1806
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.8616
#>  0.1806
#> -1.8503
#>  0.4519
#>  0.3207
#>  1.0969
#>  0.1844
#>  2.3800
#>  1.3469
#>  0.5203
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>