Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 1.2491e-01  1.7683e+00 -5.4768e-01 -3.7407e-01 -9.4012e-02  5.9044e-01
#> -7.9687e-01 -6.1545e-02 -1.4408e+00 -1.2767e-01  1.8925e+00  1.4717e-01
#> -3.9363e-01  1.9477e-01  1.7455e+00 -5.4817e-01  1.0971e+00  1.6795e+00
#>  9.7160e-02 -1.9846e-01  5.8667e-01  1.0755e-01 -1.4571e+00 -6.1535e-01
#> -2.8956e-01  4.2242e-01 -4.8203e-01 -1.9082e+00  1.0379e+00 -5.8317e-01
#>  7.6832e-01 -1.1886e+00  1.1222e+00 -1.0756e+00 -9.9421e-01  2.0022e-01
#> -1.0575e-01 -6.5801e-01 -2.7753e-01  5.5326e-01 -9.0784e-01 -5.9297e-01
#>  2.0172e-01  2.2760e-01 -3.7425e-01 -6.7962e-01  9.5969e-01  5.2122e-01
#>  2.2333e-01 -3.4176e-02 -7.8918e-01 -9.2863e-01  9.2516e-01  6.2629e-01
#>  2.7739e-01 -2.7346e-02  9.8998e-01  1.5469e+00  1.5512e+00  8.4261e-02
#> 
#> Columns 7 to 12 1.3255e+00 -2.1376e+00 -6.5116e-01 -8.4822e-01  8.8254e-01  6.4939e-01
#>  1.1740e-01 -1.7764e+00 -1.4727e+00  9.2293e-01  5.4376e-01  4.4496e-01
#> -1.9047e+00 -1.3038e-01  1.3298e-01 -1.0134e+00 -8.0794e-01  4.9169e-01
#>  6.4308e-01 -4.7634e-01 -3.5311e-01  3.0137e-02 -5.3385e-01  5.5252e-02
#> -1.1066e+00  1.8971e+00  1.3075e+00 -5.2051e-01 -3.8319e-01  8.4862e-01
#>  1.1130e+00  1.2851e+00  4.3693e-01 -5.3565e-01 -1.8442e-01 -1.0446e-01
#> -4.9035e-01 -6.6424e-01  1.3736e+00  3.8792e-01  4.7966e-02 -4.7813e-01
#>  8.5011e-02 -6.6972e-01  1.6816e+00 -1.8256e+00  6.4082e-01  8.5258e-01
#>  5.5032e-01  1.7224e-01  4.5906e-01  7.8734e-01  1.1780e+00  1.5290e-01
#>  8.1312e-02  1.3631e+00  6.9918e-01  1.9288e-01  2.6514e-01  6.9961e-01
#> 
#> Columns 13 to 18 1.8658e+00 -9.8901e-03 -9.3868e-01 -1.0161e+00  2.0434e+00  1.8976e+00
#>  1.2650e+00  8.6052e-01 -6.0119e-01  5.8355e-02  1.4500e+00 -1.1465e+00
#>  8.4356e-02  4.2054e-02 -7.3554e-01 -3.6669e-01 -2.2096e-01 -3.3673e-01
#>  1.4701e+00  8.3601e-01  3.1784e-01  2.1537e+00 -2.5517e-02  5.1632e-01
#>  2.7327e-01  1.2408e+00 -7.0671e-01 -3.9601e-01  9.6168e-01  3.4049e-01
#> -5.5063e-02  2.4795e-01  9.1700e-06  9.5936e-01  2.3277e-02  1.9274e-01
#> -1.6155e+00  8.0742e-01  1.5483e+00 -5.2979e-01 -1.5781e+00 -1.5514e+00
#>  8.3927e-01 -9.1256e-01 -7.0158e-02  3.7412e-01 -1.0933e+00  1.6362e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.6302
#>  1.6692
#>  0.5345
#>  0.1300
#> -0.6930
#> -1.5121
#> -0.1247
#>  0.1953
#> -0.2264
#>  0.2770
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  1.6692
#> -0.6930
#>  0.1300
#> -0.2264
#>  0.5345
#> -1.5121
#>  0.2770
#>  0.6302
#> -0.1247
#>  0.1953
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>