Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 6 1.4480e+00  5.6716e-01  4.3820e-01 -3.0237e-01  4.8432e-01 -3.6650e-01
#> -1.2263e+00 -3.6042e-01  3.5457e-02 -1.5140e-01 -7.2303e-01  4.6342e-01
#> -6.4429e-02 -5.5008e-01  1.0072e+00 -2.1342e-01 -2.9510e-01 -3.8252e-01
#>  4.3854e-01  1.4476e+00 -5.1865e-01 -3.8696e-01  1.8292e-02 -6.2413e-01
#> -1.2844e-01 -2.3013e-01 -8.8838e-01 -9.0187e-01  3.9742e-02  6.9560e-02
#> -9.4090e-02 -1.0222e+00  1.5580e-01 -4.0427e-01 -2.0651e-01 -8.2555e-01
#>  3.9952e-01 -4.3934e-01  6.9464e-01  1.5888e-01 -8.4979e-01  5.8959e-03
#> -5.4891e-01  2.0923e-01  1.4031e-01 -1.1991e+00 -5.2549e-01  2.5286e-01
#>  4.9390e-01 -1.9438e+00 -2.4635e-01  3.0950e-01  6.5432e-01 -1.8719e-01
#>  2.4635e+00  1.8194e-01  9.4022e-01 -1.7425e-01 -1.5488e+00 -1.4307e+00
#> 
#> Columns 7 to 12-4.7707e-02 -9.9064e-01 -1.7181e+00 -1.0968e+00  7.7334e-01  6.8556e-01
#>  3.8387e-01  1.5801e-03  4.3847e-02 -4.1943e-01 -7.4848e-01 -8.6632e-02
#> -1.7617e+00 -6.7620e-01  4.7783e-02 -5.7171e-01  3.9885e-01  5.3847e-01
#>  2.5256e-01  2.7670e-01 -1.0948e+00  4.2275e-01  4.1720e-01  5.9692e-01
#>  6.6046e-01 -1.5788e+00 -3.8121e-01 -7.4877e-01 -2.4172e-02  1.0269e+00
#>  1.2544e+00  3.9916e-01 -8.0370e-01 -1.1805e+00 -5.2683e-01 -1.1671e-01
#>  1.4193e+00  6.8329e-01  1.6263e-01  2.2112e+00 -4.0598e-01  6.0594e-01
#> -1.2466e+00 -1.6733e-01 -2.6116e+00 -5.9579e-01 -9.7343e-01  2.7521e-01
#> -7.3162e-02 -1.3609e+00  1.6213e-01 -6.4450e-01  7.2790e-01  2.3713e+00
#>  6.0797e-01 -2.4841e-02  8.5822e-01  2.5436e-01  4.6907e-02  7.8818e-01
#> 
#> Columns 13 to 18-1.2811e-01 -1.3480e+00  2.6356e-01  5.3048e-02 -1.0339e-01 -4.3345e-01
#> -1.1079e-01 -9.0715e-01  4.7864e-01  2.8790e-01  7.1554e-01  1.6047e-01
#>  9.4652e-01  3.5733e-01 -8.4672e-01 -6.8584e-01 -6.9073e-01  5.3667e-02
#> -7.7580e-01 -1.5712e-01 -8.9472e-02  6.8847e-01 -4.2095e-01 -2.9703e-01
#>  8.4743e-01  1.3242e+00  9.5820e-01 -6.3287e-01 -3.7568e-01 -1.3349e-02
#> -4.0935e-01  1.6063e+00  1.6042e+00  6.3784e-01  3.1142e-01 -8.7924e-01
#> -7.7706e-01 -6.3862e-01  1.7811e+00  1.4693e-01  4.4597e-01 -6.8494e-01
#> -3.5741e-01 -9.4129e-01  1.7700e+00 -3.5418e-02  4.0085e-01  1.2446e+00
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.3394
#>  0.5451
#> -1.0298
#>  0.0923
#> -0.1514
#> -1.2811
#> -0.0599
#>  0.8605
#>  0.6868
#> -0.2285
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#> -1.2811
#>  0.3394
#> -0.1514
#>  0.8605
#> -1.0298
#> -0.2285
#>  0.5451
#> -0.0599
#>  0.0923
#>  0.6868
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>