Mixup logic

Logic underlying luz_callback_mixup().

Usage

nnf_mixup(x, y, weight)

Arguments

x

an input batch

y

a target batch

weight

weighting coefficient to be used by torch_lerp()

Value

A list of:

x, the new, mixed-up input batch

y, a list of:

ys, a list of:

y1, the original target y1

y2, the mixed-in target y2

weight, the mixing weights

Details

Based on the passed-in input and target batches, as well as applicable mixing weights, we return new tensors intended to replace the current batch. The new input batch is a weighted linear combination of input batch items, while the new target batch bundles the original targets, as well as the mixing weights, in a nested list.

See also

Examples

if (torch::torch_is_installed()) {
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, weight)
}
#> $x
#> torch_tensor
#> Columns 1 to 10-0.3710  1.5973 -0.3849 -1.3225  1.6546 -1.4248  0.7682 -1.1697  0.7890  1.3240
#> -0.5528  0.5025 -1.0103  0.1809 -0.5656  0.3747  0.9357 -0.7457 -0.0642  1.0054
#> -0.7204  1.0797  0.9733 -2.0746  0.1190 -0.9037  0.1147  0.2290  1.6448  0.3488
#>  0.3213  0.4088 -0.7337 -0.0561  0.2936 -0.9049  0.0538  1.5715 -0.8026 -0.2936
#>  0.4696 -1.4718 -0.9381  0.4350 -0.7674 -0.0806 -0.9022 -0.4781 -0.0990 -0.0047
#> -0.1304 -0.9347  0.4201 -0.1046  2.5226 -0.6256 -0.5165  1.0131  0.6716 -0.1815
#> -0.8234  1.2019  1.0180  0.3342  0.3926 -2.3877 -0.5545  1.3314  0.2824  1.1317
#> -1.0897 -0.9428  0.3327 -0.2295  1.6597 -0.4609  0.4605 -0.0209 -0.0047  0.7232
#>  0.6847  0.5243  1.6731 -0.8387 -0.8711  2.5750  0.9113  1.9810 -1.4801  0.0660
#> -0.5212 -0.1450 -0.9373 -0.4030  0.6814 -1.2532 -1.0969  0.6584 -1.0707 -1.0461
#> 
#> Columns 11 to 20-0.8907 -1.9608 -2.2027 -0.6012  0.1468  1.2038 -0.2680 -0.8490 -2.0119 -0.7550
#> -1.1494 -0.4541 -2.5036 -1.5660  0.6335 -0.8295  0.0025 -0.0495  2.0164  1.3216
#> -0.1055  1.6910 -1.5950 -1.8300  0.2842 -0.3873 -0.4406  0.6103  0.1089  0.7816
#>  0.3011  1.1330 -1.3197  1.5306  1.9685 -0.6323  0.5659 -0.6188 -1.5105 -1.8844
#>  0.8504 -0.6697 -0.1148 -0.6446  1.1098  1.0919 -0.4556  1.3698 -0.2151 -0.2373
#> -0.5735 -1.0946  0.6954 -1.9778 -0.4580  0.3927  0.5578  1.4882 -0.4383 -0.7735
#> -0.1319 -0.8984 -0.2572 -0.8221 -1.1430  0.7035  0.2722  0.9133 -0.7502 -0.9823
#>  0.5215 -0.1527  1.1111  0.2230 -0.0478 -0.2162 -0.2292  0.8974 -0.6274 -1.7110
#> -0.6068 -0.4180  0.3338 -0.8114  0.4181  1.2410  0.5768  0.9878  0.1942  0.3799
#> -0.5190 -1.3485 -0.4214 -0.8749  1.6179 -0.5712  1.9218  0.3536 -0.9683 -1.2662
#> 
#> Columns 21 to 30 0.9042 -0.0412 -0.0500  0.4175 -0.7133 -0.4481 -0.6279  0.0300  0.7225  0.7724
#> -1.1964  0.4151  0.5797 -0.6225  0.5497  1.2321 -0.4384  0.7412  1.8507  1.5332
#>  0.0528 -1.4031 -0.3395  0.1539 -0.8282  0.4599 -1.3100  0.6836  0.6901 -1.5408
#>  1.1243 -0.5657 -1.6695  0.1477 -1.7544  0.4311 -0.1987  0.8095  2.8835  0.4514
#>  1.3616  1.4846  0.0288 -0.2093  0.4394  1.5217  0.5349  0.3800  0.6019  0.4137
#>  0.4025  1.4262  0.5062 -0.5451 -2.1657  0.4025 -0.3521 -0.2827 -0.4289  0.1972
#> -0.9984  0.6434 -1.5164 -1.0632 -0.5786 -0.2030 -1.7295 -1.6752 -0.4436 -0.6699
#> -0.3344  0.5492 -0.3282 -0.4511 -0.6396  1.7684 -0.5355  0.0574  0.9696  1.4051
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{10,768} ]
#> 
#> $y
#> $y$ys
#> $y$ys$y1
#> torch_tensor
#>  0.3176
#>  0.1448
#> -2.0424
#>  0.1949
#> -0.6014
#>  0.7148
#>  0.9054
#>  0.7388
#>  0.2836
#>  1.1405
#> [ CPUFloatType{10} ]
#> 
#> $y$ys$y2
#> torch_tensor
#>  0.1949
#>  0.1448
#>  1.1405
#>  0.7388
#>  0.7148
#>  0.3176
#>  0.9054
#> -0.6014
#>  0.2836
#> -2.0424
#> [ CPUFloatType{10} ]
#> 
#> 
#> $y$weight
#> torch_tensor
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#>  0.9000
#> [ CPUFloatType{10,1} ]
#> 
#>